### 2025-08-06 #8758 @warrickball
Hi @timlichtenberg, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. We currently have a long backlog in this domain so it will probably take at least 4 months before that happens. Sorry, and thanks for your patience!

In the meantime, I noticed that the paper is currently much too long. [Our docs say](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

You can have a look at the content that we expect in the paper. I'll try to have a closer look in the next few days to make more specific suggestions but broadly the paper should briefly motivate the existence of the software through its research context and given existing tools. Also, 1000 words isn't a hard limit but it would be good to cut the paper well under 2000 words and better still under about 1500.

None of this affects your place in the submission queue.
