### 2025-08-06 0b047f8
Inference tests work
