### 2025-08-06 c0e0c65
AGNI p_obs to 20 mbar
