+++ src/proteus/config/_atmos_clim.py
@@ -79,7 +79,7 @@ class Agni:
     spectral_group: str     = field(default=None)
     spectral_bands: str     = field(default=None)
     p_top: float            = field(default=1e-5, validator=gt(0))
-    p_obs: float            = field(default=2e-3, validator=gt(0))
+    p_obs: float            = field(default=20e-3, validator=gt(0))
     surf_material: str      = field(default="surface_albedos/Hammond24/lunarmarebasalt.dat")
     num_levels: int         = field(default=40, validator=ge(15))
     chemistry: str          = field(default="none",
