+++ tests/grid/base.toml
@@ -3,7 +3,6 @@
 # ----------------------------------------------------
 # Metadata
 version = "2.0"
-author  = "Harrison Nicholls, Tim Lichtenberg"
 
 # ----------------------------------------------------
 # Parameters
