+++ src/proteus/config/_config.py
@@ -28,9 +28,14 @@ def spada_zephyrus(instance, attribute, value):
         raise ValueError('ZEPHYRUS must be used with MORS and the Spada evolution tracks')
 
 def instmethod_dummy(instance, attribute, value):
-    # Instellation method can only be 'inst' for dummy star module
+    # Instellation method 'inst' only support for dummy star module
     if (instance.orbit.instellation_method == 'inst') and not (instance.star.module == 'dummy'):
         raise ValueError("Instellation method can only be 'inst' when star.module=dummy ")
+
+def instmethod_evolve(instance, attribute, value):
+    # Orbital evolution not supported when installation_method is 'inst'
+    if (instance.orbit.instellation_method == 'inst') and instance.orbit.evolve:
+        raise ValueError("Orbital evolution not supported for `instellation_method='inst'`")
 
 def tides_enabled_orbit(instance, attribute, value):
     # Tides in interior requires orbit module to not be None
@@ -78,7 +83,7 @@ class Config:
 
     params: Params
     star: Star
-    orbit: Orbit = field(validator=(instmethod_dummy,))
+    orbit: Orbit = field(validator=(instmethod_dummy,instmethod_evolve))
     struct: Struct
     atmos_clim: AtmosClim
     atmos_chem: AtmosChem
