+++ src/proteus/orbit/dummy.py
@@ -28,7 +28,7 @@ def run_dummy_orbit(config:Config, interior_o:Interior_t):
     Returns
     ----------
         Imk2_love: float
-            Always has a value of zero with this module.
+            Always returns the value provided in the config.
     """
 
     # Default case; zero heating throughout the mantle
@@ -50,5 +50,4 @@ def run_dummy_orbit(config:Config, interior_o:Interior_t):
                 # Heating maximised when fully liquid.
                 interior_o.tides[i] = config.orbit.dummy.H_tide * (p-Pt_va) / (1-Pt_va)
 
-    # Return Im(k2) = 0
-    return 0.0
+    return config.orbit.dummy.Imk2
