+++ input/demos/dummy.toml
@@ -4,7 +4,7 @@ version = "2.0"
 [params]
     [params.out]
         path        = "dummy"
-        logging     = "ERROR"
+        logging     = "DEBUG"
         plot_mod    = 0      # Plotting frequency, 0: wait until completion | n: every n iterations
         write_mod   = 50      # Write CSV frequency, 0: wait until completion | n: every n iterations
         archive_mod = 'none'
@@ -54,16 +54,19 @@ version = "2.0"
 
 # Orbital system
 [orbit]
-    semimajoraxis   = 1.0       # AU
-    eccentricity    = 0.0       # dimensionless
+    semimajoraxis   = 0.7       # AU
+    eccentricity    = 0.2     # dimensionless
     zenith_angle    = 48.19     # degrees
     s0_factor       = 0.375     # dimensionless
+
+    evolve = true
 
     module  = "dummy"
 
     [orbit.dummy]
-        H_tide  = 1e-9  # Fixed tidal power density [W kg-1]
+        H_tide   = 1e-9     # Fixed tidal power density [W kg-1]
         Phi_tide = "<0.3"   # Tidal heating applied when inequality locally satisfied
+        Imk2     = -1e3     # Fixed Im(k2) love number
 
     [orbit.lovepy]
         visc_thresh = 1e9   # Minimum viscosity required for heating [Pa s]

[…]
