### 2025-08-06 b543ffb
Added functions to be used for Earth-Moon orbit
