### 2025-08-06 fe392c8
Added plot_orbit function
