### 2025-08-07 c523bf3
Better plots for inference output
