### 2025-08-07 2fa79ed
Inference print results formatting
