### 2025-08-07 6f0a96a
Define some obs variables on log-scale
