### 2025-08-07 cd87038
Cleaned-up satellite orbit functions
