### 2025-08-07 7add521
Option for disabling satellite orbital evolution.
