### 2025-08-07 bdda032
Added plot for orbital parameters
