### 2025-08-07 03e53f3
Added new entries to helpfile for satellite orbit
