+++ src/proteus/utils/coupler.py
@@ -462,6 +462,9 @@ def GetHelpfileKeys():
             # Orbital period and eccentricity
             "orbital_period", "eccentricity", # [s], [1]
 
+            # Satellite orbit semi-major axis and system angular momentum
+            "semimajorax_sat", "system_am", # [m], [kg m2 s-1],
+
             # Day length
             "axial_period", # [s]
 
