### 2025-08-08 0d47085
Inference print info
