### 2025-08-08 f5e00cb
Remove option for setting initial AM
