### 2025-08-08 #8758 @timlichtenberg
Hi @warrickball, thanks for the feedback and inclusion into the waitlist!

As a start, here are some people who may be suitable as reviewers for this submission:
Steven Rieder
Lorena Acuna
Arjun Savel
Ryan MacDonald
Evert Nasedkin
Michael Radica
Zach K. Berta-Thompson

Regarding the article length: we would prefer to wait for (your's and) the reviewer's specific suggestions and feedback, and then work on the text to converge toward the requirements. In our [last JOSS submission](https://github.com/openjournals/joss-reviews/issues/7726) the reviews led to an increase in manuscript words, and I would want to avoid first cutting and then having to re-insert text.

In any case, I look forward to your suggestions and to start the review process.
