### 2025-08-08 6538f94
Option for enabling/disabling satellite. Fix orbit plot syntax/axes issues. Moon orbit now specified in m rather than AU.
