### 2025-08-08 e2dbbe9
Plot automatically
