+++ input/demos/dummy.toml
@@ -6,7 +6,7 @@ version = "2.0"
         path        = "dummy"
         logging     = "ERROR"
         plot_mod    = 0      # Plotting frequency, 0: wait until completion | n: every n iterations
-        write_mod   = 100      # Write CSV frequency, 0: wait until completion | n: every n iterations
+        write_mod   = 400      # Write CSV frequency, 0: wait until completion | n: every n iterations
         archive_mod = 'none'
 
     [params.dt]
@@ -19,7 +19,7 @@ version = "2.0"
 
         [params.dt.adaptive]
             atol         = 0.02   # Step size atol
-            rtol         = 0.12   # Step size rtol
+            rtol         = 0.10   # Step size rtol
 
     [params.stop]
         # required time constraints
@@ -54,8 +54,8 @@ version = "2.0"
 
 # Orbital system
 [orbit]
-    semimajoraxis   = 0.4       # AU
-    eccentricity    = 0.6       # dimensionless
+    semimajoraxis   = 0.02       # AU
+    eccentricity    = 0.5       # dimensionless
     zenith_angle    = 48.19     # degrees
     s0_factor       = 0.375     # dimensionless
     axial_period    = 'none'    # hours, or provide 'none' to set as tidally locked
