+++ src/proteus/orbit/satellite.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 import numpy as np
 from scipy.integrate import solve_ivp
 
-from proteus.utils.constants import AU, const_G, secs_per_hour
+from proteus.utils.constants import const_G, secs_per_hour
 
 if TYPE_CHECKING:
     from proteus.config import Config
