+++ src/proteus/plot/__init__.py
@@ -11,13 +11,13 @@ from proteus.plot.cpl_fluxes_global import plot_fluxes_global_entry
 from proteus.plot.cpl_global import plot_global_entry
 from proteus.plot.cpl_interior import plot_interior_entry
 from proteus.plot.cpl_interior_cmesh import plot_interior_cmesh_entry
+from proteus.plot.cpl_orbit import plot_orbit_entry
 from proteus.plot.cpl_population import plot_population_entry
 from proteus.plot.cpl_sflux import plot_sflux_entry
 from proteus.plot.cpl_sflux_cross import plot_sflux_cross_entry
 from proteus.plot.cpl_spectra import plot_spectra_entry
 from proteus.plot.cpl_structure import plot_structure_entry
 from proteus.plot.cpl_visual import plot_visual_entry
-from proteus.plot.cpl_orbit import plot_orbit_entry
 
 plot_dispatch = {
     'atmosphere':           plot_atmosphere_entry,
