### 2025-08-08 b133ba7
Don't allow satellite and orbital evolution at same time, for now
