### 2025-08-08 39c57ff
Show best result on plot panels@
