+++ src/proteus/utils/coupler.py
@@ -462,8 +462,8 @@ def GetHelpfileKeys():
             # Orbital period and eccentricity
             "orbital_period", "eccentricity", # [s], [1]
 
-            # Satellite orbit semi-major axis and system angular momentum
-            "semimajorax_sat", "plan_sat_am", # [m], [kg m2 s-1],
+            # Satellite orbit semi-major axis, mass, and planet-satellite system angular momentum
+            "semimajorax_sat", "M_sat", "plan_sat_am", # [m], [kg], [kg m2 s-1],
 
             # Day length
             "axial_period", # [s]
