### 2025-08-10 db6c181
Added two demo configs
