+++ input/demos/dummy_test_sat.toml
@@ -1,0 +1,198 @@
+# PROTEUS configuration file (version 2.0)
+version = "2.0"
+
+[params]
+    [params.out]
+        path        = "dummy_test_sat"
+        logging     = "INFO"
+        plot_mod    = 10      # Plotting frequency, 0: wait until completion | n: every n iterations
+        write_mod   = 10      # Write CSV frequency, 0: wait until completion | n: every n iterations
+        archive_mod = 'none'
+
+    [params.dt]
+        minimum      = 1e2    # yr, minimum time-step
+        maximum      = 3e7    # yr, maximum time-step
+        initial      = 1e2    # yr, inital step size
+        starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
+        starinst     = 1e1    # yr, interval to re-calculate the instellation
+        method       = "adaptive"  # proportional | adaptive | maximum
+
+        [params.dt.adaptive]
+            atol         = 0.02   # Step size atol
+            rtol         = 0.10   # Step size rtol
+
+    [params.stop]
+        # required time constraints
+        [params.stop.time]
+            enabled = true
+            minimum = 1.0e3     # yr, model will certainly run to t > minimum
+            maximum = 1e9     # yr, model will terminate when t > maximum
+
+        # solidification
+        [params.stop.solid]
+            enabled  = true
+            phi_crit = 0.005  # non-dim., model will terminate when global melt fraction < phi_crit
+
+        # radiative equilibrium
+        [params.stop.radeqm]
+            enabled = true
+            atol    = 0.2     # absolute tolerance [W m-2]
+            rtol    = 1e-3    # relative tolerance
+
+        [params.stop.escape]
+            enabled   = true
+            p_stop    = 1.0  # bar, model will terminate with p_surf < p_stop
+
+
+# Star
+[star]
+
+    # Physical parameters
+    mass    = 1.0       # stellar mass [M_sun]
+    age_ini = 0.100     # model initialisation age [Gyr]
+
+    module  = "dummy"
+
+    [star.dummy]
+        radius  = 1.0       # Stellar radius [R_sun]
+        Teff    = 5772.0    # Star's brightness temperature [K]
+
+# Orbital system
+[orbit]
+    semimajoraxis   = 0.02       # AU
+    eccentricity    = 0.5       # dimensionless
+    zenith_angle    = 48.19     # degrees
+    s0_factor       = 0.375     # dimensionless
+    axial_period    = 6    # hours, or provide 'none' to set as tidally locked
+
+    evolve  = false
+
+    satellite = true
+    semimajoraxis_sat = 32e6
+
+    module  = "dummy"
+
+    [orbit.dummy]
+        H_tide   = 1e-7     # Fixed tidal power density [W kg-1]
+        Phi_tide = "<0.3"   # Tidal heating applied when inequality locally satisfied
+        Imk2     = -1e5     # Fixed Im(k2) love number
+
+    [orbit.lovepy]
+        visc_thresh = 1e9   # Minimum viscosity required for heating [Pa s]
+
+# Planetary structure - physics table
+[struct]
+    mass_tot     = 1.0           # Total planet mass [M_earth]
+    corefrac     = 0.55          # non-dim., radius fraction
+    core_density = 10738.33      # Core density [kg m-3]
+    core_heatcap = 880.0         # Core specific heat capacity [J K-1 kg-1]
+
+# Atmosphere - physics table
+[atmos_clim]
+    prevent_warming = false      # do not allow the planet to heat up
+    surface_d       = 0.01      # m, conductive skin thickness
+    surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
+    cloud_enabled   = false     # enable water cloud radiative effects
[…]
