### 2025-08-11 2a91f94
Added Msat configuration option
