### 2025-08-11 1c72d3a
Updated two demo configs
