### 2025-08-11 3ceb6cf
Option for setting dummy interior tolerance. Updated config file names for clarity. Don't write status file every step
