### 2025-08-11 b64e6fd
Refactor Zalmoxis implementation: rename inner_mantle_mass_fraction to mantle_mass_fraction; clean up functions in zalmoxis.py and use imports instead from the module; remove unused earth_center_pressure constant.
