+++ src/proteus/interior/zalmoxis.py
@@ -5,22 +5,19 @@ import logging
 import os
 
 import numpy as np
-from scipy.integrate import solve_ivp
-from scipy.interpolate import interp1d
 
 from proteus.config import Config
 from proteus.utils.constants import (
     M_earth,
     R_earth,
-    const_G,
-    earth_center_pressure,
     element_list,
 )
 from proteus.utils.data import get_Seager_EOS
+from zalmoxis.zalmoxis import main
 
-log = logging.getLogger("fwl."+__name__)
+# Set up logging
+logger = logging.getLogger("fwl."+__name__)
 
-# Path to location at which to save the output data from Zalmoxis
 def get_zalmoxis_output_filepath(outdir: str):
     """Returns the output file path for Zalmoxis data.
     Args:
@@ -30,151 +27,19 @@ def get_zalmoxis_output_filepath(outdir: str):
     """
     return os.path.join(outdir, "data", "zalmoxis_output.dat")
 
-def get_tabulated_eos(pressure, material_dictionary, material, interpolation_functions={}):
-    """
-    Retrieves density from tabulated EOS data for a given material and choice of EOS."""
-    props = material_dictionary[material]
-    try:
-        eos_file = props["eos_file"]
-        # Caching: Store interpolation functions for reuse
-        if eos_file not in interpolation_functions:
-            data = np.loadtxt(eos_file, delimiter=',', skiprows=1)
-            pressure_data = data[:, 1] * 1e9
-            density_data = data[:, 0] * 1e3
-            interpolation_functions[eos_file] = interp1d(pressure_data, density_data, bounds_error=False, fill_value="extrapolate")
-
-        interpolation_function = interpolation_functions[eos_file]  # Retrieve from cache
-        density = interpolation_function(pressure)  # Call to cached function
-
-        if density is None or np.isnan(density):
-            raise ValueError(f"Density calculation failed for {material} at {pressure:.2e} Pa.")
-
-        return density
-
-    except (ValueError, OSError) as e: # Catch file errors
-        print(f"Error with tabulated EOS for {material} at {pressure:.2e} Pa: {e}")
-        return None
-    except Exception as e: # Other errors
-        print(f"Unexpected error with tabulated EOS for {material} at {pressure:.2e} Pa: {e}")
-        return None
-
-def get_density_from_eos(pressure, material, eos_choice, interpolation_functions={}):
-    """Calculates density with caching for tabulated EOS.
+def load_zalmoxis_configuration(config:Config, hf_row:dict):
+    """Loads the model configuration for Zalmoxis and calculates the dry mass of the planet based on the total mass and the mass of volatiles.
     Args:
-        pressure (float): Pressure [Pa].
-        material (str): Material type (e.g., "core", "mantle").
-        eos_choice (str): Equation of state choice.
-        interpolation_functions (dict): Cache for interpolation functions.
+        config (Config): The configuration object containing the Zalmoxis parameters.
+        hf_row (dict): A dictionary containing the mass of volatiles and other parameters.
     Returns:
-        density (float): Calculated density [kg/m^3].
-    """
-    # Get material properties for Zalmoxis
-    material_properties_iron_silicate_planets, material_properties_water_planets = get_Seager_EOS()
-
-    # Select the equation of state and calculate density
-    if eos_choice == "Tabulated:iron/silicate":
-        return get_tabulated_eos(pressure, material_properties_iron_silicate_planets, material, interpolation_functions)
-    elif eos_choice == "Tabulated:water":
-        return get_tabulated_eos(pressure, material_properties_water_planets, material, interpolation_functions)
-    else:
-        log.error(f"Invalid EOS choice: {eos_choice}.")
-
-def interior_structure_odes(radius, y, cmb_mass, inner_mantle_mass, eos_choice, interpolation_cache):
-    """
[…]
