### 2025-08-11 360f766
Tides working with Aragog. Need to test SPIDER again.
