### 2025-08-12 05bf9b3
LovePy working with dummy interior
