### 2025-08-12 ed7fcf5
Remove duplicate core_cp parameter
