+++ input/demos/tides_old.toml
@@ -1,0 +1,178 @@
+# PROTEUS configuration file (version 2.0)
+
+version = "2.0"
+
+[params]
+    [params.out]
+        path        = "tides_old_aragog"
+        logging     = "INFO"
+        plot_mod    = 40      # Plotting frequency, 0: wait until completion | n: every n iterations
+        write_mod   = 2
+        archive_mod = 250
+        remove_sf   = true
+
+    # time-stepping
+    [params.dt]
+        minimum      = 3e2    # absolute minimum time-step [years]
+        minimum_rel  = 1e-2   # relative minimum time-step [dimensionless]
+        maximum      = 1e8    # yr, maximum time-step
+        initial      = 1e4    # yr, inital step size
+        starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
+        starinst     = 1e2    # yr, interval to re-calculate the instellation
+        method       = "adaptive"  # proportional | adaptive | maximum
+        [params.dt.adaptive]
+            atol         = 0.03  # Step size atol
+            rtol         = 0.13   # Step size rtol
+
+    [params.stop]
+
+        [params.stop.time]
+            enabled = true
+            maximum = 4.89e9     # yr, model will terminate when t > maximum
+
+        # solidification
+        [params.stop.solid]
+            enabled  = true
+            phi_crit = 0.005
+
+        # radiative equilibrium
+        [params.stop.radeqm]
+            enabled = false
+
+        [params.stop.escape]
+            enabled   = false
+            p_stop    = 1.0  # bar, model will terminate with p_surf < p_stop
+
+
+# ----------------------------------------------------
+# Star
+[star]
+
+    mass    = 0.273      # M_sun
+    age_ini = 0.05     # Gyr, model initialisation/start age
+
+    module  = "mors"
+    [star.mors]
+        rot_period = 80.9       # rotation period
+        tracks  = "spada" # evolution tracks: spada | baraffe
+        age_now = 4.40     # Gyr, current age of star used for scaling
+        spec    = "stellar_spectra/Named/l-98-59.txt" # stellar spectrum
+
+
+[orbit]
+    semimajoraxis   = 0.0486    # AU
+    eccentricity    = 0.098      # dimensionless
+    zenith_angle    = 54.74     # degrees
+    s0_factor       = 0.25      # dimensionless
+
+    module  = "lovepy"
+    [orbit.lovepy]
+        visc_thresh = 1e9   # Minimum viscosity required for heating [Pa s]
+
+[struct]
+    mass_tot    = 1.0       # M_earth
+    corefrac    = 0.55      # non-dim., radius fraction
+
+[atmos_clim]
+    prevent_warming = true     # do not allow the planet to heat up
+    surface_d       = 0.01      # m, conductive skin thickness
+    surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
+    surf_state      = "skin"    # surface scheme: "mixed_layer" | "fixed" | "skin"
+    rayleigh        = false     # Enable rayleigh scattering
+
+    module  = "dummy"           # Which atmosphere module to use
+
+    [atmos_clim.dummy]
+        gamma = 0.8
+
+    [atmos_clim.agni]
+        p_top           = 2.0e-5        # bar, top of atmosphere grid pressure
+        spectral_group  = "Honeyside"   # which gas opacities to include
+        spectral_bands  = "48"         # how many spectral bands?
+        num_levels      = 40            # Number of atmospheric grid levels
+        chemistry       = "none"        # "none" | "eq"
+        surf_material   = "surface_albedos/Hammond24/lunarmarebasalt.dat"    # surface material file for scattering
[…]
