### 2025-08-12 e2fe257
Moved bulk_modulus parameter into aragog-specific part of the config. Clarify log statements. Seems to work well
