+++ input/demos/orbit_physical.toml
@@ -120,7 +120,6 @@ version = "2.0"
     tidal_heat      = true  # enable tidal heat production
     rheo_phi_loc    = 0.3    # Centre of rheological transition
     rheo_phi_wid    = 0.15   # Width of rheological transition
-    bulk_modulus    = 260e9   # Bulk modulus [Pa]
 
     module = "aragog"   # Which interior module to use
 
