### 2025-08-12 3bcfb7f
Update dummy test to include orbital evolution
