+++ tests/integration/test_integration_dummy.py
@@ -60,4 +60,3 @@ def test_dummy_physics(dummy_run):
     # reasonable surface temperatures
     assert row_1["T_surf"] < row_0["T_surf"]
     assert row_1["T_surf"] > 100.0
-
