### 2025-08-12 9457947
Added pariapsis distance calculation to orbit wrapper
