### 2025-08-12 76b61ff
Added planet disintegration termination condition. Updated planet entering star's Roche limit warning.
