### 2025-08-13 11dca62
Get ocean variables from atmosphere. Tidy up code for copying result to hf_row. Earth config
