+++ src/proteus/atmos_clim/agni.py
@@ -207,6 +207,7 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
                         use_all_gases=include_all,
                         fastchem_work = fc_dir,
                         real_gas = config.atmos_clim.agni.real_gas,
+                        check_integrity = False, # don't check thermo files every time
 
                         skin_d=config.atmos_clim.surface_d,
                         skin_k=config.atmos_clim.surface_k,
@@ -537,22 +538,34 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
     log.debug("Running AGNI...")
     time_str = "%d"%hf_row["Time"]
 
+    # ---------------------------
     # Solve atmosphere
+    # ---------------------------
+
+    # Transparent case
     if bool(atmos.transparent):
         # no opacity
         log.info("Using transparent solver")
         atmos = _solve_transparent(atmos, config)
 
+    # Opaque case
     else:
-        # has opacity
+        # full solver
         if config.atmos_clim.agni.solve_energy:
             log.info("Using nonlinear solver to conserve fluxes")
             atmos = _solve_energy(atmos, loops_total, dirs, config)
+
+        # simplified T(p)
         else:
             log.info("Using prescribed temperature profile")
             atmos = _solve_once(atmos, config)
 
+    # Set observed pressure, get observed radius
+    atmos.transspec_p = float(config.atmos_clim.agni.p_obs * 1e5) # converted to Pa
+    jl.AGNI.atmosphere.calc_observed_rho_b(atmos)
+
     # Write output data
+    log.debug("AGNI write to NetCDF file")
     ncdf_path = os.path.join(dirs["output"],"data",time_str+"_atm.nc")
     jl.AGNI.save.write_ncdf(atmos, ncdf_path)
 
@@ -561,18 +574,6 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
         cff = os.path.join(dirs["output/plots"], f"plot_cff.{config.params.out.plot_fmt}")
         jl.AGNI.plotting.plot_contfunc1(atmos, cff)
 
-    # ---------------------------
-    # Calculate observables
-    # ---------------------------
-
-    # Set observed level
-    atmos.transspec_p = float(config.atmos_clim.agni.p_obs * 1e5) # convert to Pa
-
-    # observed height and derived bulk density
-    jl.AGNI.atmosphere.calc_observed_rho_b(atmos)
-    rho_obs = float(atmos.transspec_rho)
-    p_obs   = float(atmos.transspec_p)
-    r_obs   = float(atmos.transspec_r)
 
     # ---------------------------
     # Parse results
@@ -584,6 +585,15 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
     SW_flux_up =    np.array(atmos.flux_u_sw)
     SW_flux_down =  np.array(atmos.flux_d_sw)
     T_surf =        float(atmos.tmp_surf)
+
+    # Print info to user
+    if config.atmos_clim.condensation:
+        log.info("    oceans = %.3f %% area covered"%float(atmos.ocean_areacov*100))
+        log.info("    oceans = %.3f km maxdepth"%float(atmos.ocean_maxdepth/1e3))
+    log.debug("    T_surf = %.3f K"%T_surf)
+    log.info("    R_obs  = %.3f km"%float(float(atmos.transspec_r)/1e3))
+    log.info("    F_top  = %.2e W m-2"%float(tot_flux[0]))
[…]
