### 2025-08-13 c4c0a7e
Separate options for condense and latent_heat. Ocean info printed/stored
