+++ input/planets/earth.toml
@@ -21,7 +21,7 @@ version = "2.0"
 
         [params.stop.time]
             enabled = true
-            maximum = 4.567e+9
+            maximum = 4.44e+9 # 4.54 Gyr minus age_ini
 
         [params.stop.solid]
             enabled  = false
@@ -33,8 +33,8 @@ version = "2.0"
             rtol    = 1e-3
 
         [params.stop.escape]
-            enabled   = false
-            p_stop    = 1.0
+            enabled   = true
+            p_stop    = 0.2
 
 [star]
     mass    = 1.0
@@ -60,15 +60,15 @@ version = "2.0"
 [atmos_clim]
     surf_state      = "fixed"
     rayleigh        = true
-    condensation    = true
     module          = "agni"
 
     [atmos_clim.agni]
+        condensation    = true
         spectral_group = "Dayspring"
         spectral_bands = "16"
 
 [escape]
-    module = "none"
+    module = "zephyrus"
 
 [interior]
     radiogenic_heat = false
@@ -77,7 +77,7 @@ version = "2.0"
     module = "dummy"
 
     [interior.dummy]
-        init_tmagma = 3000.0
+        ini_tmagma = 3000.0
 
     [interior.aragog]
         ini_tmagma = 3000.0
