### 2025-08-13 4b6e018
Fix print formatting
