### 2025-08-13 d33b808
Set dx_max for AGNI in PROTEUS config
