### 2025-08-13 b0c7484
Earth config. Code style
