+++ src/proteus/orbit/wrapper.py
@@ -211,4 +211,3 @@ def run_orbit(hf_row:dict, config:Config, dirs:dict, interior_o:Interior_t):
 
     # Call tides module for satellite, calculates heating rates and new love number
     # To Do
-
