+++ input/planets/earth.toml
@@ -10,13 +10,14 @@ version = "2.0"
         write_mod   = 2
 
     [params.dt]
-        starspec     = 3e6
+        starspec     = 1e7
         starinst     = 1e2
-        minimum      = 1e4
+        minimum      = 3e4
+        initial      = 3e4
 
         [params.dt.adaptive]
-            atol         = 0.02
-            rtol         = 0.10
+            atol         = 0.03
+            rtol         = 0.12
 
     [params.stop]
 
@@ -44,7 +45,7 @@ version = "2.0"
     module  = "mors"
     [star.mors]
         rot_pcntle = 50.0
-        age_now = 0.414
+        age_now = 4.57 # Gyr
         spec    = "stellar_spectra/Named/sun.txt"
 
 [orbit]
@@ -62,6 +63,7 @@ version = "2.0"
     surf_state      = "fixed"
     rayleigh        = true
     module          = "agni"
+    albedo_pl       = 0.15      # test guess, contribution from clouds
 
     [atmos_clim.agni]
         p_top          = 1e-4
@@ -79,7 +81,8 @@ version = "2.0"
     module = "dummy"
 
     [interior.dummy]
-        ini_tmagma = 2700.0
+        ini_tmagma = 2000.0
+
 
     [interior.aragog]
         ini_tmagma = 3000.0
@@ -94,17 +97,17 @@ version = "2.0"
     module = "none"
 
     [delivery.elements]
-        # H_oceans    = 0.0       # Hydrogen inventory in units of equivalent Earth oceans
-        H_ppmw      = 109.0     # Hydrogen inventory in ppmw relative to mantle mass
+        H_oceans    = 3.0       # Hydrogen inventory in units of equivalent Earth oceans
+        # H_ppmw      = 109.0     # Hydrogen inventory in ppmw relative to mantle mass
 
-        # CH_ratio    = 1.0       # C/H mass ratio in mantle/atmosphere system
-        C_ppmw      = 109.0       # Carbon inventory in ppmw relative to mantle mass
+        CH_ratio    = 1.0       # C/H mass ratio in mantle/atmosphere system
+        # C_ppmw      = 109.0       # Carbon inventory in ppmw relative to mantle mass
 
-        # NH_ratio    = 0.018       # N/H mass ratio in mantle/atmosphere system
-        N_ppmw      = 2.01       # Nitrogen inventory in ppmw relative to mantle mass
+        NH_ratio    = 0.018       # N/H mass ratio in mantle/atmosphere system
+        # N_ppmw      = 2.01       # Nitrogen inventory in ppmw relative to mantle mass
 
-        # SH_ratio    = 2.16       # S/H mass ratio in mantle/atmosphere system
-        S_ppmw      = 235.0     # Sulfur inventory in ppmw relative to mantle mass
+        SH_ratio    = 2.16       # S/H mass ratio in mantle/atmosphere system
+        # S_ppmw      = 235.0     # Sulfur inventory in ppmw relative to mantle mass
 
 
 [observe]
