### 2025-08-13 1285268
Handle case when Psurf=0
