### 2025-08-13 15e38f8
Update earth config
