### 2025-08-13 d36da65
Update earth config. Allow tests to run on draft PR
