+++ .github/workflows/tests.yaml
@@ -16,7 +16,7 @@ on:
 
 jobs:
   test:
-    if: github.event.pull_request.draft == false
+    # if: github.event.pull_request.draft == false
     name: Run Coverage and Tests
     strategy:
       matrix:
