### 2025-08-13 1f3697f
Changed termination critera from perihelion to time averaged seperation smaller than Roche limit. Added termination config to demo orbit config files. Fixed sign before offset in termination.py . New version warns user when planet is partially orbiting withing the Roche limit of star, but only terminates based on the time averaged orbital separation.
