+++ src/proteus/atmos_clim/dummy.py
@@ -116,6 +116,7 @@ def RunDummyAtm( dirs:dict, config:Config, hf_row:dict):
     output["p_xuv"]   = hf_row["P_surf"]
     output["R_xuv"]   = R_obs
     output["p_obs"]   = hf_row["P_surf"]
+    output["T_obs"]   = hf_row["T_surf"]
     output["ocean_areacov"] = 0.0
     output["ocean_maxdepth"]= 0.0
 
