### 2025-08-20 81385ae
Extract elemental mixing ratios from atmosphere profile
