+++ src/proteus/utils/constants.py
@@ -1,7 +1,6 @@
 # Physical, numerical, etc constants
 from __future__ import annotations
 
-from proteus.utils.helper import mol_to_ele
 
 # Astronomical constants
 L_sun           = 3.828e+26             # W, IAU definition
@@ -35,13 +34,13 @@ secs_per_day    = secs_per_hour * 24.0
 secs_per_year   = secs_per_day * 365.25
 
 # Values from Pierrehumbert (2010) workbook
-const_R = 8.31446261815324      # J K−1 mol−1
+const_R = 8.31446261815324   #Gas constant [J K−1 mol−1]
 const_h = 6.626075540e-34    #Planck's constant
 const_c = 2.99792458e8       #Speed of light
 const_k =1.38065812e-23      #Boltzman thermodynamic constant
 const_sigma = 5.67051196e-8  #Stefan-Boltzman constant
 const_G = 6.67428e-11        #Gravitational constant (2006 measurements)
-
+const_Nav = 6.02214076e23    #Avogadro's constant [mol-1]
 B_ein = 2.5
 
 # Supported gases
@@ -54,9 +53,6 @@ element_list = ["H", "O", "C", "N", "S", "Si", "Mg", "Fe", "Na"]
 
 # Masses of elements [kg mol-1], from https://iupac.qmul.ac.uk/AtWt/
 element_mmw = {"H":1.008000000e-03, "He":4.002000000e-03, "Li":6.940000000e-03, "Be":9.012000000e-03, "B":1.081000000e-02, "C":1.201100000e-02, "N":1.400700000e-02, "O":1.599900000e-02, "F":1.899800000e-02, "Ne":2.017970000e-02, "Na":2.298900000e-02, "Mg":2.430500000e-02, "Al":2.698100000e-02, "Si":2.808500000e-02, "P":3.097300000e-02, "S":3.206000000e-02, "Cl":3.545000000e-02, "Ar":3.995000000e-02, "K":3.909830000e-02, "Ca":4.007800000e-02, "Sc":4.495500000e-02, "Ti":4.786700000e-02, "V":5.094150000e-02, "Cr":5.199610000e-02, "Mn":5.493800000e-02, "Fe":5.584500000e-02, "Co":5.893300000e-02, "Ni":5.869340000e-02, "Cu":6.354600000e-02, "Zn":6.538000000e-02, "Ga":6.972300000e-02, "Ge":7.263000000e-02, "As":7.492100000e-02, "Se":7.897100000e-02, "Br":7.990400000e-02, "Kr":8.379800000e-02, "Rb":8.546780000e-02, "Sr":8.762000000e-02, "Y":8.890500000e-02, "Zr":9.122400000e-02, "Nb":9.290600000e-02, "Mo":9.595000000e-02, "Ru":1.010700000e-01, "Rh":1.029050000e-01, "Pd":1.064200000e-01, "Ag":1.078682000e-01, "Cd":1.124140000e-01, "In":1.148180000e-01, "Sn":1.187100000e-01, "Sb":1.217600000e-01, "Te":1.276000000e-01, "I":1.269040000e-01, "Xe":1.312930000e-01, "Cs":1.329050000e-01, "Ba":1.373270000e-01, "La":1.389050000e-01, "Ce":1.401160000e-01, "Pr":1.409070000e-01, "Nd":1.442420000e-01, "Sm":1.503600000e-01, "Eu":1.519640000e-01, "Gd":1.572500000e-01, "Tb":1.589250000e-01, "Dy":1.625000000e-01, "Ho":1.649300000e-01, "Er":1.672590000e-01, "Tm":1.689340000e-01, "Yb":1.730450000e-01, "Lu":1.74966
[…]
