### 2025-08-20 a4ae07e
Fix bug with cpl_visual not finding file. Get mixing ratios of elements from atmosphere calc.
