### 2025-08-20 2280407
input file to run toi561b grids + addition of tested parameters for the plots
