+++ input/planets/toi561b.toml
@@ -13,14 +13,16 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
         plot_mod    = 5         # Plotting frequency, 0: wait until completion | n: every n iterations
         plot_fmt    = "png"     # Plotting image file format, "png" or "pdf" recommended
         write_mod   = 1         # Write CSV frequency, 0: wait until completion | n: every n iterations
+        archive_mod = 0      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf   = true
 
     # time-stepping
     [params.dt]
         minimum      = 3e2    # yr, minimum time-step
         minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
-        maximum      = 1e11   # yr, maximum time-step
+        maximum      = 1e10   # yr, maximum time-step
         initial      = 1e3    # yr, inital step size
-        starspec     = 3e6    # yr, interval to re-calculate the stellar spectrum
+        starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
         starinst     = 1e2    # yr, interval to re-calculate the instellation
         method       = "adaptive"  # proportional | adaptive | maximum
 
@@ -72,11 +74,11 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
 
     # Physical parameters
     mass    = 0.806       # from Lacedelli et al., 2022 [M_sun]
-    age_ini = 0.05       # Gyr, model initialisation/start age
+    age_ini = 0.1         # Gyr, model initialisation/start age
 
     module  = "mors"
     [star.mors]
-        rot_pcntle = 50.0       # rotation percentile
+        rot_pcntle = 20.0       # rotation percentile
         rot_period = 'none'     # rotation period [days]
         tracks  = "spada"       # evolution tracks: spada | baraffe
         age_now = 11            # [Gyr] from Lacedelli et al., 2022
@@ -112,13 +114,13 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
 
 # Atmosphere - physics table
 [atmos_clim]
-    prevent_warming = true      # do not allow the planet to heat up
+    prevent_warming = false      # do not allow the planet to heat up
     surface_d       = 0.01      # m, conductive skin thickness
     surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
     cloud_enabled   = false     # enable water cloud radiative effects
     cloud_alpha     = 0.0       # condensate retention fraction (1 -> fully retained)
     surf_state      = "fixed"   # surface scheme: "mixed_layer" | "fixed" | "skin"
-    surf_greyalbedo = 0.2       # surface grey albedo
+    surf_greyalbedo = 0.1       # surface grey albedo
     albedo_pl       = 0.0   	# Enforced Bond albedo (do not use with `rayleigh = true`) from Lacedelli et al. 2022
     rayleigh        = true      # Enable rayleigh scattering
     tmp_minimum     = 0.5       # temperature floor on solver
@@ -130,7 +132,7 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
         spectral_group  = "Honeyside"   # which gas opacities to include
         spectral_bands  = "256"         # how many spectral bands?
-        num_levels      = 60            # Number of atmospheric grid levels
+        num_levels      = 50            # Number of atmospheric grid levels
         chemistry       = "none"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
[…]
