+++ input/demos/boreas.toml
@@ -79,7 +79,7 @@ version = "2.0"
     surf_greyalbedo  = 0.1      # surface grey albedo
     rayleigh        = false      # enable rayleigh scattering
 
-    module  = "agni"           # Which atmosphere module to use
+    module  = "dummy"           # Which atmosphere module to use
 
     [atmos_clim.dummy]
         gamma = 0.98
@@ -104,13 +104,9 @@ version = "2.0"
     module = "boreas"    # Which escape module to use
 
     [escape.boreas]
-        light_major = 'H'
-        heavy_major = 'O'
-        heavy_minor = 'CN'
         efficiency  = 0.1
         alpha_rec   = 2.6e-13
         sigma_XUV   = 1.89e-18
-        kappa_p     = 4e-3
 
     [escape.zephyrus]
         Pxuv        = 1e-4          # Pressure at which XUV radiation become opaque in the planetary atmosphere [bar]

[…]
