+++ src/proteus/escape/boreas.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 import boreas
 import numpy as np
 
-from proteus.utils.constants import element_list, gas_list
+from proteus.utils.constants import element_list
 from proteus.utils.helper import eval_gas_mmw
 
 if TYPE_CHECKING:
@@ -16,12 +16,14 @@ log = logging.getLogger("fwl."+__name__)
 
 # Supported gases
 BOREAS_GASES = ("H2O", "H2" , "O2" , "CO2", "CO" , "CH4", "N2" , "NH3")
+BOREAS_ELEMS = ('H','C','O','N')
 
 def run_boreas(config:Config, hf_row:dict):
     """Run BOREAS escape model.
 
     Calculates the mass loss rate of each element.
-    Updates the quantities in hf_row as appropriate.
+    Updates the quantities in hf_row as appropriate, including R_xuv.
+    P_xuv must then be calculated from R_xuv as required.
 
     Parameters
     ----------
@@ -31,7 +33,7 @@ def run_boreas(config:Config, hf_row:dict):
             Dictionary of helpfile variables, at this iteration only
     """
 
-    log.info("Running fractionated escape (BOREAS) ...")
+    log.info("Running fractionated escape (BOREAS)...")
 
     # Set parameters for escape
     params          = boreas.ModelParams()
@@ -47,15 +49,14 @@ def run_boreas(config:Config, hf_row:dict):
 
     # Set parameters from config provided by user
     for g in BOREAS_GASES:
-        kappa = getattr(config.escape.boreas,"kappa_"+g)
-        setattr(params, "kappa_"+g, kappa)
+        params.kappa[g] = getattr(config.escape.boreas,"kappa_"+g)
     params.sigma_EUV    = config.escape.boreas.sigma_XUV
     params.alpha_rec    = config.escape.boreas.alpha_rec
     params.eff          = config.escape.boreas.efficiency
 
     # Set parameters from atmosphere calculation
     params.Teq       = hf_row["T_obs"]              # K
-    params.FEUV      = hf_row["F_xuv"] * 1e7        # XUV flux, converted to ergs cm-2 s-1
+    params.FEUV      = hf_row["F_xuv"] * 1e3        # XUV flux, converted to ergs cm-2 s-1
     params.rplanet   = hf_row["R_obs"] * 1e2        # convert m to cm
     params.mplanet   = hf_row["M_planet"] * 1e3     # convert kg to g
 
@@ -70,27 +71,29 @@ def run_boreas(config:Config, hf_row:dict):
     # Run fractionation calculation
     fr_result = fractionation.execute(ml_result, mass_loss)[0]
 
-    # Store bulk outputs
+    # Store bulk outputs (rate, sound speed, escape level)
     hf_row["esc_rate_total"] = fr_result["Mdot"]  * 1e-3    # g/s   ->  kg/s
-    hf_row["R_xuv"]          = fr_result["REUV"]  * 1e-2    # cm    ->  m
     hf_row["cs_xuv"]         = fr_result["cs"]    * 1e-2    # cm/s  ->  m/s
+    hf_row["R_xuv"]          = fr_result["REUV"]  * 1e-2    # cm    ->  m
+    hf_row["p_xuv"]          = 0.0  # to be calc'd by atmosphere module
 
     # Convert escape fluxes to rates, and store
     for e in element_list:
-        try:
+        if e in BOREAS_ELEMS:
             # convert g/cm2/s  ->  kg/m^2/s
-            flx = fr_result["Phi_"+e] * 10
+            flx = fr_result["phi_"+e] * 10
[…]
