+++ input/case_00002.toml
@@ -1,0 +1,274 @@
+version = "2.0"
+
+[params]
+resume = false
+offline = false
+[params.out]
+path = "SNS_stellar_spectra_tests/case_00002/"
+logging = "DEBUG"
+plot_fmt = "png"
+write_mod = 1
+plot_mod = 1
+archive_mod = "none"
+remove_sf = false
+
+[params.dt]
+starspec = 300000000.0
+starinst = 100.0
+method = "adaptive"
+minimum = 300.0
+minimum_rel = 1e-06
+maximum = 10000000.0
+initial = 1000.0
+
+[params.dt.proportional]
+propconst = 52.0
+
+[params.dt.adaptive]
+atol = 0.03
+rtol = 0.15
+
+[params.stop]
+strict = false
+[params.stop.iters]
+enabled = true
+minimum = 5
+maximum = 9000
+
+[params.stop.time]
+enabled = true
+maximum = 20000000000.0
+minimum = 1000.0
+
+[params.stop.solid]
+phi_crit = 0.01
+enabled = true
+
+[params.stop.radeqm]
+enabled = true
+atol = 0.2
+rtol = 0.001
+
+[params.stop.escape]
+enabled = false
+p_stop = 1
+
+[star]
+module = "mors"
+mass = 0.465
+age_ini = 0.1
+
+[star.mors]
+age_now = 3.0
+spec = "stellar_spectra/Named/gj849.txt"
+rot_pcntle = "none"
+rot_period = "none"
+tracks = "spada"
+
+[star.dummy]
+Teff = 3467.0
+radius = "none"
+calculate_radius = true
+
+[orbit]
+module = "none"
+semimajoraxis = 0.0278256
+eccentricity = 0.0
+zenith_angle = 54.74
+s0_factor = 0.25
+instellation_method = "sma"
+instellationflux = 27.0
+
+[orbit.dummy]
+H_tide = 1e-11
+Phi_tide = "<0.3"
+
+[orbit.lovepy]
+visc_thresh = 1000000000.0
+
+[struct]
+corefrac = 0.55
+module = "self"
+zalmoxis = "none"
+core_density = 10738.33
+core_heatcap = 880.0
+mass_tot = 5.0
+radius_int = "none"
+
+[atmos_clim]
+module = "agni"
+surf_state = "skin"
+prevent_warming = false
+surface_d = 0.01
+surface_k = 2.0
+cloud_enabled = false
+cloud_alpha = 0.0
+surf_greyalbedo = 0.0
+albedo_pl = 0.0
+rayleigh = true
+tmp_minimum = 0.5
+tmp_maximum = 5000.0
+
+[atmos_clim.agni]
+spectral_group = "Dayspring"
+spectral_bands = "48"
+p_top = 1e-05
+surf_material = "greybody"
+num_levels = 30
+chemistry = "none"
+solve_energy = true
+solution_atol = 0.01
+solution_rtol = 0.1
+overlap_method = "ee"
+condensation = true
+real_gas = true
+psurf_thresh = 0.1
+
+[atmos_clim.janus]
+spectral_group = "Dayspring"
+spectral_bands = "48"
+p_top = 1e-05
+p_obs = 0.001
+F_atm_bc = 0
+num_levels = 90
+tropopause = "none"
+overlap_method = "ee"
+
+[atmos_clim.dummy]
+gamma = 0.3
+
+[atmos_chem]
+module = "none"
+when = "manually"
+photo_on = true
+Kzz_on = true
+Kzz_const = "none"
+moldiff_on = true
+updraft_const = 0.0
+
+[atmos_chem.vulcan]
[…]
