+++ input/planets/fiducial_sub_Neptune.toml
@@ -103,7 +103,7 @@ version = "2.0"
 # Orbital system
 [orbit]
     instellation_method = 'inst'        # whether to define orbit using semi major axis ('sma') or instellation flux ('inst')
-    instellationflux    = 100          # instellation flux received from the planet in Earth units.
+    instellationflux    = 27            # instellation flux received from the planet in Earth units.
     semimajoraxis       = 0.15          # AU
     eccentricity        = 0.0       	# dimensionless
     zenith_angle        = 54.74     	# degrees
