### 2025-08-21 b48cc21
Running with BOREAS to fractionate outflow. Updat escape plot and config
