+++ src/proteus/escape/boreas.py
@@ -60,6 +60,7 @@ def run_boreas(config:Config, hf_row:dict):
     params.FEUV      = hf_row["F_xuv"] * 1e3        # XUV flux, converted to ergs cm-2 s-1
     params.rplanet   = hf_row["R_obs"] * 1e2        # convert m to cm
     params.mplanet   = hf_row["M_planet"] * 1e3     # convert kg to g
+    params.mmw_outflow_eff = None
 
     # Initalise objects
     mass_loss       = boreas.MassLoss(params)
