+++ src/proteus/plot/cpl_escape.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-from proteus.utils.constants import element_list, secs_per_year
+from proteus.utils.constants import element_list, secs_per_year, R_earth
 from proteus.utils.plot import get_colour
 
 if TYPE_CHECKING:
@@ -29,18 +29,21 @@ def plot_escape(hf_all:pd.DataFrame, output_dir:str, plot_format="pdf") :
     time = np.array(hf_crop["Time"])
 
     # mass unit
-    M_uval = 1e20     # kg
-    M_ulbl = r"$10^{20}$kg"
+    M_uval = 1e18     # kg
+    M_ulbl = r"$10^{18}$kg"
 
     # create plot
     lw = 1.2
     scale = 1.2
-    fig,axs = plt.subplots(2,1, figsize=(5*scale,5*scale), sharex=True)
+    fig,axs = plt.subplots(3,1, figsize=(5*scale,6*scale), sharex=True)
 
     # get axes
-    axt = axs[0]
-    axb = axs[1]
-    axr = axb.twinx()
+    ax0 = axs[0]   # inventories
+    ax1 = axs[1]   # rates
+    ax2 = axs[2]   # Psurf and Rxuv
+    axr = ax2.twinx()
+
+    ax0.set_xlim(left=0, right=np.amax(time)*1.01)
 
     # By element
     total = np.zeros(len(time))
@@ -54,48 +57,65 @@ def plot_escape(hf_all:pd.DataFrame, output_dir:str, plot_format="pdf") :
         # Plot planetary inventory of this element
         y = np.array(hf_crop[e+"_kg_total"])/M_uval
         total += y
-        axt.plot(time, y, lw=_lw, ls='dotted', color=col)
+        ax0.plot(time, y, lw=_lw, ls='dotted', color=col)
 
         # Plot atmospheric inventory of this element
         y = np.array(hf_crop[e+"_kg_atm"])/M_uval
-        axt.plot(time, y, lw=_lw, ls='solid', color=col, label=e)
+        ax0.plot(time, y, lw=_lw, ls='solid', color=col, label=e)
+
+        # Plot escape rate of this element
+        y = np.array(hf_crop["esc_rate_"+e]) * secs_per_year*1e6 / M_uval
+        ax1.plot(time, y, lw=_lw, ls='solid', color=col)
 
     # Planetary element sum inventory
-    axt.plot(time, total, lw=lw, ls='dotted',  label='Total',  c='k')
+    ax0.plot(time, total, lw=lw, ls='dotted',  label='Total',  c='k')
 
     # Atmosphere mass
     M_atm = np.array(hf_crop["M_atm"])/M_uval
-    axt.plot(time, M_atm, lw=lw, ls='solid', label='Atm.', c='k')
+    ax0.plot(time, M_atm, lw=lw, ls='solid', label='Atm.', c='k')
 
     # Decorate top plot
-    axt.set_ylabel(r"Mass [%s]"%M_ulbl)
-    axt.set_yscale("symlog", linthresh=1e-4)
-    axt.legend(loc='upper left', bbox_to_anchor=(1.0, 1.02), labelspacing=0.2)
+    ax0.set_ylabel(r"Mass [%s]"%M_ulbl)
+    ax0.set_yscale("symlog", linthresh=1e-3)
+    ax0.legend(loc='upper left', bbox_to_anchor=(1.0, 0.99), labelspacing=0.2)
+    ax0.set_ylim(0, np.amax(total)*1.5)
 
-    # Plot escape rate (kg / yr)
[…]
