+++ src/proteus/config/_escape.py
@@ -71,33 +71,33 @@ class EscapeBoreas:
     sigma_XUV: float
         Absorption cross-section in XUV [cm2 molecule-1]
     kappa_H2O: float
-        H2O opacity in XUV [cm2 g-1]
+        Grey H2O opacity in IR [cm2 g-1]
     kappa_H2: float
-        H2 opacity in XUV [cm2 g-1]
+        Grey H2 opacity in IR [cm2 g-1]
     kappa_O2: float
-        O2 opacity in XUV [cm2 g-1]
+        Grey O2 opacity in IR [cm2 g-1]
     kappa_CO2: float
-        CO2 opacity in XUV [cm2 g-1]
+        Grey CO2 opacity in IR [cm2 g-1]
     kappa_CO: float
-        CO opacity in XUV [cm2 g-1]
+        Grey CO opacity in IR [cm2 g-1]
     kappa_CH4: float
-        CH4 opacity in XUV [cm2 g-1]
+        Grey CH4 opacity in IR [cm2 g-1]
     kappa_N2: float
-        N2 opacity in XUV [cm2 g-1]
+        Grey N2 opacity in IR [cm2 g-1]
     kappa_NH3: float
-        NH3 opacity in XUV [cm2 g-1]
+        Grey NH3 opacity in IR [cm2 g-1]
     """
     efficiency: float = field(default=0.1,      validator=(ge(0), le(1)))
     alpha_rec: float  = field(default=2.6e-13,  validator=ge(0))
     sigma_XUV: float  = field(default=1.89e-18, validator=ge(0))
-    kappa_H2:  float  = field(default=1e-2,   validator=ge(0))
-    kappa_H2O: float  = field(default=1e-0,   validator=ge(0))
-    kappa_O2:  float  = field(default=1e-0,   validator=ge(0))
-    kappa_CO2: float  = field(default=1e-0,   validator=ge(0))
-    kappa_CO:  float  = field(default=1e-0,   validator=ge(0))
-    kappa_CH4: float  = field(default=1e-0,   validator=ge(0))
-    kappa_N2:  float  = field(default=1e-0,   validator=ge(0))
-    kappa_NH3: float  = field(default=1e-0,   validator=ge(0))
+    kappa_H2:  float  = field(default=1e-2,     validator=ge(0))
+    kappa_H2O: float  = field(default=1e-0,     validator=ge(0))
+    kappa_O2:  float  = field(default=1e-0,     validator=ge(0))
+    kappa_CO2: float  = field(default=1e-0,     validator=ge(0))
+    kappa_CO:  float  = field(default=1e-0,     validator=ge(0))
+    kappa_CH4: float  = field(default=1e-0,     validator=ge(0))
+    kappa_N2:  float  = field(default=1e-0,     validator=ge(0))
+    kappa_NH3: float  = field(default=1e-0,     validator=ge(0))
 
 def valid_reservoir(instance, attribute, value):
 
