+++ input/demos/boreas.toml
@@ -127,7 +127,7 @@ version = "2.0"
         ini_tmagma = 3000.0
 
 [outgas]
-    fO2_shift_IW    = 5         # log10(ΔIW), atmosphere/interior boundary oxidation state
+    fO2_shift_IW    = -4         # log10(ΔIW), atmosphere/interior boundary oxidation state
 
     module = "calliope"         # Which outgassing module to use
 
