### 2025-08-21 49695a8
Update escape plot. Reinit opacities for boreas calc
