+++ src/proteus/grid/run_grid_analysis.py
@@ -8,8 +8,25 @@
 # update the related plotting variables accordingly. This can be done in the `run_grid_analyze` function (see below).
 from __future__ import annotations
 
-from proteus.grid.post_processing_grid import *
+import os
+
 import matplotlib.cm as cm
+
+from proteus.grid.post_processing_grid import (
+    ecdf_grid_plot,
+    ecdf_single_plots,
+    extract_grid_output,
+    extract_solidification_time,
+    get_grid_parameters,
+    group_output_by_parameter,
+    load_extracted_data,
+    load_grid_cases,
+    plot_dir_exists,
+    plot_grid_status,
+    save_error_running_cases,
+    save_grid_data_to_csv,
+)
+
 
 def run_grid_analyze(path_to_grid: str, grid_name: str, update_csv: bool = True):
     """
@@ -129,20 +146,20 @@ def run_grid_analyze(path_to_grid: str, grid_name: str, update_csv: bool = True)
         'M_planet':            {"label": r"M$_p$ [M$_\oplus$]",             "log_scale": False,  "scale": 1.0/5.9722e24},
         'H_kg_atm':            {"label": r"[H$_{atm}$] [kg]",               "log_scale": True,  "scale": 1.0}
         }
-    #ecdf_single_plots(grid_params=grid_params, grouped_data=grouped_data, param_settings=param_settings_single, output_settings=output_settings_single, plots_path=plots_path)
+    ecdf_single_plots(grid_params=grid_params, grouped_data=grouped_data, param_settings=param_settings_single, output_settings=output_settings_single, plots_path=plots_path)
 
     # ECDF Grid Plot
     # The user needs to comment the parameters he didn't used in the grid/ add the ones non-listed here. Same for the outputs.
     param_settings_grid = {
-        "atmos_clim.module":          {"label": "Atmosphere module",                      "colormap": cm.rainbow,    "log_scale": False},
-        "orbit.semimajoraxis":        {"label": "a [AU]",                                 "colormap": cm.plasma,   "log_scale": False},
+        #"atmos_clim.module":          {"label": "Atmosphere module",                      "colormap": cm.rainbow,    "log_scale": False},
+        #"orbit.semimajoraxis":        {"label": "a [AU]",                                 "colormap": cm.plasma,   "log_scale": False},
         "escape.zephyrus.efficiency": {"label": r"$\epsilon$",                            "colormap": cm.spring,   "log_scale": False},
-        "escape.zephyrus.Pxuv":       {"label": r"$P_{XUV}$ [bar]",                       "colormap": cm.cividis,  "log_scale": True},
-        "outgas.fO2_shift_IW":        {"label": r"$\log_{10}(fO_2 / IW)$",                "colormap": cm.coolwarm, "log_scale": False},
+        #"escape.zephyrus.Pxuv":       {"label": r"$P_{XUV}$ [bar]",                       "colormap": cm.cividis,  "log_scale": True},
+        #"outgas.fO2_shift_IW":        {"label": r"$\log_{10}(fO_2 / IW)$",                "colormap": cm.coolwarm, "log_scale": False},
         "delivery.elements.CH_ratio": {"label": "C/H ratio",                              "colormap": cm.copper,   "log_scale": False},
         "delivery.elements.H_oceans": {"label": "[H] [oceans]",                           "colormap": cm.winter,   "log_scale": True},
-        #"delivery.elements.SH_ratio": {"label": "S/H ratio",                              "colormap": cm.autumn,   "log_scale": False},
-        #"escape.reservoir":           {"label": "Reservoir",                              "colormap": cm.viridis,  "log_scale": False}
[…]
