### 2025-08-27 9d8d21d
update files for more efficient grid runs
