+++ input/ensembles/escape_comparison_on_off.toml
@@ -13,8 +13,8 @@ ref_config = "input/demos/escape_comparison.toml"
 use_slurm = true
 
 # Execution limits
-max_jobs = 2       # maximum number of concurrent tasks (e.g. 500 on Habrok)
-max_days = 3        # maximum number of days to run (e.g. 1)
+max_jobs = 5       # maximum number of concurrent tasks (e.g. 500 on Habrok)
+max_days = 5        # maximum number of days to run (e.g. 1)
 max_mem  = 3        # maximum memory per CPU in GB (e.g. 3)
 
 # Now define grid axes...
@@ -23,7 +23,12 @@ max_mem  = 3        # maximum memory per CPU in GB (e.g. 3)
 #    Each table name must be written in double quotes.
 #    See examples below
 
-# Atmosphere module set directly
+# Escape module set directly
 ["escape.module"]
     method = "direct"
     values = ['none', 'zephyrus']
+
+# Atmosphere module set directly
+["atmos_clim.module"]
+    method = "direct"
+    values = ['janus', 'agni']
