### 2025-08-28 8a25af0
changes for toi561b grid with correct stellar spectrum
