### 2025-08-29 d97d510
Cleaned up the code and added comments
