+++ src/proteus/orbit/wrapper.py
@@ -117,7 +117,7 @@ def update_rochelimit(hf_row:dict):
 
     hf_row["roche_limit"] = Rpl * (2 * Mst/Mpl)**(1.0/3)
 
-def update_energy_budget(hf_row:dict, dirs:dict, total_E_last):
+def update_energy_budget(hf_row:dict, dirs:dict, dt:float, total_E_last:float):
     '''
     Calculate total dissipated energy.
 
@@ -132,52 +132,55 @@ def update_energy_budget(hf_row:dict, dirs:dict, total_E_last):
             Current helpfile row
         dirs: dict
             Dictionary of directories.
+        dt : float
+            Time interval [yr]
+        total_E_last: float
+            Energy disipated by tides up till previous timestep.
     '''
 
-    from proteus.interior.spider import get_all_output_times, get_dict_surface_values_for_specific_time
-    current_time = get_all_output_times(dirs["output"])[-1]
+    # Time step
+    current_time = float(hf_row["Time"])
 
     if current_time <= 1:
         log.info("Skipping energy budget in first iterations")
         return 0
     else:
-        sim_time = get_all_output_times(dirs["output"])[-1]  # yr, as an integer value
-        last_sim_time = get_all_output_times(dirs["output"])[-2]  # yr, as an integer value
+        # Find previous_time from which to evolve orbit to current_time
+        previous_time = current_time - dt
 
+    # Define keys for data extraction
     keys_t = (('data', 'rho_s'), ('data', 'radius_s'), ('data', 'Htidal_s'))
 
-    # data_a = get_dict_surface_values_for_specific_time(keys_t, sim_time, indir=dirs["output"])
-    data_b = get_dict_surface_values_for_specific_time(keys_t, last_sim_time, indir=dirs["output"])
+    # Extract interior data
+    from proteus.interior.spider import get_dict_surface_values_for_specific_time
+    data_previous = get_dict_surface_values_for_specific_time(keys_t, previous_time, indir=dirs["output"])
+    area_current  = get_dict_surface_values_for_specific_time((('data', 'area_b'),), previous_time, indir=dirs["output"])[0]
 
-    area = get_dict_surface_values_for_specific_time((('data', 'area_b'),), last_sim_time, indir=dirs["output"])[0]
+    # Calculate power input by tides at current time
+    power_current = hf_row["F_tidal"] * area_current[0] # current power
 
-    data_a = hf_row["F_tidal"] * area[0] # current power
-    # data_b =
+    # Find the increase in energy dissipated by tides
+    if data_previous is not None and len(data_previous) > 0:
 
-    if data_b is not None and len(data_b) > 0:
-        #sep_a = np.abs(np.diff(data_a[1]))
-        #vol_a = 4/3 * np.pi * ((data_a[1][1:] + sep_a)**3 - (data_a[1][1:])**3)
+        # Calculate the volume of the interior shells
+        sep_b = np.abs(np.diff(data_previous[1]))
+        vol_b = 4/3 * np.pi * ((data_previous[1][1:] + sep_b)**3 - (data_previous[1][1:])**3)
 
-        #avg_rho_a = np.convolve(data_a[0], [0.5, 0.5], "valid")
-        #avg_Htide_a = np.convolve(data_a[2], [0.5, 0.5], "valid")
+        # Find the density and tidal power input in each shell
+        avg_rho_b = np.convolve(data_previous[0], [0.5, 0.5], "valid")
+        avg_Htide_b = np.convolve(data_previous[2], [0.5, 0.5], "valid")
 
-        sep_b = np.abs(np.diff(data_b[1]))
[…]
