### 2025-08-30 394e199
Aragog return total heat flux for Fint
