### 2025-09-01 641353c
First pass on "update-all"
1) This first pass on "update-all" has been ChatGPT generated based on "install-all"
2) Reformatting from Black.
