### 2025-09-01 02ca206
Rename tests for clarity. physical_agni->dummy_agni. physical->aragog_janus. Updated test data. Added comments to test files, so it's clear what they do
