+++ tests/integration/test_integration_dummy_agni.py
@@ -1,3 +1,4 @@
+# This test runs PROTEUS with dummy interior and AGNI atmosphere, then chemistry with VULCAN
 from __future__ import annotations
 
 import os
@@ -15,13 +16,13 @@ from proteus.plot.cpl_atmosphere_cbar import plot_atmosphere_cbar_entry
 from proteus.plot.cpl_chem_atmosphere import plot_chem_atmosphere
 from proteus.utils.coupler import ReadHelpfileFromCSV
 
-out_dir = PROTEUS_ROOT / 'output' / 'physical_agni'
-ref_dir = PROTEUS_ROOT / 'tests' / 'data' / 'integration' / 'physical_agni'
-config_path = PROTEUS_ROOT /'tests' / 'integration' / 'physical_agni.toml'
+out_dir = PROTEUS_ROOT / 'output' / 'dummy_agni'
+ref_dir = PROTEUS_ROOT / 'tests' / 'data' / 'integration' / 'dummy_agni'
+config_path = PROTEUS_ROOT /'tests' / 'integration' / 'dummy_agni.toml'
 
 
 @pytest.fixture(scope="module")
-def agni_run():
+def dummy_agni_run():
 
     # Run simulation
     runner = Proteus(config_path=config_path)
@@ -32,7 +33,7 @@ def agni_run():
 
     return runner
 
-def test_agni_run(agni_run):
+def test_dummy_agni_run(dummy_agni):
     '''
     Test that the AGNI run completes without error and produces correct output
     '''
@@ -51,14 +52,14 @@ def test_agni_run(agni_run):
     assert_frame_equal(hf_all, hf_ref, rtol=5e-3)
 
 @pytest.mark.dependency()
-def test_agni_archive(agni_run):
+def test_dummy_agni_archive(dummy_agni):
     '''
     Test that the archive files are able to be extracted and created again
 
     The archive files were automatically created during the run.
     '''
 
-    data_dir = agni_run.directories["output/data"]
+    data_dir = dummy_agni.directories["output/data"]
 
     # Check that the archive file exists
     assert os.path.isfile(os.path.join(data_dir, "data.tar"))
@@ -67,13 +68,13 @@ def test_agni_archive(agni_run):
     assert not os.path.isfile(os.path.join(data_dir, "0_atm.nc"))
 
     # Extract archives
-    agni_run.extract_archives()
+    dummy_agni.extract_archives()
 
     # Check that the extracted files now exist
     assert os.path.isfile(os.path.join(data_dir, "0_atm.nc"))
 
-@pytest.mark.dependency(depends=["test_agni_archive"])
-def test_agni_atmosphere(agni_run):
+@pytest.mark.dependency(depends=["test_dummy_agni_archive"])
+def test_dummy_agni_atmosphere(dummy_agni):
     '''
     Test that the modelled profiles match the reference data
     '''
@@ -87,7 +88,7 @@ def test_agni_atmosphere(agni_run):
     out = read_atmosphere(_out, extra_keys=fields)
 
     # Compare to config
-    assert len(out["tmpl"]) == agni_run.config.atmos_clim.agni.num_levels+1
+    assert len(out["tmpl"]) == dummy_agni.config.atmos_clim.agni.num_levels+1
     assert np.all(out["Kzz"] >= 0)
     assert np.all(out["rl"][:-1] - out["rl"][1:] > 0)
 
@@ -99,14 +100,14 @@ def test_agni_atmosphere(agni_run):
         assert_allclose(out[key], ref[key], rtol=1e-3,
                         err_msg=f"Key {key} does not match reference data")
 
-@pytest.mark.dependency(depends=["test_agni_archive"])
-def test_agni_offchem(agni_run):
[…]
