### 2025-09-01 115225f
Code style. Moved tests into common folder
