+++ tests/test_config.py
@@ -1,4 +1,4 @@
-# Test whether config parser works 
+# Test whether config parser works
 from __future__ import annotations
 
 from itertools import chain
