### 2025-09-02 0117b17
Append to helpfile on disk when possible
