### 2025-09-03 fb286e3
Added stopping criterion based on planet axial period, updated stopping criterion based on roche limit. Both criteria are now listed under the StopDisint class in the configuration.
