### 2025-09-03 c677297
Updated demo and all_options configuration files to include the new disint criteria.
