### 2025-09-03 669e95a
different kernles and acquisition functions, proper seeding, aligned step counts with data in print statements
