### 2025-09-04 b2165f3
Add support for 'fwl-zalmoxis' module in PROTEUS configuration and validation
