+++ pyproject.toml
@@ -15,7 +15,7 @@ authors = [
     {name="Laurent Soucasse",        email="l.soucasse@esciencecenter.nl"},
     {name="Mariana Sastre",          email="m.c.villamil.sastre@rug.nl"},
     {name="Emma Postolec",           email="e.n.postolec@rug.nl"},
-    {name="Flavia Pascal",           email="f.pascal@student.rug.nl"},
+    {name="Flavia Pascal",           email="fpascal@astro.rug.nl"},
     {name="Ben Riegler",             email="ben.riegler@tum.de"},
     {name="Hanno Spreeuw",           email="h.spreeuw@esciencecenter.nl"},
     {name="Dan J. Bower",            email="dbower@ethz.ch"},
@@ -50,6 +50,7 @@ dependencies = [
     "fwl-calliope>=25.05.01",
     "fwl-zephyrus>=25.03.11",
     "aragog>=v0.2.8-alpha",
+    "fwl-zalmoxis>=25.09.02",
     "cmcrameri",
     "juliacall",
     "matplotlib",
