### 2025-09-05 64deb33
Improved function description
