### 2025-09-05 9bda25c
Updated all_options configuration file to include default values
