### 2025-09-07 56fff6d
Update configuration and logging for Zalmoxis module - Updated fwl-zalmoxis dependency version in pyproject.toml - Added verbose parameter to Zalmoxis class in _struct.py - Enhanced logging in zalmoxis.py to include convergence status
