+++ pyproject.toml
@@ -50,7 +50,7 @@ dependencies = [
     "fwl-calliope>=25.05.01",
     "fwl-zephyrus>=25.03.11",
     "aragog>=v0.2.8-alpha",
-    "fwl-zalmoxis>=25.09.02",
+    "fwl-zalmoxis>=25.09.07",
     "cmcrameri",
     "juliacall",
     "matplotlib",
