### 2025-09-10 6098967
"psutil" may be handy when determining disk space
"psutil" may be handy when determining available disk space for installing proteus.
Using "psutil.disk_usage('.').free" for the current working directory.
20 GB is recommended.
When updating, we should require a lot less, perhaps 5 GB?
