### 2025-09-10 804671a
Increase default values for num_levels, max_iterations_outer, and tolerance_outer in Zalmoxis class
