+++ environment.yml
@@ -6,7 +6,6 @@ dependencies:
   - python=3.13
   - pip
   - pytest
-  - psutil
 
   # Non-Python dependencies (system-level)
   - wget
