### 2025-09-10 66a72fb
Increase num_levels, max_iterations_outer, and tolerance_outer for Zalmoxis in all_options.toml
