### 2025-09-10 5cd21f9
"wget" is now in "environment.yml"
so no need to install it separately.
