### 2025-09-10 a91e576
Add "update-all" option to documentation.
