### 2025-09-10 d680bdb
"$HOME/miniconda3" not default install folder
"$HOME/miniconda3" should not be the default folder for installing miniconda, unless you have plenty of disk space there.
