### 2025-09-11 adf8b82
Instructions for capped home directories
Fixes #503, i.e. accommodate PROTEUS installations on clusters with capped home directories, i.e. the disk quota on home directories are very stringent, e.g. 9 GB.
