### 2025-09-11 1100d86
This comment is not helpful
1) Downloading Python from the official website is not helpful in installing PROTEUS. Also "pyenv" is not sufficient, since there are a number of non-Python packages required for the user install, as listed in "environment.yml". "pyenv" cannot provide that.
2) Removed redundant blank line.
