### 2025-09-11 e62dc97
Fix "❌ Aborting installation"
Fixes CI failure:
"
Run proteus install-all --export-env
⚠️ You have 14.914 GB of disk space at your disposal.
   To be safe, make sure you have at least 20 GB of free disk space.
❌ Aborting installation — 'proteus install-all'.
Error: Process completed with exit code 1.
"

GH has 14.914 GB of free disk space, while we require 20 GB.
20 GB is actually more than needed since we do the disk space check as part of step 6 of the user install.
Conda has already been installed then, this takes about 9 GB.
PROTEUS has already been git cloned.
So 10 GB should be enough.
