### 2025-09-15 6653715
EOS bug fixed and Zalmoxis clean up (#504)
* Fixed Zalmoxis EOS_material_properties already downloaded bug

* Refactor Zalmoxis implementation: rename inner_mantle_mass_fraction to mantle_mass_fraction; clean up functions in zalmoxis.py and use imports instead from the module; remove unused earth_center_pressure constant.

* Add support for 'fwl-zalmoxis' module in PROTEUS configuration and validation

* Update configuration and logging for Zalmoxis module
- Updated fwl-zalmoxis dependency version in pyproject.toml
- Added verbose parameter to Zalmoxis class in _struct.py
- Enhanced logging in
[…]
