+++ input/all_options.toml
@@ -156,16 +156,16 @@ version = "2.0"
         coremassfrac                = 0.325     # core mass fraction [non-dim.]
         inner_mantle_mass_fraction  = 0         # inner mantle mass fraction [non-dim.]
         weight_iron_frac            = 0.325     # iron fraction in the planet [non-dim.]
-        num_levels                  = 100       # number of Zalmoxis radius layers
+        num_levels                  = 150       # number of Zalmoxis radius layers
         EOSchoice = "Tabulated:iron/silicate"   # iron/silicate for super-Earths, water for water planets with Earth-like rocky cores
-        max_iterations_outer        = 20        # max. iterations for the outer loop
-        tolerance_outer             = 1e-3      # tolerance for the outer loop
+        max_iterations_outer        = 100       # max. iterations for the outer loop
+        tolerance_outer             = 3e-3      # tolerance for the outer loop
         max_iterations_inner        = 100       # max. iterations for the inner loop
         tolerance_inner             = 1e-4      # tolerance for the inner loop
         relative_tolerance          = 1e-5      # relative tolerance for solve_ivp
         absolute_tolerance          = 1e-6      # absolute tolerance for solve_ivp
         target_surface_pressure     = 101325    # target surface pressure
-        pressure_tolerance          = 1e11      # tolerance surface pressure
+        pressure_tolerance          = 1e9       # tolerance surface pressure
         max_iterations_pressure     = 200       # max. iterations for the innermost loop
         pressure_adjustment_factor  = 1.1       # factor for adjusting the pressure in the innermost loop
 
