+++ src/proteus/utils/helper.py
@@ -10,52 +10,57 @@ import shutil
 
 import numpy as np
 
-log = logging.getLogger("fwl."+__name__)
+log = logging.getLogger('fwl.' + __name__)
+
 
 def get_proteus_dir():
-    '''
+    """
     Get absolute path to PROTEUS directory.
 
     This should be a directory containing `pyproject.toml`.
-    '''
+    """
 
     # Assuming that this file is in `PROTEUS/src/proteus/utils/`
     utils = os.path.dirname(os.path.abspath(__file__))
 
     # Work upwards from utils
-    root = os.path.abspath(os.path.join(utils,"..","..",".."))
+    root = os.path.abspath(os.path.join(utils, '..', '..', '..'))
 
     # Check that this path is reasonable
-    if "pyproject.toml" not in os.listdir(root):
+    if 'pyproject.toml' not in os.listdir(root):
         raise EnvironmentError(f"Cannot locate PROTEUS directory. Tried '{root}' ")
 
     return root
 
+
 def PrintSeparator():
-    log.info("===================================================")
+    log.info('===================================================')
     pass
+
 
 def PrintHalfSeparator():
-    log.info("---------------------------------------------------")
+    log.info('---------------------------------------------------')
     pass
 
-def multiple(a:int,b:int) -> bool:
-    '''
+
+def multiple(a: int, b: int) -> bool:
+    """
     Return true if a is an integer multiple of b. Otherwise, return false.
 
     This is a more robust version of the modulo operator, which can fail if a or b are None, or if b is 0.
-    '''
+    """
     if (a is None) or (b is None) or (b == 0):
         return False
     else:
-        return bool(a%b == 0)
+        return bool(a % b == 0)
 
-def mol_to_ele(mol:str):
-    '''
+
+def mol_to_ele(mol: str):
+    """
     Return the number of atoms of each element in a given molecule, as a dictionary
 
     https://codereview.stackexchange.com/a/232664
-    '''
+    """
 
     # Validate
     if not str(mol[0]).isupper:
@@ -77,29 +82,34 @@ def mol_to_ele(mol:str):
 
     return elems
 
+
 # String sorting inspired by natsorted
 def natural_sort(lst):
     def convert(text):
         return int(text) if text.isdigit() else text.lower()
+
     def alphanum_key(key):
-        return [convert(c) for c in re.split("([0-9]+)", key)]
-    return sorted(lst, key = alphanum_key)
+        return [convert(c) for c in re.split('([0-9]+)', key)]
+
+    return sorted(lst, key=alphanum_key)
+
 
 # Create a temporary folder
 def create_tmp_folder():
-    tmp_dir = "/tmp/proteus_%d/" % np.random.randint(int(1e12),int(1e13-1))
+    tmp_dir = '/tmp/proteus_%d/' % np.random.randint(int(1e12), int(1e13 - 1))
     if os.path.exists(tmp_dir):
-        shutil.rmtree(tmp_dir,ignore_errors=True)
+        shutil.rmtree(tmp_dir, ignore_errors=True)
     os.makedirs(tmp_dir)
     return tmp_dir
 
-def safe_rm(fpath:str):
-    '''
-    Safely remove a file or folder
-    '''
 
-    if fpath=="":
-        log.warning("Could not remove file at empty path")
+def safe_rm(fpath: str):
+    """
+    Safely remove a file or folder
+    """
+
+    if fpath == '':
+        log.warning('Could not remove file at empty path')
         return
 
     fpath = os.path.abspath(fpath)
@@ -108,72 +118,76 @@ def safe_rm(fpath:str):
             os.remove(fpath)
 
         elif os.path.isdir(fpath):
[…]
