### 2025-09-16 0d2750a
Revert "linting"
This reverts commit c89a927799ac80ff2750f10be4ba077e9c183871.
