+++ src/proteus/plot/cpl_sflux_cross.py
@@ -15,7 +15,7 @@ from proteus.utils.helper import find_nearest, natural_sort
 if TYPE_CHECKING:
     from proteus import Proteus
 
-log = logging.getLogger('fwl.' + __name__)
+log = logging.getLogger("fwl."+__name__)
 
 
 def planck_function(lam, T):
@@ -26,22 +26,22 @@ def planck_function(lam, T):
     erg s-1 cm-2 nm-1
     """
 
-    x = lam * 1.0e-9  # convert nm -> m
-    hc_by_kT = const_h * const_c / (const_k * T)
+    x = lam * 1.0e-9   # convert nm -> m
+    hc_by_kT = const_h*const_c / (const_k*T)
 
-    planck_func = 1.0 / ((x**5.0) * (np.exp(hc_by_kT / x) - 1.0))
-    planck_func *= 2 * const_h * const_c * const_c  #  w m-2 sr-1 s-1 m-1
+    planck_func = 1.0/( (x ** 5.0) * ( np.exp( hc_by_kT/ x) - 1.0 ) )
+    planck_func *= 2 * const_h * const_c * const_c #  w m-2 sr-1 s-1 m-1
     planck_func *= np.pi * 1.0e3 * 1.0e-9  # erg s-1 cm-2 nm-1
 
     return planck_func
 
 
 def plot_sflux_cross(
-    output_dir: str,
-    wl_targets: list | None = None,
-    modern_age: float = -1,
-    plot_format='pdf',
-):
+        output_dir: str,
+        wl_targets: list | None=None,
+        modern_age: float=-1,
+        plot_format="pdf",
+    ):
     """Plots stellar flux vs time, for a set of wavelengths.
 
     Note that this function will plot the flux from EVERY file it finds.
@@ -62,14 +62,14 @@ def plot_sflux_cross(
         wl_targets = [1.0, 12.0, 50.0, 121.0, 200.0, 400.0, 500.0, 2000.0]
 
     # Find and sort files
-    files = glob.glob(output_dir + '/data/*.sflux')
+    files = glob.glob(output_dir+"/data/*.sflux")
     files = natural_sort(files)
 
     if len(files) <= 1:
-        log.warning('Insufficient data to make plot_sflux_cross')
+        log.warning("Insufficient data to make plot_sflux_cross")
         return
 
-    log.info('Plot stellar flux (crossection)')
+    log.info("Plot stellar flux (crossection)")
 
     # Arrays for storing data over time
     time_t = []
@@ -77,11 +77,11 @@ def plot_sflux_cross(
     flux_t = []
     for f in files:
         # Load data
-        X = np.loadtxt(f, skiprows=1, delimiter='\t').T
+        X = np.loadtxt(f,skiprows=1,delimiter='\t').T
 
         # Parse data
         time = int(f.split('/')[-1].split('.')[0])
-        if time < 0:
+        if (time < 0):
             continue
         wave = X[0]
         flux = X[1]
@@ -96,67 +96,61 @@ def plot_sflux_cross(
     flux_t = np.array(flux_t)
 
     # Create figure
-    fig, ax = plt.subplots(1, 1)
+    fig,ax = plt.subplots(1,1)
 
-    ax.set_yscale('log')
-    ax.set_ylabel('Flux [erg / (s cm$^2$ nm)]')
+    ax.set_yscale("log")
+    ax.set_ylabel("Flux [erg / (s cm$^2$ nm)]")
 
-    ax.set_xscale('log')
-    ax.set_xlabel('Time [yr]')
-    ax.set_title('TOA flux versus time')
+    ax.set_xscale("log")
+    ax.set_xlabel("Time [yr]")
+    ax.set_title("TOA flux versus time")
 
[…]
