+++ src/proteus/star/wrapper.py
@@ -11,15 +11,14 @@ import numpy as np
 from proteus.utils.constants import AU, M_sun, R_sun, const_sigma, ergcm2stoWm2
 from proteus.utils.helper import UpdateStatusfile
 
-log = logging.getLogger('fwl.' + __name__)
+log = logging.getLogger("fwl."+__name__)
 
 if TYPE_CHECKING:
     from proteus import Proteus
     from proteus.config import Config
 
-
-def init_star(handler: Proteus):
-    """
+def init_star(handler:Proteus):
+    '''
     Star-related things to be done when the simulation begins.
     This includes:
         - Preparing the stellar model
@@ -29,9 +28,9 @@ def init_star(handler: Proteus):
     ----------
         handler : Proteus
             Proteus object instance
-    """
+    '''
 
-    log.info('Preparing stellar model')
+    log.info("Preparing stellar model")
 
     # Dummy star module does not require preparation
 
@@ -42,17 +41,18 @@ def init_star(handler: Proteus):
         # Path to the modern spectrum
         #   i.e. that observed by a telescope, or derived from various observations.
         #   This is what we download from OSF.
-        star_modern_path = os.path.join(
-            handler.directories['fwl'], handler.config.star.mors.spec
-        )
+        star_modern_path = os.path.join(handler.directories["fwl"],
+                                        handler.config.star.mors.spec)
 
         # Copy modern spectrum to output folder, for posterity.
-        star_backup_path = os.path.join(handler.directories['output/data'], '-1.sflux')
+        star_backup_path = os.path.join(handler.directories["output/data"], "-1.sflux")
         shutil.copyfile(star_modern_path, star_backup_path)
 
         match handler.config.star.mors.tracks:
+
             case 'spada':
-                age_now_Myr = handler.config.star.mors.age_now * 1000  # convert Gyr to Myr
+
+                age_now_Myr = handler.config.star.mors.age_now * 1000 # convert Gyr to Myr
 
                 # Get rotation period or percentile (one of these is None)
                 pcntle = handler.config.star.mors.rot_pcntle
@@ -72,12 +72,9 @@ def init_star(handler: Proteus):
 
                 # load and fit track data
                 try:
-                    handler.stellar_track = mors.Star(
-                        Mstar=handler.config.star.mass,
-                        Age=age_rot_Myr,
-                        percentile=pcntle,
-                        Prot=period,
-                    )
+                    handler.stellar_track = mors.Star(Mstar = handler.config.star.mass,
+                                                        Age = age_rot_Myr,
+                                                        percentile = pcntle, Prot=period)
                 except Exception as e:
                     UpdateStatusfile(handler.directories, 23)
                     raise e
@@ -90,8 +87,7 @@ def init_star(handler: Proteus):
 
                 # calculate other properties from modern spectrum
                 handler.star_props = get_spada_synthesis_properties(
-                    handler.stellar_track, age_now_Myr
-                )
+                                                handler.stellar_track, age_now_Myr )
 
             case 'baraffe':
                 # creates track data
[…]

[…]
