### 2025-09-16 d12d470
Fix error when Aragog/Zalmoxis not installed, by avoiding calls to import them. Update demo config. Fix keyerror when atmosphere escapes
