+++ input/ensembles/example.infer.toml
@@ -25,10 +25,12 @@ n_samples  = 1000     # Raw samples for acquisition optimization
 [parameters]
 "struct.mass_tot" = [0.7, 3.0]
 "struct.corefrac" = [0.3, 0.9]
-"delivery.elements.H_ppmw" = [6e3, 2e4]
+"delivery.elements.H_ppmw" = [1e3, 2e4]
 "outgas.fO2_shift_IW" = [-3.0, 5.0]
+"escape.dummy.rate" = [1e5, 1e7]
 
 # Target observables to match by optimisation
 [observables]
-"R_obs" = 7.9950245489e6
-"H2O_vmr" = 0.41
+"R_obs" = 6.4781639626e+6
+"H2O_vmr" = 7.5032517428e-1
+"P_surf" = 3.8198606763e+3
