### 2025-09-18 90d5bd5
A reasonable default path should be returned.
We shouldn't be raising an OSError.
