### 2025-09-18 99bea4f
Take account of pip cache sizes as well.
