### 2025-09-19 bef9dd6
Fix two linter warnings
1) Linter complained about the line "return Path(os.environ.get("FWL_DATA"))" with "Argument of type 'str | None' cannot be assigned to parameter 'args' of type 'StrPath' in function '__new__'". This fixes that.
2) Linter complained about the line "def append_to_shell_rc(var: str, value: str, shell: str = None) -> Path | None:" with "Expression of type 'None' cannot be assigned to parameter of type 'str' 'None' is not assignable to 'str'". This is also fixed now.
3) Reformatting from Black.
