### 2025-09-19 ec076eb
Updating the Python packages had been omitted
Thanks @nichollsh for this catch: the Python packages were not updated.
Overseen because it was a separate, previous step in the install process.
