### 2025-09-19 5221648
Return an absolute path, not a relative one.
