### 2025-09-24 b04cd7d
"update-all" should include updating input data
We can deploy "download_sufficient_data". That needs a config, but that we can extract from "input/all_options.toml" (default) or any other path to a .toml file that a user might want.
