### 2025-09-24 f5be9cf
Show message when additional data downloaded
Show message when additional data has been downloaded.
