### 2025-09-25 551627b
Option for scaling stellar flux to match observations
