+++ input/planets/l9859d.toml
@@ -14,14 +14,14 @@ version = "2.0"
     # time-stepping
     [params.dt]
         minimum      = 3e2    # absolute minimum time-step [years]
-        minimum_rel  = 1e-2   # relative minimum time-step [dimensionless]
+        minimum_rel  = 2e-2   # relative minimum time-step [dimensionless]
         maximum      = 1e8    # yr, maximum time-step
         initial      = 1e4    # yr, inital step size
         starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
         starinst     = 1e2    # yr, interval to re-calculate the instellation
         method       = "adaptive"  # proportional | adaptive | maximum
         [params.dt.adaptive]
-            atol         = 0.03  # Step size atol
+            atol         = 0.04  # Step size atol
             rtol         = 0.14   # Step size rtol
 
     [params.stop]
@@ -50,12 +50,13 @@ version = "2.0"
 
     mass    = 0.2923      # M_sun
     age_ini = 0.05     # Gyr, model initialisation/start age
+    bol_scale  = 1.1162
 
     module  = "mors"
     [star.mors]
         rot_period = 77.5       # rotation period
         tracks  = "spada" # evolution tracks: spada | baraffe
-        age_now = 2.8     # Gyr, current age of star used for scaling
+        age_now = 2.5     # Gyr, current age of star used for scaling
         spec    = "stellar_spectra/Named/l-98-59.txt" # stellar spectrum
 
 
@@ -71,7 +72,7 @@ version = "2.0"
 
 [struct]
     mass_tot    = 1.64       # M_earth
-    corefrac    = 0.2      # non-dim., radius fraction
+    corefrac    = 0.15      # non-dim., radius fraction
 
 [atmos_clim]
     prevent_warming = true     # do not allow the planet to heat up
@@ -83,15 +84,15 @@ version = "2.0"
     module  = "agni"           # Which atmosphere module to use
 
     [atmos_clim.agni]
-        p_top           = 2.0e-5        # bar, top of atmosphere grid pressure
+        p_top           = 4.0e-5        # bar, top of atmosphere grid pressure
         spectral_group  = "Honeyside"   # which gas opacities to include
         spectral_bands  = "48"         # how many spectral bands?
-        num_levels      = 40            # Number of atmospheric grid levels
+        num_levels      = 45            # Number of atmospheric grid levels
         chemistry       = "none"        # "none" | "eq"
         surf_material   = "surface_albedos/Hammond24/lunarmarebasalt.dat"    # surface material file for scattering
         solve_energy    = true         # solve for energy-conserving atmosphere profile
         solution_atol   = 2.0          # solver absolute tolerance
-        solution_rtol   = 0.13          # solver relative tolerance
+        solution_rtol   = 0.14          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
         condensation    = false         # volatile condensation
         real_gas        = false          # use real-gas equations of state
@@ -127,7 +128,7 @@ version = "2.0"
 
 
 [outgas]
-    fO2_shift_IW    = -5         # log10(ΔIW), atmosphere/interior boundary oxidation state
+    fO2_shift_IW    = -6         # log10(ΔIW), atmosphere/interior boundary oxidation state
     module = "calliope"         # Which outgassing module to use
 
 [delivery]
@@ -135,10 +136,10 @@ version = "2.0"
     module = "none"
 
     [delivery.elements]
-        H_ppmw      = 35000.0     # Hydrogen inventory in ppmw relative to mantle mass
+        H_ppmw      = 40000.0     # Hydrogen inventory in ppmw relative to mantle mass
[…]
