### 2025-10-01 e7b5462
better grid and time-stepping
