+++ src/proteus/interior/timestep.py
@@ -16,7 +16,7 @@ log = logging.getLogger("fwl."+__name__)
 
 # Constants
 LBAVG = 3      # Number of steps to average over
-SFINC = 1.5    # Scale factor for step size increase
+SFINC = 1.6    # Scale factor for step size increase
 SFDEC = 0.8    # Scale factor for step size decrease
 SMALL = 1e-8   # Small number
 
@@ -227,8 +227,8 @@ def next_step(config:Config, dirs:dict, hf_row:dict, hf_all:pd.DataFrame, step_s
 
         # Min step size
         dtminimum = config.params.dt.minimum
-        if step_sf > 0.99:
-            dtminimum += config.params.dt.minimum_rel * hf_row["Time"]
+        # if step_sf > 0.99:
+        dtminimum += config.params.dt.minimum_rel * hf_row["Time"]
         dtswitch = max(dtswitch, dtminimum)
 
     log.info("New time-step target is %.2e years" % dtswitch)
