+++ input/planets/l9859d.toml
@@ -6,8 +6,8 @@ version = "2.0"
     [params.out]
         path        = "l9859d"
         logging     = "INFO"
-        plot_mod    = 40      # Plotting frequency, 0: wait until completion | n: every n iterations
-        write_mod   = 2
+        plot_mod    = 100      # Plotting frequency, 0: wait until completion | n: every n iterations
+        write_mod   = 5
         archive_mod = 250
         remove_sf   = true
 
@@ -33,7 +33,7 @@ version = "2.0"
         # solidification
         [params.stop.solid]
             enabled  = true
-            phi_crit = 0.005
+            phi_crit = 0.01
 
         # radiative equilibrium
         [params.stop.radeqm]
@@ -114,7 +114,7 @@ version = "2.0"
     module = "spider"   # Which interior module to use
 
     [interior.spider]
-        num_levels      = 190       # Number of SPIDER grid levels
+        num_levels      = 160       # Number of SPIDER grid levels
         mixing_length   = 2         # Mixing length parameterization
         tolerance       = 1e-10      # absolute solver tolerance
         tolerance_rel   = 1e-9      # relative solver tolerance
@@ -128,7 +128,7 @@ version = "2.0"
 
 
 [outgas]
-    fO2_shift_IW    = -6         # log10(ΔIW), atmosphere/interior boundary oxidation state
+    fO2_shift_IW    = -4         # log10(ΔIW), atmosphere/interior boundary oxidation state
     module = "calliope"         # Which outgassing module to use
 
 [delivery]
