+++ input/planets/l9859d.grid.toml
@@ -1,7 +1,7 @@
 # Config file for running a grid of forward models
 
 # Path to output folder where grid will be saved (relative to PROTEUS output folder)
-output = "scratch/l98d_updated5/"
+output = "scratch/l98d_updated6/"
 
 # Make `output` a symbolic link to this absolute location. To disable: set to empty string.
 symlink = ""
@@ -13,7 +13,7 @@ ref_config = "input/planets/l9859d.toml"
 use_slurm = true
 
 # Execution limits
-max_jobs = 250       # maximum number of concurrent tasks (e.g. 500 on Habrok)
+max_jobs = 300       # maximum number of concurrent tasks (e.g. 500 on Habrok)
 max_days = 2        # maximum number of days to run (e.g. 1)
 max_mem  = 3        # maximum memory per CPU in GB (e.g. 3)
 
@@ -32,25 +32,25 @@ max_mem  = 3        # maximum memory per CPU in GB (e.g. 3)
 ["delivery.elements.H_ppmw"]
     method = "arange"
     start    = 10000
-    stop     = 40000
+    stop     = 35000
     step     = 5000
 
 # Mantle redox state set directly
 ["outgas.fO2_shift_IW"]
     method = "direct"
-    values = [-5.0, -4.5, -4.0, -3.5, -3.0]
+    values = [-5.0, -4.5, -4.0, -3.5, -3.0, -2.5]
 
 # Planet bulk S/H ratio set directly
 ["delivery.elements.SH_ratio"]
-    method = "arange"
-    start    = 2
-    stop     = 22
-    step     = 5
+    method = "linspace"
+    start    = 1
+    stop     = 16
+    count    = 5
 
 # Core radius fraction set by linspace
 ["struct.corefrac"]
     method = "direct"
-    values = [0.15, 0.20, 0.25, 0.30, 0.40]
+    values = [0.15, 0.20, 0.25, 0.30, 0.35]
 
 # Escape efficiency set by arange
 # ["escape.zephyrus.efficiency"]
