### 2025-10-01 d8caf5b
Permit small steps when SPIDER fails
