+++ src/proteus/interior/timestep.py
@@ -226,9 +226,8 @@ def next_step(config:Config, dirs:dict, hf_row:dict, hf_all:pd.DataFrame, step_s
         dtswitch = min(dtswitch, config.params.dt.maximum )
 
         # Min step size
-        dtminimum = config.params.dt.minimum
-        # if step_sf > 0.99:
-        dtminimum += config.params.dt.minimum_rel * hf_row["Time"]
+        dtminimum = config.params.dt.minimum # absolute
+        dtminimum += config.params.dt.minimum_rel * hf_row["Time"] * 0.01 # allow small steps
         dtswitch = max(dtswitch, dtminimum)
 
     log.info("New time-step target is %.2e years" % dtswitch)
