### 2025-10-01 5bbeddf
Update config to new boreas version
