### 2025-10-01 ebd6665
Add config for GJ 9827 d, and new config option
