### 2025-10-02 5dbd421
Option for passing MLT stability criterion via config
