+++ src/proteus/atmos_clim/agni.py
@@ -7,6 +7,7 @@ import os
 from typing import TYPE_CHECKING
 
 import numpy as np
+from juliacall import convert
 from juliacall import Main as jl
 from scipy.interpolate import PchipInterpolator
 
@@ -208,6 +209,7 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
                         fastchem_work = fc_dir,
                         real_gas = config.atmos_clim.agni.real_gas,
                         check_integrity = False, # don't check thermo files every time
+                        mlt_criterion = convert(jl.Char,config.atmos_clim.agni.mlt_criterion),
 
                         skin_d=config.atmos_clim.surface_d,
                         skin_k=config.atmos_clim.surface_k,
@@ -364,24 +366,26 @@ def _solve_energy(atmos, loops_total:int, dirs:dict, config:Config):
         log.info("Attempt %d" % attempts)
 
         # default parameters
-        linesearch  = 2
-        easy_start  = False
-        dx_max      = float(config.atmos_clim.agni.dx_max)
-        ls_increase = 1.01
-        perturb_all = True
-        max_steps   = 70
-        chem_type   = int(config.atmos_clim.agni.chemistry_int)
+        linesearch   = 2
+        easy_start   = False
+        dx_max       = float(config.atmos_clim.agni.dx_max)
+        ls_increase  = 1.01
+        ls_max_steps = 20
+        ls_min_scale = 1e-5
+        perturb_chem = False
+        perturb_all  = bool(config.atmos_clim.agni.perturb_all)
+        max_steps    = int(config.atmos_clim.agni.max_steps)
+        chem_type    = int(config.atmos_clim.agni.chemistry_int)
 
-        # first few iterations
+        # parameters during initial few iterations
         if loops_total < 3:
-            dx_max = 200.0
+            dx_max = 300.0
             ls_increase = 1.1
             max_steps   = 200
 
-        # very first iteration parameters
+        # parameters for the first iteration
         if loops_total == 0:
             easy_start  = True
-            dx_max      = 300.0
 
         # try different solver parameters if struggling
         if attempts == 2:
@@ -409,8 +413,10 @@ def _solve_energy(atmos, loops_total:int, dirs:dict, config:Config):
                             conv_rtol=float(config.atmos_clim.agni.solution_rtol),
 
                             ls_increase=float(ls_increase), ls_method=int(linesearch),
+                            ls_max_steps=int(ls_max_steps), ls_min_scale=float(ls_min_scale),
+
                             dx_max=float(dx_max), easy_start=easy_start,
-                            perturb_all=perturb_all,
+                            perturb_all=perturb_all, perturb_chem=perturb_chem,
 
                             save_frames=False, modplot=int(modplot),
                             plot_jacobian=plot_jacobian
