### 2025-10-03 615e7b4
Adjust required disk space for CI
