+++ src/proteus/config/_atmos_clim.py
@@ -30,10 +30,6 @@ def valid_agni(instance, attribute, value):
     if (not instance.agni.solve_energy) and (instance.surf_state == 'skin'):
         raise ValueError("Must set `agni.solve_energy=true` if using `surf_state='skin'`")
 
-    # cannot set condensation and chemistry at the same time
-    if instance.agni.chemistry and instance.agni.condensation:
-        raise ValueError("`atmos_clim.agni`: Cannot enable condensation and chemistry at the same time")
-
     if instance.agni.latent_heat and not instance.agni.condensation:
         raise ValueError("`atmos_clim.agni`: Must set `condensation=true` if setting `latent_heat=true`")
 
