### 2025-10-03 7e7ec72
Model stability with chem enabled
