+++ src/proteus/atmos_clim/agni.py
@@ -386,12 +386,15 @@ def _solve_energy(atmos, loops_total:int, dirs:dict, config:Config):
         # parameters for the first iteration
         if loops_total == 0:
             easy_start  = True
+            perturb_chem = True
 
         # try different solver parameters if struggling
         if attempts == 2:
             linesearch  = 1
             dx_max     *= 2.0
             ls_increase = 1.1
+            perturb_chem = True
+            perturb_all = True
 
         log.debug("Solver parameters:")
         log.debug("    ls_method=%d, easy_start=%s, dx_max=%.1f, ls_increase=%.2f"%(
