+++ input/planets/gj9827d.toml
@@ -7,10 +7,11 @@ version = "2.0"
 [params]
     # output files
     [params.out]
-        path        = "gj9827d_1"
-        logging     = "DEBUG"
+        path        = "gj9827d_5"
+        logging     = "INFO"
         plot_mod    = 50
         write_mod   = 5
+        archive_mod = 0
 
      # time-stepping
     [params.dt]
@@ -92,7 +93,7 @@ version = "2.0"
     [atmos_clim.agni]
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
         spectral_group  = "Honeyside"   # which gas opacities to include
-        spectral_bands  = "16"         # how many spectral bands?
+        spectral_bands  = "48"         # how many spectral bands?
         num_levels      = 45            # Number of atmospheric grid levels
         chemistry       = "eq"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
@@ -143,7 +144,7 @@ version = "2.0"
 
 # Outgassing - physics table
 [outgas]
-    fO2_shift_IW    = 0         # log10(ΔIW), atmosphere/interior boundary oxidation state
+    fO2_shift_IW    = 4         # log10(ΔIW), atmosphere/interior boundary oxidation state
     module = "calliope"         # Which outgassing module to use
 
      [outgas.calliope]
@@ -160,7 +161,7 @@ version = "2.0"
     # Set initial volatile inventory by planetary element abundances
     [delivery.elements]
         H_ppmw      = 2e4       # Hydrogen inventory in ppmw relative to mantle mass
-        CH_ratio    = 1.0       # C/H mass ratio in mantle/atmosphere system
+        CH_ratio    = 0.0016       # C/H mass ratio in mantle/atmosphere system
         NH_ratio    = 0.1       # N/H mass ratio in mantle/atmosphere system
         SH_ratio    = 2.0       # S/H mass ratio in mantle/atmosphere system
 
