+++ input/demos/boreas.toml
@@ -13,24 +13,24 @@ version = "2.0"
 
     # time-stepping
     [params.dt]
-        minimum      = 3e2    # yr, minimum time-step
+        minimum      = 1e2    # yr, minimum time-step
         maximum      = 8e7    # yr, maximum time-step
-        initial      = 1e4    # yr, inital step size
-        starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
+        initial      = 1e2    # yr, inital step size
+        starspec     = 1e10   # yr, interval to re-calculate the stellar spectrum
         starinst     = 100    # yr, interval to re-calculate the instellation
 
         method = 'maximum'
 
         [params.dt.adaptive]
-            atol         = 0.02   # Step size atol
-            rtol         = 0.07   # Step size rtol
+            atol         = 0.01   # Step size atol
+            rtol         = 0.06   # Step size rtol
 
     [params.stop]
 
         [params.stop.time]
             enabled = true
             minimum = 8e8     # yr, model will certainly run to t > minimum
-            maximum = 2e9     # yr, model will terminate when t > maximum
+            maximum = 1e9     # yr, model will terminate when t > maximum
 
         [params.stop.solid]
             enabled  = false
@@ -49,25 +49,25 @@ version = "2.0"
 [star]
 
     # Physical parameters
-    mass    = 1.0       # M_sun
-    age_ini = 0.100     # Gyr, model initialisation/start age
+    mass    = 0.1       # M_sun
+    age_ini = 0.10      # Gyr, model initialisation/start age
 
     module  = "mors"
     [star.mors]
-        rot_pcntle = 50.0       # rotation percentile
+        rot_pcntle = 20.0       # rotation percentile
         tracks  = "spada"   # evolution tracks: spada | baraffe
         age_now = 4.567     # Gyr, current age of star used for scaling
-        spec    = "stellar_spectra/Named/sun.txt" # stellar spectrum
+        spec    = "stellar_spectra/Named/trappist-1.txt" # stellar spectrum
 
 [orbit]
-    semimajoraxis   = 0.5       # AU
+    semimajoraxis   = 0.01       # AU
     eccentricity    = 0.0       # dimensionless
     zenith_angle    = 48.19     # degrees
     s0_factor       = 0.375     # dimensionless
     module  = "none"
 
 [struct]
-    mass_tot    = 1.0           # Total planet mass [M_earth]
+    mass_tot    = 0.9           # Total planet mass [M_earth]
     corefrac    = 0.55          # non-dim., radius fraction
 
 # Atmosphere - physics table
@@ -82,7 +82,7 @@ version = "2.0"
     module  = "dummy"           # Which atmosphere module to use
 
     [atmos_clim.dummy]
-        gamma = 0.98
+        gamma = 0.99
 
     [atmos_clim.agni]
         solve_energy = false
@@ -104,8 +104,8 @@ version = "2.0"
     module = "boreas"    # Which escape module to use
 
     [escape.boreas]
-        fractionate = false
-        efficiency = 0.1
+        fractionate = true
+        efficiency = 0.01
 
         alpha_rec = 2.6e-13
 
@@ -173,7 +173,7 @@ version = "2.0"
 
     # Set initial volatile inventory by partial pressures in atmosphere
     [delivery.elements]
-        H_oceans = 30.0
+        H_oceans = 200.0
         CH_ratio = 1.0
         SH_ratio = 2.0
         N_ppmw   = 3.0
