### 2025-10-03 6c378dd
Update all_options.toml to include all of the options
