+++ input/planets/gj9827d.toml
@@ -124,11 +124,11 @@ version = "2.0"
     module = "spider"   # Which interior module to use
 
     [interior.spider]
-        num_levels      = 180       # Number of SPIDER grid levels
+        num_levels      = 120       # Number of SPIDER grid levels
         mixing_length   = 2         # Mixing length parameterization
-        tolerance       = 1e-10      # absolute solver tolerance
-        tolerance_rel   = 1e-9      # relative solver tolerance
-        tsurf_atol      = 5.0      # tsurf_poststep_change
+        tolerance       = 1e-6      # absolute solver tolerance
+        tolerance_rel   = 1e-5      # relative solver tolerance
+        tsurf_atol      = 10.0      # tsurf_poststep_change
         tsurf_rtol      = 0.02      # tsurf_poststep_change_frac
         ini_entropy     = 3150.0    # Surface entropy conditions [J K-1 kg-1]
         ini_dsdr        = -4.698e-6 # Interior entropy gradient [J K-1 kg-1 m-1]
