+++ input/all_options.toml
@@ -11,17 +11,17 @@
 # ----------------------------------------------------
 
 # The general structure is:
-#   [root]     metadata
-#   [params]   parameters for code execution, output files, time-stepping, convergence
-#   [star]     stellar parameters, model selection
-#   [orbit]    planetary orbital parameters
-#   [struct]   planetary structure (mass, radius)
-#   [atmos]    atmosphere parameters, model selection
-#   [escape]   escape parameters, model selection
-#   [interior] magma ocean model selection and parameters
-#   [outgas]   outgassing parameters (fO2) and included volatiles
-#   [delivery] initial volatile inventory, and delivery model selection
-#   [observe]  synthetic observations
+#   [params]        parameters for code execution, output files, time-stepping, convergence
+#   [star]          stellar parameters, model selection
+#   [orbit]         planetary orbital parameters
+#   [struct]        planetary structure (mass, radius)
+#   [atmos_clim]    atmosphere climate parameters, model selection
+#   [atmos_chem]    atmosphere chemistry parameters, model selection
+#   [escape]        escape parameters, model selection
+#   [interior]      magma ocean model selection and parameters
+#   [outgas]        outgassing parameters (fO2) and included volatiles
+#   [delivery]      initial volatile inventory, and delivery model selection
+#   [observe]       synthetic observations
 
 # ----------------------------------------------------
 
@@ -37,6 +37,7 @@ version = "2.0"
         plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
         write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
         archive_mod = 5      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf = false    # Remove SOCRATES spectral file when simulation ends.
 
     # time-stepping
     [params.dt]
@@ -134,6 +135,11 @@ version = "2.0"
     evolve  = false             # whether to evolve the SMaxis and eccentricity
     module  = "none"            # module used to calculate tidal heating
 
+    axial_period = "none"       # planet's initial day length; will use orbital period if 'none'
+    satellite = false           # include satellite (moon)?
+    mass_sat = 7.347e+22        # mass of satellite [kg]
+    semimajoraxis_sat = 3e8     # initial SMA of satellite's orbit
+
     [orbit.dummy]
         H_tide  = 1e-7      # Fixed tidal power density [W kg-1]
         Phi_tide = "<0.3"   # Tidal heating applied when inequality locally satisfied
@@ -153,6 +159,7 @@ version = "2.0"
     module       = "self"       # self | zalmoxis
 
     [struct.zalmoxis]
+        verbose                     = false     # verbose printing?
         coremassfrac                = 0.325     # core mass fraction [non-dim.]
         inner_mantle_mass_fraction  = 0         # inner mantle mass fraction [non-dim.]
         weight_iron_frac            = 0.325     # iron fraction in the planet [non-dim.]
@@ -187,6 +194,7 @@ version = "2.0"
 
     [atmos_clim.agni]
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
+        p_obs           = 0.02          # bar, level probed in transmission
         spectral_group  = "Honeyside"   # which gas opacities to include
         spectral_bands  = "48"          # how many spectral bands?
         num_levels      = 40            # Number of atmospheric grid levels
@@ -197,7 +205,13 @@ version = "2.0"
         solution_rtol   = 2e-2          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
         condensation    = true          # volatile condensation
+        latent_heat     = false         # include latent heat release by phase change?
[…]
