### 2025-10-03 88f7402
Solver options
