### 2025-10-05 b6f8d55
Update input/all_options.toml
Co-authored-by: Tim Lichtenberg <lichtenberg.tim@gmail.com>
