+++ input/all_options.toml
@@ -249,7 +249,7 @@ version = "2.0"
     tidal_heat      = false         # enable tidal heat production
     rheo_phi_loc    = 0.6           # Centre of rheological transition
     rheo_phi_wid    = 0.2           # Width of rheological transition
-    melting_dir     = "Monteux-600" # Name of folder constaining melting data
+    melting_dir     = "Monteux-600" # Name of folder constaining melting curves
 
 
     module = "aragog"       # Which interior module to use
