### 2025-10-05 24cad30
Default profile as linear
