### 2025-10-05 1657b52
Align all_options with minimal.toml config
