+++ input/minimal.toml
@@ -26,7 +26,7 @@ version = "2.0"
     eccentricity    = 0.0       # [dimensionless]
     zenith_angle    = 48.19     # [degrees]
     s0_factor       = 0.375     # [dimensionless]
-    module          = "dummy"
+    module          = "lovepy"
 
 [struct]
     mass_tot    = 1.0           # [M_earth]
@@ -36,8 +36,8 @@ version = "2.0"
     module  = "agni"            # Which atmosphere module to use
 
     [atmos_clim.agni]
-        spectral_group  = "Dayspring"   # Which gas opacities to include
-        spectral_bands  = "16"  # How many spectral bands
+        spectral_group  = "Honeyside"   # Which gas opacities to include
+        spectral_bands  = "48"  # How many spectral bands
 
 [escape]
     module = "zephyrus"
@@ -49,7 +49,7 @@ version = "2.0"
         ini_tmagma = 3300.0     # Initial magma surface temperature [K]
 
 [outgas]
-    fO2_shift_IW    = 2         # Atmosphere/interior boundary oxidation state [log10(ΔIW)]
+    fO2_shift_IW    = 4         # Atmosphere/interior boundary oxidation state [log10(ΔIW)]
     module = "calliope"         # Which outgassing module to use
 
 [delivery]
@@ -58,9 +58,9 @@ version = "2.0"
 
     # Set initial volatile inventory by planetary element abundances
     [delivery.elements]
-        H_oceans    = 1.0       # Hydrogen inventory in units of equivalent Earth oceans
+        H_oceans    = 5.0       # Hydrogen inventory in units of equivalent Earth oceans
         CH_ratio    = 1.0       # C/H mass ratio in mantle/atmosphere system
-        N_ppmw      = 1.0       # Nitrogen inventory in ppmw relative to mantle mass
+        NH_ratio    = 0.5       # N/H mass ratio in mantle/atmosphere system
         SH_ratio    = 2.0       # S/H mass ratio in mantle/atmosphere system
 
 # Calculate simulated observations
