### 2025-10-08 e020ebd
Start config for CHILI
