+++ input/chili_mip/protocol.toml
@@ -1,0 +1,325 @@
+# Configuration file used for CHILI Model Intercomparison Project
+
+version = "2.0"
+
+# Parameters
+[params]
+    # output files
+    [params.out]
+        path        = "chili_protocol"
+        logging     = "INFO"
+        plot_mod    = 3      # Plotting frequency, 0: wait until completion | n: every n iterations | none: do not plot
+        plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
+        write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
+        archive_mod = 5      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf = false    # Remove SOCRATES spectral file when simulation ends.
+
+    # time-stepping
+    [params.dt]
+        minimum      = 1e4    # absolute minimum time-step [years]
+        minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
+        maximum      = 1e7    # maximum time-step [yr]
+        initial      = 3e1    # inital step size [yr]
+        starspec     = 1e9    # interval to re-calculate the stellar spectrum [yr]
+        starinst     = 1e1    # interval to re-calculate the instellation [yr]
+        method       = "adaptive"  # proportional | adaptive | maximum
+
+        [params.dt.proportional]
+            propconst    = 52.0   # Proportionality constant
+
+        [params.dt.adaptive]
+            atol         = 0.02   # Step size atol
+            rtol         = 0.10   # Step size rtol
+
+    # Termination criteria
+    #     Set enabled=true/false in each section to enable/disable that termination criterion
+    [params.stop]
+
+        # Require criteria to be satisfied twice before model will exit?
+        strict = false
+
+        # required number of iterations
+        [params.stop.iters]
+            enabled = true
+            minimum = 5
+            maximum = 9000
+
+        # required time constraints
+        [params.stop.time]
+            enabled = true
+            minimum = 1.0e3      # model will certainly run to t > minimum [yr]
+            maximum = 7.0e+9   # model will terminate when t > maximum [yr]
+
+        # solidification
+        [params.stop.solid]
+            enabled  = true
+            phi_crit = 0.01      # non-dim., model will terminate when global melt fraction < phi_crit
+
+        # radiative equilibrium
+        [params.stop.radeqm]
+            enabled = false
+            atol    = 0.2        # absolute tolerance [W m-2]
+            rtol    = 1e-3       # relative tolerance
+
+        [params.stop.escape]
+            enabled = false
+            p_stop  = 5.0        # Stop surface pressure is less than this value
+
+        # disintegration
+        [params.stop.disint]
+            enabled   = false
+
+            roche_enabled = true
+            offset_roche  = 0 # correction to calculated Roche limit [m]
+
+            spin_enabled  = true
+            offset_spin   = 0 # correction to calculated Breakup period [s]
+
+
+# ----------------------------------------------------
+# Star
+[star]
+
+    # Physical parameters
+    mass    = 1.0       # stellar mass [M_sun]
+    age_ini = 0.100     # model initialisation age [Gyr]
+
+    module  = "mors"
+
+    [star.mors]
+        rot_pcntle  = 50.0      # rotation percentile
+        rot_period  = "none"    # rotation period [days]
+        tracks      = "bhac"   # evolution tracks: spada | baraffe
+        age_now     = 4.567     # current age of star [Gyr]
+        spec        = "stellar_spectra/Named/sun.txt" # path to stellar spectrum file
+
[…]
