### 2025-10-08 c88a497
Allow setting H_kg and C_kg directly
