+++ input/chili/protocol/earth.toml
@@ -6,13 +6,13 @@ version = "2.0"
 [params]
     # output files
     [params.out]
-        path        = "chili_protocol"
+        path        = "chili_protocol_earth"
         logging     = "INFO"
-        plot_mod    = 3      # Plotting frequency, 0: wait until completion | n: every n iterations | none: do not plot
+        plot_mod    = 5      # Plotting frequency, 0: wait until completion | n: every n iterations | none: do not plot
         plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
         write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
-        archive_mod = 5      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
-        remove_sf = false    # Remove SOCRATES spectral file when simulation ends.
+        archive_mod = 50      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf   = true    # Remove SOCRATES spectral file when simulation ends.
 
     # time-stepping
     [params.dt]
@@ -82,14 +82,14 @@ version = "2.0"
 
     # Physical parameters
     mass    = 1.0       # stellar mass [M_sun]
-    age_ini = 0.100     # model initialisation age [Gyr]
+    age_ini = 0.050     # model initialisation age [Gyr]
 
     module  = "mors"
 
     [star.mors]
         rot_pcntle  = 50.0      # rotation percentile
         rot_period  = "none"    # rotation period [days]
-        tracks      = "bhac"   # evolution tracks: spada | baraffe
+        tracks      = "spada"   # evolution tracks: spada | baraffe
         age_now     = 4.567     # current age of star [Gyr]
         spec        = "stellar_spectra/Named/sun.txt" # path to stellar spectrum file
 
@@ -106,7 +106,7 @@ version = "2.0"
     s0_factor = 0.375           # instellation scale factor [dimensionless]
 
     evolve  = false             # whether to evolve the SMaxis and eccentricity
-    module  = "lovepy"          # module used to calculate tidal heating
+    module  = "dummy"          # module used to calculate tidal heating
 
     axial_period = "none"       # planet's initial day length [hours]; will use orbital period if 'none'
     satellite = false           # include satellite (moon)?
@@ -132,7 +132,7 @@ version = "2.0"
 
 # Atmosphere - physics table
 [atmos_clim]
-    prevent_warming = true      # do not allow the planet to heat up
+    prevent_warming = false      # do not allow the planet to heat up
     surface_d       = 0.01      # m, conductive skin thickness
     surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
     cloud_enabled   = false     # enable water cloud radiative effects
@@ -150,16 +150,16 @@ version = "2.0"
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
         p_obs           = 0.02          # bar, level probed in transmission
         spectral_group  = "Honeyside"   # which gas opacities to include
-        spectral_bands  = "48"          # how many spectral bands?
-        num_levels      = 40            # Number of atmospheric grid levels
+        spectral_bands  = "16"          # how many spectral bands?
+        num_levels      = 42            # Number of atmospheric grid levels
         chemistry       = "none"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
         solve_energy    = true          # solve for energy-conserving atmosphere profile
         solution_atol   = 0.5           # solver absolute tolerance
         solution_rtol   = 0.15          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
[…]
