### 2025-10-08 1baebd2
Update config for chili
