### 2025-10-09 3c650c6
Add script for postprocessing CHILI MIP outputs
