### 2025-10-09 0cad17c
Model stability. CHILI configs
