### 2025-10-09 46489ba
Store global melt volume fraction
