### 2025-10-09 3a3ba08
Merge mariana interpolate function
