### 2025-10-09 193d143
Calc phi_global_vol
