+++ input/chili/protocol/tr1b.toml
@@ -152,7 +152,7 @@ version = "2.0"
         spectral_group  = "Dayspring"   # which gas opacities to include
         spectral_bands  = "48"          # how many spectral bands?
         num_levels      = 40            # Number of atmospheric grid levels
-        chemistry       = "none"        # "none" | "eq"
+        chemistry       = "eq"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
         solve_energy    = true          # solve for energy-conserving atmosphere profile
         solution_atol   = 0.1           # solver absolute tolerance
