### 2025-10-09 0bb28aa
Output potential temperature. Fastchem floor set by config
