+++ src/proteus/atmos_clim/agni.py
@@ -207,6 +207,7 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
                         condensates=condensates,
                         use_all_gases=include_all,
                         fastchem_work = fc_dir,
+                        fastchem_floor = config.atmos_clim.agni.fastchem_floor,
                         real_gas = config.atmos_clim.agni.real_gas,
                         check_integrity = False, # don't check thermo files every time
                         mlt_criterion = convert(jl.Char,config.atmos_clim.agni.mlt_criterion),
