+++ src/proteus/interior/aragog.py
@@ -352,8 +352,9 @@ class AragogRunner():
             output["M_mantle_solid"] = (output["M_mantle"] *
                                         (1.0 - output["Phi_global"]))
 
-        # CHANGE ME
-        output["Phi_global_vol"]    = output["Phi_global"]
+        # TODO: CHANGE THESE TO USE REALISTIC CALCULATION FOR ARAGOG
+        output["Phi_global_vol"]    = float(output["Phi_global"])
+        output["T_pot"]             = float(output["T_magma"])
 
         # Calculate surface area
         radii = aragog_output.radii_km_basic * 1e3 # [m]
