+++ input/chili/protocol/earth.toml
@@ -166,6 +166,7 @@ version = "2.0"
         max_steps       = 70            # max steps allowed by solver during each iteration
         perturb_all     = true          # updated entire jacobian each step?
         mlt_criterion   = "l"           # MLT convection stability criterion; (l)edoux or (s)chwarzschild
+        fastchem_floor  = 100.0
 
     [atmos_clim.dummy]
         gamma           = 0.01          # atmosphere opacity between 0 and 1
