+++ input/chili/protocol/tr1b.toml
@@ -166,6 +166,7 @@ version = "2.0"
         max_steps       = 70            # max steps allowed by solver during each iteration
         perturb_all     = true          # updated entire jacobian each step?
         mlt_criterion   = "l"           # MLT convection stability criterion; (l)edoux or (s)chwarzschild
+        fastchem_floor  = 100.0
 
     [atmos_clim.dummy]
         gamma           = 0.01          # atmosphere opacity between 0 and 1
@@ -259,7 +260,7 @@ version = "2.0"
 
     # Set initial volatile inventory by planetary element abundances if [initial = 'elements']
     [delivery.elements]
-        C_kg   = 5.458e20       # Carbon inventory  kg of C
+        C_kg   = 5.458e20       # Carbon inventory
         H_kg   = 1.119e20       # Hydrogen inventory
         N_kg   = 0.0       # Nitrogen inventory
         S_kg   = 0.0       # Sulfur inventory
