+++ input/chili/protocol/tr1b.toml
@@ -151,8 +151,8 @@ version = "2.0"
         p_obs           = 0.02          # bar, level probed in transmission
         spectral_group  = "Dayspring"   # which gas opacities to include
         spectral_bands  = "48"          # how many spectral bands?
-        num_levels      = 40            # Number of atmospheric grid levels
-        chemistry       = "eq"        # "none" | "eq"
+        num_levels      = 42            # Number of atmospheric grid levels
+        chemistry       = "none"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
         solve_energy    = true          # solve for energy-conserving atmosphere profile
         solution_atol   = 0.1           # solver absolute tolerance
@@ -232,9 +232,9 @@ version = "2.0"
         include_SO2  = false     # Include SO2 compound
         include_H2S  = false     # Include H2S compound
         include_NH3  = false     # Include NH3 compound
-        include_H2   = false     # Include H2 compound
-        include_CH4  = false     # Include CH4 compound
-        include_CO   = false     # Include CO compound
+        include_H2   = true     # Include H2 compound
+        include_CH4  = true     # Include CH4 compound
+        include_CO   = true     # Include CO compound
         T_floor      = 700.0    # Temperature floor applied to outgassing calculation [K].
         rtol         = 0.0001   # Relative mass tolerance
         xtol         = 1e-06    # Absolute mass tolerance
