### 2025-10-09 c585655
add plot in chili_postproc.py
