+++ tools/chili_postproc.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 import os
 import sys
 
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
@@ -28,7 +29,7 @@ hf_all = pd.read_csv(hfpath, delimiter=r'\s+')
 out = {}
 out["t(yr)"]            = np.array(hf_all["Time"].iloc[:])
 out["Tsurf(K)"]         = np.array(hf_all["T_pot"].iloc[:])
-out["Tpot(K)"]         = np.array(hf_all["T_surf"].iloc[:])
+out["Tpot(K)"]          = np.array(hf_all["T_surf"].iloc[:])
 out["phi(vol_frac)"]    = np.array(hf_all["Phi_global_vol"].iloc[:])
 out["massC_solid(kg)"]  = np.array(hf_all["C_kg_solid"].iloc[:])
 out["massC_melt(kg)"]   = np.array(hf_all["C_kg_liquid"].iloc[:])
@@ -48,3 +49,86 @@ outpath = os.path.join(simdir, "chili.csv")
 pd.DataFrame(out).to_csv(outpath, sep=',', index=False, float_format="%.10e")
 
 print("Done")
+
+
+# Plot results
+fig, axes = plt.subplots(2, 3, figsize=(20, 10), sharex=False)
+# Panel 1 : Surface Temperature
+axes[0, 0].plot(out["t(yr)"], out["Tsurf(K)"], linewidth=2)
+axes[0, 0].set_xlabel('Time [yr]', fontsize=14)
+axes[0, 0].set_ylabel(r'$\rm T_{surf}$ [K]', fontsize=14)
+axes[0, 0].set_xscale('log')
+axes[0, 0].tick_params(axis='both', labelsize=14)
+axes[0, 0].legend(loc='best', fontsize=14)
+# Panel 2 : Partial pressures
+axes[0, 1].plot(out["t(yr)"], out["pH2O(bar)"], label=r'$\rm H_2O$', color='royalblue', linewidth=2)
+axes[0, 1].plot(out["t(yr)"], out["pCO2(bar)"], label=r'$\rm CO_2$', color='red', linewidth=2)
+axes[0, 1].plot(out["t(yr)"], out["pCO(bar)"], label=r'$\rm CO$', color='green', linewidth=2)
+axes[0, 1].plot(out["t(yr)"], out["pH2(bar)"], label=r'$\rm H_2$', color='red', linewidth=2)
+axes[0, 1].plot(out["t(yr)"], out["pCH4(bar)"], label=r'$\rm CH_4$', color='purple', linewidth=2)
+axes[0, 1].plot(out["t(yr)"], out["pO2(bar)"], label=r'$\rm O_2$', color='brown', linewidth=2)
+axes[0, 1].set_xlabel('Time [yr]', fontsize=14)
+axes[0, 1].set_ylabel(r'$\rm p_{i}$ [bar]', fontsize=14)
+axes[0, 1].set_xscale('log')
+axes[0, 1].set_yscale('log')
+axes[0, 1].tick_params(axis='both', labelsize=14)
+axes[0, 1].legend(loc='best', fontsize=10)
+# Panel 3 : Transit radius
+axes[0, 2].plot(out["t(yr)"], out["Rtrans(Re)"], label='PROTEUS', linewidth=2)
+axes[0, 2].set_xlabel('Time [yr]', fontsize=14)
+axes[0, 2].set_ylabel(r'$\rm R_{trans}$ [R$_{\oplus}$]', fontsize=14)
+axes[0, 2].set_xscale('log')
+axes[0, 2].tick_params(axis='both', labelsize=14)
[…]
