+++ tools/chili_postproc.py
@@ -12,6 +12,7 @@ import numpy as np
 import pandas as pd
 
 from proteus.utils.constants import R_earth
+from proteus.utils.plot import get_colour, latexify
 
 # Simulation folder
 if len(sys.argv) != 2:
@@ -48,52 +49,42 @@ out["Rtrans(Re)"]       = np.array(hf_all["R_obs"].iloc[:]) / R_earth
 outpath = os.path.join(simdir, "chili.csv")
 pd.DataFrame(out).to_csv(outpath, sep=',', index=False, float_format="%.10e")
 
-print("Done")
-
-
-# Plot results
+print("Make plots")
 fig, axes = plt.subplots(2, 3, figsize=(20, 10), sharex=False)
+
 # Panel 1 : Surface Temperature
 axes[0, 0].plot(out["t(yr)"], out["Tsurf(K)"], linewidth=2)
 axes[0, 0].set_xlabel('Time [yr]', fontsize=14)
 axes[0, 0].set_ylabel(r'$\rm T_{surf}$ [K]', fontsize=14)
-axes[0, 0].set_xscale('log')
 axes[0, 0].tick_params(axis='both', labelsize=14)
-axes[0, 0].legend(loc='best', fontsize=14)
+
 # Panel 2 : Partial pressures
-axes[0, 1].plot(out["t(yr)"], out["pH2O(bar)"], label=r'$\rm H_2O$', color='royalblue', linewidth=2)
-axes[0, 1].plot(out["t(yr)"], out["pCO2(bar)"], label=r'$\rm CO_2$', color='red', linewidth=2)
-axes[0, 1].plot(out["t(yr)"], out["pCO(bar)"], label=r'$\rm CO$', color='green', linewidth=2)
-axes[0, 1].plot(out["t(yr)"], out["pH2(bar)"], label=r'$\rm H_2$', color='red', linewidth=2)
-axes[0, 1].plot(out["t(yr)"], out["pCH4(bar)"], label=r'$\rm CH_4$', color='purple', linewidth=2)
-axes[0, 1].plot(out["t(yr)"], out["pO2(bar)"], label=r'$\rm O_2$', color='brown', linewidth=2)
+for g in ["H2O", "CO2", "CO", "H2", "CH4", "O2"]:
+    axes[0, 1].plot(out["t(yr)"], out[f"p{g}(bar)"], label=latexify(g), color=get_colour(g), linewidth=2)
 axes[0, 1].set_xlabel('Time [yr]', fontsize=14)
 axes[0, 1].set_ylabel(r'$\rm p_{i}$ [bar]', fontsize=14)
-axes[0, 1].set_xscale('log')
 axes[0, 1].set_yscale('log')
 axes[0, 1].tick_params(axis='both', labelsize=14)
 axes[0, 1].legend(loc='best', fontsize=10)
+
 # Panel 3 : Transit radius
-axes[0, 2].plot(out["t(yr)"], out["Rtrans(Re)"], label='PROTEUS', linewidth=2)
+axes[0, 2].plot(out["t(yr)"], out["Rtrans(Re)"], linewidth=2)
 axes[0, 2].set_xlabel('Time [yr]', fontsize=14)
 axes[0, 2].set_ylabel(r'$\rm R_{trans}$ [R$_{\oplus}$]', fontsize=14)
-axes[0, 2].set_xscale('log')
 axes[0, 2].tick_params(axis='both', labelsize=14)
-axes[0, 2].legend(loc='best', fontsize=14)
+
 # Panel 4 : Potential temperature
-axes[1, 0].plot(out["t(yr)"], out["Tpot(K)"], label='PROTEUS', linewidth=2)
[…]
