### 2025-10-10 b888f57
Update configs. Store profiles
