+++ tools/chili_postproc.py
@@ -7,12 +7,19 @@ from __future__ import annotations
 import os
 import sys
 
+import matplotlib as mpl
+mpl.use("Agg") # noqa
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import netCDF4 as nc
 
-from proteus.utils.constants import R_earth
 from proteus.utils.plot import get_colour, latexify
+from proteus.utils.constants import R_earth, vol_list
+from proteus.config import read_config_object
+
+# Target times for sampling profile [years]
+tau_target = [1e3, 1e4, 1e5, 1e7, 1e9]
 
 # Simulation folder
 if len(sys.argv) != 2:
@@ -26,52 +33,82 @@ if not os.path.isfile(hfpath):
    raise FileNotFoundError(f"Cannot find {hfpath}")
 hf_all = pd.read_csv(hfpath, delimiter=r'\s+')
 
+# Read config
+config = read_config_object(os.path.join(simdir, "init_coupler.toml"))
+
 # Write to expected format
 out = {}
 out["t(yr)"]            = np.array(hf_all["Time"].iloc[:])
-out["Tsurf(K)"]         = np.array(hf_all["T_pot"].iloc[:])
-out["Tpot(K)"]          = np.array(hf_all["T_surf"].iloc[:])
+out["T_surf(K)"]        = np.array(hf_all["T_surf"].iloc[:])
+out["T_pot(K)"]         = np.array(hf_all["T_pot"].iloc[:])
 out["phi(vol_frac)"]    = np.array(hf_all["Phi_global_vol"].iloc[:])
 out["massC_solid(kg)"]  = np.array(hf_all["C_kg_solid"].iloc[:])
 out["massC_melt(kg)"]   = np.array(hf_all["C_kg_liquid"].iloc[:])
+out["massC_atm(kg)"]    = np.array(hf_all["C_kg_atm"].iloc[:])
 out["massH_solid(kg)"]  = np.array(hf_all["H_kg_solid"].iloc[:])
 out["massH_melt(kg)"]   = np.array(hf_all["H_kg_liquid"].iloc[:])
+out["massH_atm(kg)"]    = np.array(hf_all["H_kg_atm"].iloc[:])
 out["massO_atm(kg)"]    = np.array(hf_all["O_kg_atm"].iloc[:])
-out["massO_atm(kg)"]    = np.array(hf_all["O_kg_atm"].iloc[:])
-out["pH2O(bar)"]        = np.array(hf_all["H2O_bar"].iloc[:])
-out["pCO2(bar)"]        = np.array(hf_all["CO2_bar"].iloc[:])
-out["pCO(bar)"]         = np.array(hf_all["CO_bar"].iloc[:])
-out["pH2(bar)"]         = np.array(hf_all["H2_bar"].iloc[:])
-out["pCH4(bar)"]        = np.array(hf_all["CH4_bar"].iloc[:])
-out["pO2(bar)"]         = np.array(hf_all["O2_bar"].iloc[:])
-out["Rtrans(Re)"]       = np.array(hf_all["R_obs"].iloc[:]) / R_earth
+out["mmw(kg/mol)"]      = np.array(hf_all["atm_kg_per_mol"].iloc[:])
+out["p_surf(bar)"]      = np.array(hf_all["P_surf"].iloc[:])
+for gas in vol_list:
+    out[f"p_{gas}(bar)"]       = np.array(hf_all[f"{gas}_bar"].iloc[:])
+out["fO2_liquid(bar)"]  = out["p_O2(bar)"]
+out["fO2_solid(bar)"]   = np.zeros_like(out["t(yr)"])
[…]
