+++ input/chili/protocol/earth.toml
@@ -1,302 +1,312 @@
-# Configuration file used for CHILI Model Intercomparison Project
-
 version = "2.0"
 
-# Parameters
 [params]
-    # output files
-    [params.out]
-        path        = "chili_protocol_earth"
-        logging     = "INFO"
-        plot_mod    = 20      # Plotting frequency, 0: wait until completion | n: every n iterations | none: do not plot
-        plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
-        write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
-        archive_mod = 'none'      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
-        remove_sf   = true    # Remove SOCRATES spectral file when simulation ends.
+resume = false
+offline = false
 
-    # time-stepping
-    [params.dt]
-        minimum      = 1e4    # absolute minimum time-step [years]
-        minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
-        maximum      = 1e6    # maximum time-step [yr]
-        initial      = 3e1    # inital step size [yr]
-        starspec     = 1e7    # interval to re-calculate the stellar spectrum [yr]
-        starinst     = 1e1    # interval to re-calculate the instellation [yr]
-        method       = "adaptive"  # proportional | adaptive | maximum
+[params.out]
+path = "chili_protocol_earth"
+logging = "INFO"
+plot_fmt = "png"
+write_mod = 1
+plot_mod = 20
+archive_mod = "none"
+remove_sf = true
 
-        [params.dt.proportional]
-            propconst    = 52.0   # Proportionality constant
+[params.dt]
+starspec = 10000000.0
+starinst = 10.0
+method = "adaptive"
+minimum = 10000.0
+minimum_rel = 1e-05
+maximum = 10000000.0
+initial = 30.0
 
-        [params.dt.adaptive]
-            atol         = 0.02   # Step size atol
-            rtol         = 0.09   # Step size rtol
+[params.dt.proportional]
+propconst = 52.0
 
-    # Termination criteria
-    #     Set enabled=true/false in each section to enable/disable that termination criterion
-    [params.stop]
+[params.dt.adaptive]
+atol = 0.02
+rtol = 0.09
 
-        # Require criteria to be satisfied twice before model will exit?
-        strict = false
+[params.stop]
+strict = false
 
-        # required number of iterations
-        [params.stop.iters]
-            enabled = true
-            minimum = 5
-            maximum = 9000
+[params.stop.iters]
+enabled = true
+minimum = 5
+maximum = 9000
 
-        # required time constraints
-        [params.stop.time]
-            enabled = true
-            minimum = 1.0e3      # model will certainly run to t > minimum [yr]
-            maximum = 5.0e+9   # model will terminate when t > maximum [yr]
+[params.stop.time]
+enabled = true
+maximum = 5000000000.0
+minimum = 1000.0
 
-        # solidification
-        [params.stop.solid]
-            enabled  = true
-            phi_crit = 0.02      # non-dim., model will terminate when global melt fraction < phi_crit
+[params.stop.solid]
+phi_crit = 0.03
+enabled = true
 
-        # radiative equilibrium
-        [params.stop.radeqm]
-            enabled = false
-            atol    = 0.2        # absolute tolerance [W m-2]
-            rtol    = 1e-3       # relative tolerance
+[params.stop.radeqm]
+enabled = false
+atol = 0.2
[…]

[…]
