+++ CITATION.cff
@@ -62,7 +62,7 @@ authors:
   orcid: "https://orcid.org/0000-0002-5887-1197"
 
 title: "PROTEUS"
-version: 25.07.31
+version: 25.10.15
 doi: 10.0000/xxxx/xxxx
-date-released: 2025-07-31
+date-released: 2025-10-15
 url: "https://github.com/FormingWorlds/PROTEUS"
