### 2025-10-16 d7074ce
Config options for fastchem via agni
