+++ input/all_options.toml
@@ -212,7 +212,11 @@ version = "2.0"
         max_steps       = 70            # max steps allowed by solver during each iteration
         perturb_all     = true          # updated entire jacobian each step?
         mlt_criterion   = "l"           # MLT convection stability criterion; (l)edoux or (s)chwarzschild
-        
+        fastchem_floor        = 150.0   # Minimum temperature allowed to be sent to FC
+        fastchem_maxiter_chem = 60000   # Maximum FC iterations (chemistry)
+        fastchem_maxiter_solv = 20000   # Maximum FC iterations (internal solver)
+        fastchem_xtol_chem    = 1e-4    # FC solver tolerance (chemistry)
+        fastchem_xtol_elem    = 1e-4    # FC solver tolerance (elemental)
 
     [atmos_clim.janus]
         p_top           = 1.0e-6        # bar, top of atmosphere grid pressure
