### 2025-10-20 1a8d46f
Start with CLI commands
