### 2025-10-20 6e6fb73
Add test for grid_summarise. Fix syntax. Add CLI comments
