### 2025-10-20 14a8adf
Gridsummarise CLI works
