### 2025-10-20 a9713c0
update summarise return code. add pack script
