+++ src/proteus/grid/pack.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import os
 from glob import glob
 from pathlib import Path
-from shutil import copyfile, copytree, rmtree
+from shutil import copyfile, rmtree
 from zipfile import ZIP_DEFLATED, ZipFile
 
 
@@ -27,39 +27,51 @@ def pack(grid:str, plots:bool=True, zip:bool=True):
         raise FileNotFoundError("Cannot find any subfolders containing grid cases!")
 
     # copy top-level files in grid output folder
+    print("Copy top-level files...")
     for tf in ["manager.log", "ref_config.toml", "copy.grid.toml"]:
         copyfile(os.path.join(grid, tf), os.path.join(pack, tf))
 
     # copy per-case data
-    print("Copy results")
+    print("Copy results for each grid point...")
+    print(f"Found {len(case_dirs)} subfolders")
     for case in case_dirs:
-        num = int(os.path.dirname(case).split("_")[-1])
-
-        print(f"   {num:06d}")
-        dest = os.path.join(pack, os.path.dirname(case))
+        print("   "+os.path.basename(case))
+        dest = os.path.join(pack, os.path.basename(case))
         os.mkdir(dest)
 
         # lower level files
         llfs = ["runtime_helpfile.csv", "init_coupler.toml", "status"]
         llfs.extend([f"proteus_{i:02d}.log" for i in range(100)])
         for lf in llfs:
-            copyfile(os.path.join(case, lf), os.path.join(dest, lf))
+            try:
+                copyfile(os.path.join(case, lf), os.path.join(dest, lf))
+            except FileNotFoundError:
+                pass
 
         # plots directory
         if plots:
-            copytree(os.path.join(case, "plots"), os.path.join(dest,"plots"))
+            for pf in glob("plot_*",root_dir=os.path.join(case,"plots")):
+                copyfile(os.path.join(case,"plots",pf), os.path.join(dest,pf))
 
     # create zip at grid/pack.zip containing "pack/..."
-    zip_path = os.path.join(grid,"pack.zip")
-    print(f"Make zip: {zip_path}")
-    with ZipFile(zip_path, "w", compression=ZIP_DEFLATED) as zf:
-        for root, dirs, files in os.walk(pack):
-            root_path = Path(root)
-            for name in files:
-                file_path = root_path / name
-                # keep "pack" as top-level folder inside the zip
-                arcname = Path("pack") / file_path.relative_to(pack)
-                zf.write(file_path, arcname)
+    if zip:
+        zip_path = os.path.join(grid,"pack.zip")
+        print(f"Make zip: {zip_path}")
+        if os.path.isfile(zip_path):
+            os.remove(zip_path)
+        with ZipFile(zip_path, "w", compression=ZIP_DEFLATED) as zf:
+            for root, dirs, files in os.walk(pack):
+                root_path = Path(root)
+                for name in files:
+                    file_path = root_path / name
+                    # keep "pack" as top-level folder inside the zip
+                    arcname = Path("pack") / file_path.relative_to(pack)
+                    zf.write(file_path, arcname)
 
-    print("Done!")
+        # give size of file
+        print(f"Archive size is {os.path.getsize(zip_path)/1e6:.1f} MB")
+
+        # remove `pack` folder now that we have zipped it
+        rmtree(pack)
+
     return True
