### 2025-10-20 55ad6ef
Fix doctor script for getting non-python repos
