+++ src/proteus/__init__.py
@@ -2,6 +2,6 @@ from __future__ import annotations
 
 from .proteus import Proteus
 
-__version__ = '25.10.15'
+__version__ = '25.10.20'
 
 __all__ = ['Proteus']
