### 2025-10-20 6c7f643
Improve test coverage
