### 2025-10-21 2f1cadd
JANUS only receive positive values for Psurf
