### 2025-10-21 9e33fee
add a validator which requires the user enable the escape stopping condition if using JANUS
