+++ src/proteus/config/_config.py
@@ -52,6 +52,11 @@ def observe_resolved_atmosphere(instance, attribute, value):
     if (instance.observe.synthesis is not None) and (instance.atmos_clim.module == "dummy"):
         raise ValueError("Observational synthesis requires that atmos_clim != dummy")
 
+def janus_escape_atmosphere(instance, attribute, value):
+    # Using escape.zephyrus with JANUS requires params.stop.escape to be true
+    if (instance.escape.module == "zephyrus") and (instance.atmos_clim.module != "janus"):
+        raise ValueError("Using ZEHPYRUS escape module with JANUS atmosphere module requires params.stop.escape to be enabled")
+
 @define
 class Config:
     """Root config parameters.
