### 2025-10-21 821e5bc
add pstop > 0 fro the user
