### 2025-10-21 826c40c
UpdateStateAtm() update to include config in argument
