### 2025-10-22 9c6dd44
Make download from zenodo itself more robust
