### 2025-10-24 d006237
Add base config for chili
