+++ input/chili/readme.txt
@@ -1,0 +1,9 @@
+Configuration files for the CHILI model intercomparison project.
+Website: https://nexss.info/cuisines/chili-mip/
+
+The `protocol/` subfolder contains the two configurations used in the protocol paper.
+
+To ensure that the PROTEUS are consistent with each other for the main intercomparison,
+there is a single base configuration file `intercomp/_base.toml` from which planet-specific
+configuration files are generated in the `intercomp/` directory. These configurations may
+be generated by running the `chili_gen.py` script.
