+++ input/chili/intercomp/_base.toml
@@ -1,0 +1,237 @@
+version = "2.0"
+
+[params]
+resume = false
+offline = false
+
+[params.out]
+path = "chili_base"
+logging = "INFO"
+plot_fmt = "png"
+write_mod = 1
+plot_mod = 20
+archive_mod = "none"
+remove_sf = true
+
+[params.dt]
+starspec = 10000000.0
+starinst = 10.0
+method = "adaptive"
+minimum = 10000.0
+minimum_rel = 1e-05
+maximum = 10000000.0
+initial = 30.0
+[params.dt.adaptive]
+atol = 0.03
+rtol = 0.10
+
+[params.stop]
+strict = false
+
+[params.stop.iters]
+enabled = true
+minimum = 5
+maximum = 9000
+
+[params.stop.time]
+enabled = true
+maximum = 5000000000.0
+minimum = 1000.0
+
+[params.stop.solid]
+phi_crit = 0.05
+enabled = true
+
+[params.stop.radeqm]
+enabled = false
+atol = 0.2
+rtol = 0.001
+
+[params.stop.escape]
+enabled = false
+p_stop = 5.0
+
+[params.stop.disint]
+enabled = false
+roche_enabled = true
+offset_roche = 0.0
+spin_enabled = true
+offset_spin = 0.0
+
+[star]
+module = "mors"
+mass = 1.0
+age_ini = 0.05
+
+[star.mors]
+age_now = 4.567
+spec = "stellar_spectra/Named/sun.txt"
+rot_pcntle = 50.0
+rot_period = "none"
+tracks = "spada"
+
+[star.dummy]
+Teff = 5772.0
+radius = 1.0
+calculate_radius = false
+
+[orbit]
+module = "none"
+semimajoraxis = 1.0
+eccentricity = 0.0
+zenith_angle = 48.19
+s0_factor = 0.375
+evolve = false
+axial_period = "none"
+satellite = false
+mass_sat = 7.347e+22
+semimajoraxis_sat = 300000000.0
+instellation_method = "sma"
+instellationflux = 1.0
+
+[struct]
+corefrac = 0.55
+module = "self"
+core_density = 10738.33
+core_heatcap = 880.0
+mass_tot = 1.0
+radius_int = "none"
+
+[atmos_clim]
+module = "agni"
+surf_state = "fixed"
+prevent_warming = true
+surface_d = 0.01
+surface_k = 2.0
+cloud_enabled = false
+cloud_alpha = 0.0
+surf_greyalbedo = 0.0
+albedo_pl = 0.1
+rayleigh = false
+tmp_minimum = 0.5
+tmp_maximum = 5000.0
+
+[atmos_clim.agni]
+spectral_group = "Dayspring"
+spectral_bands = "48"
+p_top = 1e-05
+p_obs = 0.02
+surf_material = "greybody"
+num_levels = 42
+chemistry = "none"
+solve_energy = true
+solution_atol = 0.1
+solution_rtol = 0.1
+overlap_method = "ee"
+condensation = false
+latent_heat = false
+real_gas = false
+psurf_thresh = 0.1
+dx_max = 10.0
+max_steps = 70
+perturb_all = true
+mlt_criterion = "l"
+fastchem_floor = 100.0
+
+[atmos_clim.dummy]
+gamma = 0.01
+height_factor = 3.0
+
+[escape]
+module = "zephyrus"
+reservoir = "outgas"
+
+[escape.zephyrus]
+Pxuv = 5e-05
+efficiency = 0.1
+tidal = true
+
+[interior]
+module = "spider"
+radiogenic_heat = true
[…]
