### 2025-10-24 eaaf3be
Add PR template and update issue templates
