### 2025-10-24 3a920ba
Configuration script for chili. Base and grid configs. Gitignore
