+++ input/chili/readme.txt
@@ -6,4 +6,4 @@ The `protocol/` subfolder contains the two configurations used in the protocol p
 To ensure that the PROTEUS are consistent with each other for the main intercomparison,
 there is a single base configuration file `intercomp/_base.toml` from which planet-specific
 configuration files are generated in the `intercomp/` directory. These configurations may
-be generated by running the `chili_gen.py` script.
+be generated by running the `intercomp/_generate.py` script.
