### 2025-10-24 4831c09
Update links to other repos in docs. Update bolometric flux scaling in CHILI
