### 2025-10-24 fe471a6
Fix grid config syntax. Move generate script to tools/ folder. Update postproc to allow grids
