+++ tools/chili_postproc.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 # Postprocess output into CHILI-MIP format
 
 # Import modules
@@ -6,7 +7,7 @@ from __future__ import annotations
 
 import os
 import sys
-
+from glob import glob
 import matplotlib as mpl
 
 mpl.use("Agg") # noqa
@@ -22,144 +23,170 @@ from proteus.utils.plot import get_colour, latexify
 # Target times for sampling profile [years]
 tau_target = [1e3, 1e4, 1e5, 1e7, 1e9]
 
-# Simulation folder
-if len(sys.argv) != 2:
-    raise ValueError("Must provide path to simulation folder")
-simdir = os.path.abspath(sys.argv[1])
-print(f"Sim dir: {simdir}")
+def postproc_once(simdir:str):
+    '''Run postprocessing steps for a single simulation'''
+    print(f"Simulation dir: {simdir}")
 
-# Read simulation helpfile
-hfpath = os.path.join(simdir, "runtime_helpfile.csv")
-if not os.path.isfile(hfpath):
-   raise FileNotFoundError(f"Cannot find {hfpath}")
-hf_all = pd.read_csv(hfpath, delimiter=r'\s+')
+    # Read simulation helpfile
+    hfpath = os.path.join(simdir, "runtime_helpfile.csv")
+    if not os.path.isfile(hfpath):
+        raise FileNotFoundError(f"Cannot find {hfpath}")
+    hf_all = pd.read_csv(hfpath, delimiter=r'\s+')
 
-# Read config
-config = read_config_object(os.path.join(simdir, "init_coupler.toml"))
+    # Read config
+    config = read_config_object(os.path.join(simdir, "init_coupler.toml"))
 
-# Write to expected format
-out = {}
-out["t(yr)"]            = np.array(hf_all["Time"].iloc[:])
-out["T_surf(K)"]        = np.array(hf_all["T_surf"].iloc[:])
-out["T_pot(K)"]         = np.array(hf_all["T_pot"].iloc[:])
-out["phi(vol_frac)"]    = np.array(hf_all["Phi_global_vol"].iloc[:])
-out["massC_solid(kg)"]  = np.array(hf_all["C_kg_solid"].iloc[:])
-out["massC_melt(kg)"]   = np.array(hf_all["C_kg_liquid"].iloc[:])
-out["massC_atm(kg)"]    = np.array(hf_all["C_kg_atm"].iloc[:])
-out["massH_solid(kg)"]  = np.array(hf_all["H_kg_solid"].iloc[:])
-out["massH_melt(kg)"]   = np.array(hf_all["H_kg_liquid"].iloc[:])
-out["massH_atm(kg)"]    = np.array(hf_all["H_kg_atm"].iloc[:])
-out["massO_atm(kg)"]    = np.array(hf_all["O_kg_atm"].iloc[:])
-out["mmw(kg/mol)"]      = np.array(hf_all["atm_kg_per_mol"].iloc[:])
-out["p_surf(bar)"]      = np.array(hf_all["P_surf"].iloc[:])
-for gas in vol_list:
-    out[f"p_{gas}(bar)"]       = np.array(hf_all[f"{gas}_bar"].iloc[:])
-out["fO2_liquid(bar)"]  = out["p_O2(bar)"]
-out["fO2_solid(bar)"]   = np.zeros_like(out["t(yr)"])
-out["R_trans(m)"]       = np.array(hf_all["R_obs"].iloc[:])
-out["flux_surf(W/m2)"]  = np.array(hf_all["F_int"].iloc[:])
[…]
