+++ input/chili/readme.txt
@@ -6,4 +6,13 @@ The `protocol/` subfolder contains the two configurations used in the protocol p
 To ensure that the PROTEUS are consistent with each other for the main intercomparison,
 there is a single base configuration file `intercomp/_base.toml` from which planet-specific
 configuration files are generated in the `intercomp/` directory. These configurations may
-be generated by running the `intercomp/_generate.py` script.
+be generated by running the `tools/chili_generate.py` script.
+
+To run the protocol, the steps would then be:
+1) Install PROTEUS
+2) Generate configs: `./tools/chili_generate.py `
+3) Run Earth grid: `proteus grid -c input/chili/intercomp/earth.grid.toml`
+4) Run Venus grid: `proteus grid -c input/chili/intercomp/venus.grid.toml`
+6) Run TRAPPIST-1 b: `proteus start -c input/chili/intercomp/tr1b.toml`
+7) Run TRAPPIST-1 e: `proteus start -c input/chili/intercomp/tr1e.toml`
+8) Run TRAPPIST-1 α: `proteus start -c input/chili/intercomp/tr1a.toml`
