+++ input/chili/intercomp/_base.grid.toml
@@ -9,8 +9,8 @@ max_mem  = 3        # maximum memory per CPU in GB (e.g. 3)
 
 ["delivery.elements.H_kg"]
     method = "direct"
-    values = ["4.70e20", "1.60e20", "7.80e20", "16e20"]
+    values = [4.70e20, 1.60e20, 7.80e20, 16e20]
 
 ["delivery.elements.C_kg"]
     method = "direct"
-    values = ["2.73e20", "1.36e20", "2.72e20", "5.44e20"]
+    values = [2.73e20, 1.36e20, 2.72e20, 5.44e20]
