### 2025-10-24 6aca467
Update chili readme
