+++ input/chili/readme.md
@@ -8,11 +8,16 @@ there is a single base configuration file `intercomp/_base.toml` from which plan
 configuration files are generated in the `intercomp/` directory. These configurations may
 be generated by running the `tools/chili_generate.py` script.
 
-To run the protocol, the steps would then be:
+To execute steps for the intercomparison:
+
 1) Install PROTEUS
-2) Generate configs: `./tools/chili_generate.py `
-3) Run Earth grid: `proteus grid -c input/chili/intercomp/earth.grid.toml`
-4) Run Venus grid: `proteus grid -c input/chili/intercomp/venus.grid.toml`
-6) Run TRAPPIST-1 b: `proteus start -c input/chili/intercomp/tr1b.toml`
-7) Run TRAPPIST-1 e: `proteus start -c input/chili/intercomp/tr1e.toml`
-8) Run TRAPPIST-1 α: `proteus start -c input/chili/intercomp/tr1a.toml`
+2) Generate configuration files
+    - $ ./tools/chili_generate.py
+3) Run Earth/Venus grids and process them
+    - E.g. for Earth...
+    - $ proteus grid -c input/chili/intercomp/earth.grid.toml
+    - $ ./tools/chili_postproc.py output/chili_earth_grid
+4) Run TRAPPIST-1 cases and process them
+    - E.g. for TRAPPIST-1 b...
+    - $ proteus start -c input/chili/intercomp/tr1b.toml
+    - $ ./tools/chili_postproc.py output/chili_tr1b
