+++ tools/chili_postproc.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 import os
 import sys
 from glob import glob
+
 import matplotlib as mpl
 
 mpl.use("Agg") # noqa
