### 2025-10-24 ccafced
Postproc script renames files into CHILI syntax
