### 2025-10-25 80d5b1c
Tweaks to chili configs. Add validator for AGNI chemistry which checks if fastchem is installed. Update habrok guide with module required for fastchem
