+++ input/chili/intercomp/_base.toml
@@ -4,7 +4,7 @@ version = "2.0"
 
 [params.out]
 path = "chili_base"
-logging = "INFO"
+logging = "DEBUG"
 plot_fmt = "png"
 write_mod = 1
 plot_mod = 20
@@ -112,11 +112,11 @@ tmp_maximum = 5000.0
 [atmos_clim.agni]
 spectral_group = "Dayspring"
 spectral_bands = "48"
-p_top = 1e-05
+p_top = 1e-04
 p_obs = 0.02
 surf_material = "greybody"
-num_levels = 42
-chemistry = "none"
+num_levels = 45
+chemistry = "eq"
 solve_energy = true
 solution_atol = 0.1
 solution_rtol = 0.1
@@ -154,9 +154,9 @@ rheo_phi_loc = 0.5
 rheo_phi_wid = 0.2
 
 [interior.spider]
-ini_entropy = 3000.0
+ini_entropy = 3300.0
 ini_dsdr = -4.698e-06
-num_levels = 140
+num_levels = 150
 mixing_length = 2
 tolerance = 1e-08
 tolerance_rel = 1e-08
