### 2025-10-25 7c6b05e
Stop Tr1 simulations at 7.6 Gyr
