### 2025-10-25 9cbbfed
Stop time key
