### 2025-10-25 a6defd3
Update julia in workflow to 1.12
