+++ .github/workflows/tests.yaml
@@ -62,7 +62,7 @@ jobs:
       - name: Setup Julia
         uses: julia-actions/setup-julia@v2
         with:
-          version: '1.11.3'
+          version: '1.12'
 
       - name: Cache Julia
         uses: julia-actions/cache@v2
