### 2025-10-25 f21855e
OPENSSL_3.3.0 library path for julia 1.12
