### 2025-10-25 e1e4735
Add info about Julia1.11 in docs. Add juliapkg file to tell juliacall which version of Julia to use. Update chili configs
