### 2025-10-25 1680890
Add option for manually choosing initial atmosphere profile shape
