### 2025-10-25 5dd8b23
Update configs and examples
