+++ tools/chili_generate.py
@@ -66,7 +66,7 @@ for p in ("tr1a","tr1b","tr1e"):
     cfg[p]["star"]["mass"] = 0.1      # Msun
 cfg["tr1b"]["orbit"]["semimajoraxis"] = 1.154e-2 # AU
 cfg["tr1e"]["orbit"]["semimajoraxis"] = 2.295e-2 # AU
-cfg["tr1a"]["orbit"]["semimajoraxis"] = 3.0e-3 #6.750e-4 # AU
+cfg["tr1a"]["orbit"]["semimajoraxis"] = 6.750e-4 # AU     2.0e-3
 
 
 # ------------------------
