+++ input/all_options.toml
@@ -204,10 +204,11 @@ version = "2.0"
         solution_atol   = 0.5           # solver absolute tolerance
         solution_rtol   = 0.15          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
-        condensation    = false         # volatile condensation
+        condensation    = false         # include volatile rainout upon condensation
         latent_heat     = false         # include latent heat release by phase change?
         real_gas        = false         # use real-gas equations of state
         psurf_thresh    = 0.1           # bar, surface pressure where we switch to 'transparent' mode
+        dx_max_ini      = 300.0         # initial maximum temperature step [kelvin] allowed by solver
         dx_max          = 35.0          # maximum temperature step [kelvin] allowed by solver
         max_steps       = 70            # max steps allowed by solver during each iteration
         perturb_all     = true          # updated entire jacobian each step?
@@ -217,6 +218,7 @@ version = "2.0"
         fastchem_maxiter_solv = 20000   # Maximum FC iterations (internal solver)
         fastchem_xtol_chem    = 1e-4    # FC solver tolerance (chemistry)
         fastchem_xtol_elem    = 1e-4    # FC solver tolerance (elemental)
+        ini_profile     = 'loglinear'   # Initial guess for temperature profile shape
 
     [atmos_clim.janus]
         p_top           = 1.0e-6        # bar, top of atmosphere grid pressure
