+++ input/chili/intercomp/_base.toml
@@ -115,21 +115,22 @@ spectral_group = "Dayspring"
 spectral_bands = "48"
 p_top = 1e-04
 p_obs = 0.02
-surf_material = "greybody"
+surf_material = "surface_albedos/Hammond24/lunarmarebasalt.dat"
 num_levels = 45
 chemistry = "none"
 solve_energy = true
 solution_atol = 0.1
 solution_rtol = 0.1
 overlap_method = "ee"
-condensation = false
+condensation = true
 latent_heat = false
-real_gas = false
+real_gas = true
 psurf_thresh = 0.1
-dx_max = 10.0
+dx_max     = 10.0
+dx_max_ini = 200.0
 max_steps = 70
 perturb_all = true
-mlt_criterion = "l"
+mlt_criterion = "s"
 fastchem_floor = 273.0
 ini_profile = "isothermal"
 
