### 2025-10-26 5b4fa57
create a new csv file with only completed cases and add number in the panels fro the ecdf gridplot. Update the toml file for comparison escape on off
