+++ src/proteus/grid/post_processing_grid.py
@@ -459,6 +459,103 @@ def save_error_running_cases(grid_name: str, cases_data: List[Dict[str, Any]], g
 
     print(f"→ Error/Running (and missing‐output) CSV saved to: {err_csv}")
 
+def save_completed_cases(
+    grid_name: str,
+    cases_data: List[Dict[str, Any]],
+    grid_parameters: Dict[str, List[Any]],
+    case_params: Dict[int, Dict[str, Any]],
+    extracted_value: Dict[str, List[Any]],
+    output_to_extract: List[str],
+    output_dir: Path,
+) -> None:
+    """
+    Save all cases whose status starts with 'Completed' into a separate CSV file
+    named '{grid_name}_filtered.csv'.
+
+    Parameters
+    ----------
+    grid_name : str
+        Name of the grid.
+
+    cases_data : list
+        List of dictionaries containing simulation data.
+
+    grid_parameters : dict
+        A dictionary where each key is a parameter name, and its corresponding values
+        used for the entire grid is a list.
+
+    case_params : dict
+        Dictionary containing each case number with the name and values of the
+        tested parameters in this grid.
+
+    extracted_value : dict
+        A dictionary containing the extracted values of the specified parameter
+        for all cases of the grid.
+
+    output_to_extract : list
+        List of output values extracted from each simulation in the grid.
+
+    output_dir : Path
+        Directory where the generated CSV file will be saved.
+        Created if it does not exist.
+    """
+    output_dir = Path(output_dir)
+    output_dir.mkdir(parents=True, exist_ok=True)
+
+    # Identify indices where status starts with "Completed" (case-insensitive, trimmed)
+    completed_indices = []
+    for idx, case_data in enumerate(cases_data):
+        status_raw = case_data.get("status", "")
+        # Normalize status string safely
+        if not isinstance(status_raw, str):
+            continue
+        status_clean = status_raw.strip().lower().replace('\xa0', ' ')  # replace non-breaking spaces if any
+        if status_clean.startswith("completed"):
+            completed_indices.append(idx)
+
+    if not completed_indices:
+        print("→ No cases with status starting with 'Completed' found; skipping filtered CSV.")
+        return
+
+    # Build output path
+    filtered_csv = output_dir / f"{grid_name}_filtered.csv"
+
+    with open(filtered_csv, mode="w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+
+        # Header block
+        writer.writerow(["#############################################################################################################"])
+        writer.writerow([f"Grid name:                               {grid_name}"])
+        writer.writerow([f"Total number of 'Completed*' cases:      {len(completed_indices)}"])
+        writer.writerow(["----------------------------------------------------------"])
+        writer.writerow(
+            ["Case number", "Status"]
+            + list(grid_parameters.keys())
+            + list(extracted_value.keys())
+        )
+        writer.writerow(["#############################################################################################################"])
+        writer.writerow([])
+
+        # Write each completed case row
+        for case_index in completed_indices:
+            status = cases_data[case_index].get("status", "Unknown") or "Unknown"
+            row = [case_index, f"'{status}'"]
+
+            # Add grid parameter values
+            for param in grid_parameters.keys():
+                row.append(case_params.get(case_index, {}).get(param, "NA"))
+
+            # Add extracted outputs
+            for param in extracted_value.keys():
+                vals = extracted_value[param]
+                if case_index < len(vals):
+                    row.append(vals[case_index])
[…]
