### 2025-10-26 fb82ac1
Add info to docs about running single jobs on Habrok
