### 2025-10-26 339e394
Update stellar mass validator to enable correct calculation for TRAPPIST-1 planets. Fix docstrings on some functions
