### 2025-10-27 bbf58f4
Add plot for postproc CHILI grids. Allow sim output to scratch/ folder
