### 2025-10-27 179fd2c
Now using a single calculation of rho_obs from R_obs. Fixes to AGNI transparent solver so that it uses a more appropriate p_boa value. Option for user-selected linesearch method
