+++ src/proteus/atmos_clim/agni.py
@@ -268,7 +268,7 @@ def deallocate_atmos(atmos):
     safe_rm(str(atmos.fastchem_work))
 
 
-def update_agni_atmos(atmos, hf_row:dict, dirs:dict, transparent:bool):
+def update_agni_atmos(atmos, hf_row:dict, dirs:dict, config:Config):
     """Update atmosphere struct.
 
     Sets the new boundary conditions and composition.
@@ -281,8 +281,8 @@ def update_agni_atmos(atmos, hf_row:dict, dirs:dict, transparent:bool):
             Dictionary containing simulation variables for current iteration
         dirs : dict
             Directories dictionary
-        transparent : bool
-            If True, the atmosphere is configured to be transparent to radiation
+        config: Config
+            PROTEUS config object
 
     Returns
     ----------
@@ -308,10 +308,15 @@ def update_agni_atmos(atmos, hf_row:dict, dirs:dict, transparent:bool):
 
     # ---------------------
     # Transparent mode?
-    if transparent:
+    if hf_row["P_surf"] < config.atmos_clim.agni.psurf_thresh:
+
+        # set p_boa to threshold value [bar]
+        atmos.p_boa = float(config.atmos_clim.agni.psurf_thresh) * 1.0e-5
+
+        # update struct to handle this mode of operation
         jl.AGNI.atmosphere.make_transparent_b(atmos)
-        atmos.tmp[:]  = float(atmos.tmp_surf)
-        atmos.tmpl[:] = float(atmos.tmp_surf)
+
+        # return here - don't do anything else to the `atmos` struct
         return atmos
 
     # ---------------------
@@ -383,7 +388,7 @@ def _solve_energy(atmos, loops_total:int, dirs:dict, config:Config):
         log.info("Attempt %d" % attempts)
 
         # default parameters
-        linesearch   = 2
+        linesearch   = int(config.atmos_clim.agni.ls_default)
         easy_start   = False
         dx_max       = float(config.atmos_clim.agni.dx_max)
         ls_increase  = 1.01
@@ -537,7 +542,7 @@ def _solve_transparent(atmos, config:Config):
     return atmos
 
 def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
-                hf_row:dict, transparent:bool):
+                hf_row:dict):
     """Run AGNI atmosphere model.
 
     Calculates the temperature structure of the atmosphere and the fluxes, etc.
@@ -555,8 +560,6 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
             Configuration options and other variables
         hf_row : dict
             Dictionary containing simulation variables for current iteration
-        transparent : bool
-            Find solution assuming a transparent atmosphere
 
     Returns
     ----------
@@ -568,7 +571,6 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
 
     # Inform
     log.debug("Running AGNI...")
-    time_str = "%d"%hf_row["Time"]
 
     # ---------------------------
     # Solve atmosphere
@@ -578,10 +580,15 @@ def run_agni(atmos, loops_total:int, dirs:dict, config:Config,
     if bool(atmos.transparent):
         # no opacity
         log.info("Using transparent solver")
+        atmos.transspec_p = float(atmos.p_boa)
         atmos = _solve_transparent(atmos, config)
 
     # Opaque case
     else:
+
+        # Set observed pressure
+        atmos.transspec_p = float(config.atmos_clim.agni.p_obs * 1e5) # converted to Pa
+
[…]
