+++ input/chili/intercomp/_base.toml
@@ -15,14 +15,14 @@ remove_sf = true
 starspec = 10000000.0
 starinst = 10.0
 method = "adaptive"
-minimum = 10000.0
+minimum = 1000.0
 minimum_rel = 1e-05
 maximum = 10000000.0
 initial = 30.0
 
 [params.dt.adaptive]
 atol = 0.03
-rtol = 0.10
+rtol = 0.09
 
 [params.stop]
 strict = false
@@ -112,7 +112,7 @@ tmp_maximum = 5000.0
 
 [atmos_clim.agni]
 spectral_group = "Dayspring"
-spectral_bands = "256"
+spectral_bands = "48"
 p_top = 1e-04
 p_obs = 0.02
 surf_material = "surface_albedos/Hammond24/lunarmarebasalt.dat"
@@ -128,6 +128,7 @@ real_gas = true
 psurf_thresh = 0.1
 dx_max     = 15.0
 dx_max_ini = 200.0
+ls_default = 1
 max_steps = 70
 perturb_all = true
 mlt_criterion = "s"
@@ -157,7 +158,7 @@ rheo_phi_loc = 0.5
 rheo_phi_wid = 0.2
 
 [interior.spider]
-ini_entropy = 3200.0
+ini_entropy = 3900.0
 ini_dsdr = -4.698e-06
 num_levels = 150
 mixing_length = 2
