### 2025-10-27 846559b
Indicate on plot when T(p) is from transparent solver
