### 2025-10-27 4e8b5bb
Our Aragog fork now has an entry on PyPI
We will be using the "Forming Worlds" entry of Aragog on PyPI: "fwl-aragog".
That entry comes from the "Forming Worlds" fork of Aragog at https://github.com/FormingWorlds/aragog.
