### 2025-10-27 1d37a7f
sbatch in docs
