### 2025-10-27 b339b46
Update fwl-aragog dependency version to use calver
