### 2025-10-27 46430ab
Add boreas to pyprojet
