### 2025-10-27 e053b97
Add boreas options to all_options
