### 2025-10-27 96dfdc6
Tidy up module version checker code using boolean algebra. Set BOREAS parameters.
