+++ input/demos/boreas.toml
@@ -7,7 +7,7 @@ version = "2.0"
 # Parameters
 [params]
     [params.out]
-        logging     = "INFO"
+        logging     = "DEBUG"
         path        = "boreas"
         plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
 
