### 2025-10-27 7511687
Fix broken URL. Get boreas gases/elements dynamically
