### 2025-10-28 f4fe62a
Final step in fixing 426 (#549)
* Our Aragog fork now has an entry on PyPI

We will be using the "Forming Worlds" entry of Aragog on PyPI: "fwl-aragog".
That entry comes from the "Forming Worlds" fork of Aragog at https://github.com/FormingWorlds/aragog.

* The developer install should also use the fork.

The developer install should also use the "Forming Worlds" fork of Aragog.

* Update fwl-aragog dependency version to use calver

---------

Co-authored-by: Harrison Nicholls <32022679+nichollsh@users.noreply.github.com>
