+++ pyproject.toml
@@ -49,7 +49,7 @@ dependencies = [
     "fwl-mors>=25.07.15",
     "fwl-calliope>=25.05.01",
     "fwl-zephyrus>=25.03.11",
-    "aragog>=v0.2.8-alpha",
+    "fwl-aragog>=25.10.27",
     "fwl-zalmoxis>=25.09.07",
     "cmcrameri",
     "juliacall",
