### 2025-10-28 120e9ab
Try/catch in NetCDF flag parser. Plot title
