### 2025-10-28 c939ae6
Use TR1 spectrum for TR1 planets. Normalise  stellar fluxes to be consistent with baraffe tracks. Update SMAs for TR1 planets to equal protocol specification. Ensure phi_vol is reasonable
