+++ tools/chili_generate.py
@@ -39,10 +39,17 @@ print("Removed old configs")
 
 # ------------------------
 # Earth and Venus (Table 2 of protocol paper)
+tnow = 4.567 # Gyr
 for p in ("earth","venus"):
     cfg[p] = deepcopy(cfg["base"])
+
+    # output
     cfg[p]["params"]["out"]["path"] = f"chili_{p}"
+
+    # star
     cfg[p]["star"]["mass"] = 1.0  # Msun
+    cfg[p]["star"]["mors"]["spec"] = "stellar_spectra/Named/sun.txt"
+    cfg[p]["star"]["mors"]["age_now"] = tnow   # Gyr
 
 # Planet mass
 cfg["earth"]["struct"]["mass_tot"] = 1.0    # Mearth
@@ -52,7 +59,7 @@ cfg["venus"]["struct"]["mass_tot"] = 0.815  # Mearth
 cfg["earth"]["orbit"]["semimajoraxis"] = 1.0    # AU
 cfg["venus"]["orbit"]["semimajoraxis"] = 0.723  # AU
 
-# Scale bolometric flux to match Baraffe tracks at t=t0
+# Scale flux to match Baraffe tracks at t=t0
 cfg["earth"]["star"]["bol_scale"] = 920.0 / 1005.3
 cfg["venus"]["star"]["bol_scale"] = 1760.0 / 1923.1
 
@@ -65,17 +72,39 @@ for p in ("earth","venus"):
 
 # ------------------------
 # TRAPPIST-1 b/e/α (Table 4 of protocol paper)
+tnow = 7.6 # Gyr
 for p in ("tr1a","tr1b","tr1e"):
     cfg[p] = deepcopy(cfg["base"])
+
+    # output
     cfg[p]["params"]["out"]["path"]  = "scratch/" if use_scratch else ""
     cfg[p]["params"]["out"]["path"] += f"chili_{p}/"
-    cfg[p]["params"]["stop"]["time"]["maximum"] = (7.6 - cfg[p]["star"]["age_ini"])*1e9
-    cfg[p]["star"]["mass"] = 0.09      # Msun
-cfg["tr1b"]["orbit"]["semimajoraxis"] = 1.154e-2 # AU
-cfg["tr1e"]["orbit"]["semimajoraxis"] = 2.295e-2 # AU
-cfg["tr1a"]["orbit"]["semimajoraxis"] = 6.750e-4 # AU     2.0e-3
+
+    # star
+    cfg[p]["star"]["mass"]            = 0.0898 # Msun
+    cfg[p]["star"]["mors"]["spec"]    = "stellar_spectra/Named/trappist-1.txt"
+    cfg[p]["star"]["mors"]["age_now"] = tnow   # Gyr
+
+# Scale flux to match Baraffe tracks at t=t0
+cfg["tr1b"]["star"]["bol_scale"] = 3.628e+04 / 4.439e+04
+cfg["tr1e"]["star"]["bol_scale"] = 5.648e+03 / 6.909e+03
+cfg["tr1a"]["star"]["bol_scale"] = 1.687e+06 / 4.218e+06
+
+# Orbital distances
+cfg["tr1b"]["orbit"]["semimajoraxis"] = 0.01154 # AU
+cfg["tr1e"]["orbit"]["semimajoraxis"] = 0.02925 # AU
+cfg["tr1a"]["orbit"]["semimajoraxis"] = 6.750e-4 # AU
+
+# Initial ages
+cfg["tr1b"]["star"]["age_ini"] = 0.05 # Gyr
+cfg["tr1e"]["star"]["age_ini"] = 0.05 # Gyr
+cfg["tr1a"]["star"]["age_ini"] = 1.00 # Gyr  <- different from b/e
 
 # ------------------------
[…]
