### 2025-10-28 e031c45
Parse Netcdf var
