### 2025-10-28 54d81ec
NetCDF masked array parser
