### 2025-10-28 c9ca7e4
update ecdf grid function to plot only values > 0 for species plot
