+++ tools/nogit_grid_proteus.py
@@ -1,612 +1,0 @@
-#!/usr/bin/env python3
-
-# Run PROTEUS for a grid of parameters
-
-from __future__ import annotations
-
-import gc
-import itertools
-import logging
-import multiprocessing
-import os
-import shutil
-import subprocess
-import sys
-import time
-import csv
-from copy import deepcopy
-from datetime import datetime
-from getpass import getuser
-import argparse
-
-import numpy as np
-
-from proteus.config import Config, read_config_object
-from proteus.utils.helper import get_proteus_dir
-
-PROTEUS_DIR=get_proteus_dir()
-
-# Custom logger instance
-def setup_logger(logpath:str="new.log",level=1,logterm=True):
-
-    # https://stackoverflow.com/a/61457119
-
-    custom_logger = logging.getLogger()
-    custom_logger.handlers.clear()
-
-    if os.path.exists(logpath):
-        os.remove(logpath)
-
-    fmt = logging.Formatter("[%(asctime)s] %(message)s", datefmt='%Y-%m-%d %H:%M:%S')
-
-    level_code = logging.INFO
-    match level:
-        case 0:
-            level_code = logging.DEBUG
-        case 2:
-            level_code = logging.WARNING
-        case 3:
-            level_code = logging.ERROR
-        case 4:
-            level_code = logging.CRITICAL
-
-    # Add terminal output to logger
-    if logterm:
-        sh = logging.StreamHandler(sys.stdout)
-        sh.setFormatter(fmt)
-        sh.setLevel(level_code)
-        custom_logger.addHandler(sh)
-
-    # Add file output to logger
-    fh = logging.FileHandler(logpath)
-    fh.setFormatter(fmt)
-    fh.setLevel(level)
-    custom_logger.addHandler(fh)
-    custom_logger.setLevel(level_code)
-
-    # Capture unhandled exceptions
-    # https://stackoverflow.com/a/16993115
-    def handle_exception(exc_type, exc_value, exc_traceback):
-        if issubclass(exc_type, KeyboardInterrupt):
-            custom_logger.error("KeyboardInterrupt")
-            sys.__excepthook__(exc_type, exc_value, exc_traceback)
-            return
-        custom_logger.critical("Uncaught exception", exc_info=(exc_type, exc_value, exc_traceback))
-    sys.excepthook = handle_exception
-
-    return
-
-# Object for handling the parameter grid
-class Grid():
-
-    CONFIG_BASENAME = "case_%05d.toml"
-
-    def __init__(self, name:str, base_config_path:str, custom_grid_flat:list=None, symlink_dir:str="_UNSET"):
-
-        # Grid's own name (for versioning, etc.)
-        self.name = str(name).strip()
-        self.outdir = PROTEUS_DIR+"/output/"+self.name+"/"
-        self.conf = str(base_config_path)
-        if not os.path.exists(self.conf):
-            raise Exception("Base config file '%s' does not exist!" % self.conf)
-
-        self.symlink_dir = symlink_dir
-        if self.symlink_dir in ["","."]:
-            raise Exception("Symlinked directory is set to a blank path")
-
-        # Paths
-        self.outdir = os.path.abspath(self.outdir) + "/"
-        self.cfgdir = os.path.join(self.outdir, "cfgs") + "/"
-        self.logdir = os.path.join(self.outdir, "logs") + "/"
-
-        # Remove old output location
-        if os.path.exists(self.outdir):
-            if os.path.islink(self.outdir):
-                os.unlink(self.outdir)
-            if os.path.isdir(self.outdir):
-                print("Removing old files at '%s'" % self.outdir)
-                subfolders = [ f.path.split("/")[-1].lower() for f in os.scandir(self.outdir) if f.is_dir() ]
[…]
