+++ src/proteus/escape/wrapper.py
@@ -30,7 +30,7 @@ def run_escape(config:Config, hf_row:dict, dt:float, stellar_track):
 
     if not config.escape.module:
         hf_row["esc_rate_total"] = 0.0
-        log.info(f"Escape is disabled, bulk rate = {hf_row["esc_rate_total"]:.2e} kg s-1")
+        log.info(f"Escape is disabled, bulk rate = {hf_row['esc_rate_total']:.2e} kg s-1")
         return
 
     elif config.escape.module == 'zephyrus':
@@ -42,7 +42,7 @@ def run_escape(config:Config, hf_row:dict, dt:float, stellar_track):
     else:
         raise ValueError(f"Invalid escape model: {config.escape.module}")
 
-    log.info(f"Bulk escape rate = {hf_row["esc_rate_total"]:.2e} kg s-1")
+    log.info(f"Bulk escape rate = {hf_row['esc_rate_total']:.2e} kg s-1")
 
     # calculate new elemental inventories
     solvevol_target = calc_new_elements(hf_row, dt,

[…]
