### 2025-10-28 abe4921
Style: apply ruff import formatting (pre-commit)
