### 2025-10-28 5e842fc
Clarify docstring
