### 2025-10-28 72fec7a
Fix docstrings. Gas mixing ratios stored by *_vmr_xuv key, rather than *_xuv, for clarity.
