### 2025-10-30 bf7ab1f
SPIDER config options: flux terms, smoothing width
