### 2025-11-03 #8758 @timlichtenberg
@warrickball Is there already an indication of when we can expect the paper to be assigned an editor? Thanks a lot in advance.
