### 2025-11-06 5e0d381
toy example and more
