### 2025-11-11 1b3c9ac
Allow reading bond albedo from lookup table
