+++ src/proteus/proteus.py
@@ -106,7 +106,7 @@ class Proteus:
 
         #    atmosphere solver
         from proteus.atmos_clim import run_atmosphere
-        from proteus.atmos_clim.common import Atmos_t
+        from proteus.atmos_clim.common import Atmos_t, Albedo_t
 
         #    escape and outgas
         from proteus.escape.wrapper import run_escape
@@ -220,13 +220,17 @@ class Proteus:
         # Download basic data
         download_sufficient_data(self.config)
 
-        # Initialise interior and atmosphere objects
+        # Initialise interior object
         if self.config.interior.module == "spider":
             spider_dir = self.directories["spider"]
         else:
             spider_dir = None
         self.interior_o = Interior_t(get_nlevb(self.config), spider_dir=spider_dir)
-        self.atmos_o    = Atmos_t()
+
+        # Initialise atmosphere object
+        self.atmos_o  = Atmos_t()
+        if self.config.atmos_clim.albedo_from_file:
+            self.atmos_o.albedo_o = Albedo_t(self.config.atmos_clim.albedo_pl)
 
         # Is the model resuming from a previous state?
         if not self.config.params.resume:
