### 2025-11-11 477d0e2
Example config and syntax fixes
