### 2025-11-11 a20c1c9
Ensure type
