+++ src/proteus/atmos_clim/agni.py
@@ -186,7 +186,7 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
 
                         hf_row["F_ins"],
                         config.orbit.s0_factor,
-                        hf_row["albedo_pl"],
+                        float(hf_row["albedo_pl"]),
                         config.orbit.zenith_angle,
 
                         hf_row["T_surf"],
