### 2025-11-11 dbe2235
Read table as dataframe. Update type
