### 2025-11-11 b43d972
Safety checks on albedo
