+++ src/proteus/proteus.py
@@ -232,6 +232,9 @@ class Proteus:
         if self.config.atmos_clim.albedo_from_file:
             log.debug("Reading albedo data from file")
             self.atmos_o.albedo_o = Albedo_t(self.config.atmos_clim.albedo_pl)
+            if not self.atmos_o.albedo_o.ok:
+                UpdateStatusfile(self.directories, 22)
+                raise RuntimeError("Problem when loading albedo data file")
 
         # Is the model resuming from a previous state?
         if not self.config.params.resume:
