### 2025-11-11 a4c6893
Add test for albedo lookup example
