+++ tests/data/albedo_lookup/config.toml
@@ -10,7 +10,9 @@ version = "2.0"
 
 [params]
     [params.out]
-        path        = "albedo_table"
+        logging     = "ERROR"
+        path        = "albedo_lookup"
+        plot_mod    = 0
 
 [star]
     mass    = 1.0               # [M_sun]
@@ -38,7 +40,7 @@ version = "2.0"
 
     rayleigh        = false
     cloud_enabled   = false
-    albedo_pl       = "nogit_albedo_cloudy.csv"
+    albedo_pl       = "tests/data/albedo_lookup/cloudy.csv"
     surf_greyalbedo = 0.3
 
     [atmos_clim.agni]
