### 2025-11-11 638c78e
Add check to config
