### 2025-11-12 010c8e3
Move test data for albedo_lookup and add better tests
