### 2025-11-13 a1e1570
Update AGNI call syntax and required version
