### 2025-11-13 9d6ede0
Update keys for AGNI functon calls. Config options
