+++ src/proteus/inference/BO.py
@@ -18,11 +18,12 @@ import matplotlib.pyplot as plt
 import torch
 from botorch.acquisition import LogExpectedImprovement, UpperConfidenceBound, qLogExpectedImprovement
 from botorch.fit import fit_gpytorch_mll
+from botorch.optim.fit import fit_gpytorch_mll_torch
 from botorch.models import SingleTaskGP
 from botorch.models.transforms import Normalize, Standardize
 from botorch.optim import optimize_acqf
 from gpytorch.mlls import ExactMarginalLogLikelihood
-
+from utils import get_kernel_w_prior
 
 # Set tensor dtype for consistent precision
 dtype = torch.double
@@ -39,6 +40,172 @@ def unit_bounds(d):
     # Build bounds [[0,...,0], [1,...,1]]
     bounds = torch.tensor([[0] * d, [1] * d], dtype=dtype)
     return bounds
+
+
+def BO_step(D, B, f, k, acqf, lock, worker_id, x_in = None):
+
+    """Perform a single Bayesian optimization step.
+
+    Fits a GP to current data, optimizes an acquisition function,
+    updates busy points, evaluates the objective, and logs timing.
+
+    Args:
+        D (dict): Shared dict containing 'X' and 'Y' lists of data.
+        B (dict): Shared dict mapping worker IDs to busy input points.
+        f (callable): Objective function to evaluate.
+        k (Kernel): GPyTorch kernel for the GP covariance.
+        acqf (str): Acquisition function.
+        lock (multiprocessing.Lock): Lock for synchronizing shared state.
+        worker_id (int): ID of the calling worker.
+        x_in (torch.Tensor, optional): Initial input for the first iteration.
+
+    Returns:
+        tuple: (x_next, y_next, bo_duration, eval_duration,
+                lock_duration, fit_duration, acq_duration, min_dist)
+    """
+
+    t_0_bo = time.perf_counter()
+    noise = 1e-4
+
+    if x_in is None:
+
+        t_0_lock = time.perf_counter()
+        with lock:
+            X = D["X"]
+            Y = D["Y"]
+            busys = list(B.values())
+
+        t_1_lock = time.perf_counter()
+
+        busys = torch.cat(busys, dim=0)
+
+        d = X.shape[-1]
+
+        # print("worker ", worker_id, "samples ", len(X), "\n")
+
+        best = Y.max().item()
+
+        t_0_fit = time.perf_counter()
+        gp = SingleTaskGP(  train_X=X,
+                            train_Y=Y,
+                            covar_module=k,
+                            train_Yvar=torch.full_like(Y, noise),
+                            input_transform=Normalize(d=d),
+                            outcome_transform=Standardize(m=1),
+                            )
+
+        lik = gp.likelihood
+        mll = ExactMarginalLogLikelihood(lik, gp)
+        fit_gpytorch_mll(mll, optimizer=fit_gpytorch_mll_torch,
+                         kwargs={"pick_best_of_all_attempts" : True, "max_attempts" : 10})
+
+        t_1_fit = time.perf_counter()
+
+        t_0_ac = time.perf_counter()
+
+        # build acquisition function
+        if acqf == "UCB":
+            acqf = UpperConfidenceBound(gp, beta = 2.)
+        elif acqf =="LogEI":
+            acqf = LogExpectedImprovement(gp, best_f=best)
+        elif acqf =="E-LogEI":
+            acqf = qLogExpectedImprovement(gp, best_f=best, X_pending=busys)
+        else:
+            raise ValueError("Unknown acquisition function, choices are UCB or LogEI")
+
+        x, _ = optimize_acqf(   acq_function=acqf, # expects outputs shape (N)
[…]
