### 2025-11-15 bee5866
init batch kernel
