### 2025-11-17 3544a0a
Fixes 437
1) Launches proteus with "all_options.toml" configuration
2) Uploads its plots to "output/full_physical/plot"
