+++ .github/workflows/tests.yaml
@@ -132,6 +132,16 @@ jobs:
             output/physical/plot/plot_*.png
             output/physical/runtime_helpfile.csv
 
+      # Full physical integration
+      - name: Full physical integration
+        run: proteus start -c input/all_options.toml
+        uses: actions/upload-artifact@v4
+        with:
+          name: full-physical-integration
+          path: |
+            output/full_physical/plot/plot_*.png
+            output/full_physical/runtime_helpfile.csv
+
       - name: Report coverage
         run: |
           coverage json
