### 2025-11-17 925ce78
Previous version returned "Invalid workflow file"
Apparently, something wrong with the formatting of the new entry in "tests.yaml" from the previous commit, it returned "Invalid workflow file (Line: 138, Col: 9): Unexpected value 'uses', (Line: 139, Col: 9): Unexpected value 'with'".
Hopefully, this will fix it.
