### 2025-11-17 27bd419
Update AGNI call and default config
