+++ src/proteus/atmos_clim/agni.py
@@ -23,6 +23,7 @@ log = logging.getLogger("fwl."+__name__)
 
 # Constant
 AGNI_LOGFILE_NAME="agni_recent.log"
+ALWAYS_DRY = ("CO", "N2", "H2")
 
 def sync_log_files(outdir:str):
     # Logfile paths
@@ -85,6 +86,32 @@ def _construct_voldict(hf_row:dict, dirs:dict):
 
     return vol_dict
 
+def _determine_condensates(vol_list:list):
+    """Determine which gases will be condensable.
+
+    Need to ensure that there's at least one 'dry' gas with non-zero opacity.
+
+    Parameters
+    -----------
+        config : Config
+            Configuration options and other variables
+        vol_list: list
+            List of included gases.
+
+    Returns
+    ----------
+        condensates : list
+            List of allowed-condensable gases
+    """
+
+    # single-gas case must be dry
+    if len(vol_list) == 1:
+        log.warning("Cannot include rainout condensation with only one gas!")
+        return []
+
+    # all dry gases...
+    return [v for v in vol_list if v not in ALWAYS_DRY]
+
 
 def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
     """Initialise atmosphere struct for use by AGNI.
@@ -128,37 +155,22 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
         input_sf =      get_spfile_path(dirs["fwl"], config)
         input_star =    sflux_path
 
+    # Fast I/O folder
+    if config.params.out.logging == "DEBUG":
+        io_dir = dirs["output"]
+    else:
+        io_dir = create_tmp_folder()
+
     # composition
     vol_dict = _construct_voldict(hf_row, dirs)
 
     # set condensation
     condensates = []
     if config.atmos_clim.agni.condensation:
-        if len(vol_dict) == 1:
-            # single-gas case
-            condensates = list(vol_dict.keys())
-        else:
-            # get sorted gases (in order of decreasing VMR at the surface)
-            vol_sorted = sorted(vol_dict.items(), key=lambda item: item[1])[::-1]
-
-            # Set gases as condensates...
-            condensates = ['H2O'] # always prefer H2O
-            for v in vol_sorted:
-                # add gas if it has non-zero abundance
-                if (v[1] > 1e-30) and (v[0] not in condensates):
-                    condensates.append(v[0])
-
-            # Remove the least abundant gas from the list, so that we have something to
-            #     fill the background with if everything else condenses at a given level
-            condensates = condensates[:-1]
+        condensates = _determine_condensates(vol_dict.keys())
 
     # Chemistry
-    include_all = False
-    fc_dir = create_tmp_folder()
-    if config.atmos_clim.agni.chemistry == 'eq':
-        include_all = True
-        condensates = []
-        log.debug("Fastchem work folder: '%s'"%fc_dir)
+    include_all = bool(config.atmos_clim.agni.chemistry == 'eq')
 
     # Surface single-scattering albedo
     surface_material = config.atmos_clim.agni.surf_material
@@ -197,6 +209,7 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
 
                         vol_dict, "",
 
+                        IO_DIR = io_dir,
                         flag_rayleigh = config.atmos_clim.rayleigh,
                         flag_cloud    = config.atmos_clim.cloud_enabled,
                         overlap_method  = config.atmos_clim.agni.overlap_method,
@@ -209,7 +222,6 @@ def init_agni_atmos(dirs:dict, config:Config, hf_row:dict):
                         evap_efficiency=config.atmos_clim.agni.evap_efficiency,
 
                         use_all_gases=include_all,
[…]
