+++ input/all_options.toml
@@ -186,7 +186,7 @@ version = "2.0"
     surf_state      = "fixed"   # surface scheme: "mixed_layer" | "fixed" | "skin"
     surf_greyalbedo = 0.1       # surface grey albedo
     albedo_pl       = 0.0   	# Enforced input value of bond albedo; float (0 to 1) or str (path to CSV)
-    rayleigh        = false     # enable rayleigh scattering
+    rayleigh        = true     # enable rayleigh scattering
     tmp_minimum     = 0.5       # temperature floor on solver
     tmp_maximum     = 5000.0    # temperature ceiling on solver
 
@@ -204,10 +204,11 @@ version = "2.0"
         solution_atol   = 0.5           # solver absolute tolerance
         solution_rtol   = 0.15          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
-        rainout         = false         # include volatile rainout upon condensation
-        latent_heat     = false         # include latent heat release by phase change? Requires `rainout=true`
+        condensation    = true         # include volatile rainout upon condensation
+        latent_heat     = false         # include latent heat release when `condensation=true`?
         convection      = true          # include convective heat transport, with MLT
-        conduction      = false         # include conductive heat transport, with Fourier's law
+        conduction      = true         # include conductive heat transport, with Fourier's law
+        sens_heat       = true          # include sensible heat flux near surface, with TKE scheme
         real_gas        = false         # use real-gas equations of state
         psurf_thresh    = 0.1           # bar, surface pressure where we switch to 'transparent' mode
         dx_max_ini      = 300.0         # initial maximum temperature step [kelvin] allowed by solver
@@ -222,6 +223,7 @@ version = "2.0"
         fastchem_xtol_elem    = 1e-4    # FC solver tolerance (elemental)
         ini_profile     = 'loglinear'   # Initial guess for temperature profile shape
         ls_default      = 2             # Default linesearch method (0:none, 1:gs, 2:bt)
+        fdo             = 2             # finite-difference order (options: 2, 4)
 
     [atmos_clim.janus]
         p_top           = 1.0e-6        # bar, top of atmosphere grid pressure
