+++ input/planets/trappist1c.toml
@@ -7,7 +7,8 @@ version = "2.0"
 [params]
     # output files
     [params.out]
-        path        = "hd63433d"
+        logging     = "DEBUG"
+        path        = "trappist1c"
         plot_mod    = 5      # Plotting frequency, 0: wait until completion | n: every n iterations
 
 # ----------------------------------------------------
@@ -51,16 +52,16 @@ version = "2.0"
     [atmos_clim.agni]
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
         spectral_group  = "Honeyside"   # which gas opacities to include
-        spectral_bands  = "256"         # how many spectral bands?
-        num_levels      = 60            # Number of atmospheric grid levels
-        chemistry       = "none"        # "none" | "eq"
+        spectral_bands  = "48"         # how many spectral bands?
+        num_levels      = 40            # Number of atmospheric grid levels
+        chemistry       = "eq"        # "none" | "eq"
         surf_material   = "greybody"    # surface material file for scattering
         solve_energy    = true         # solve for energy-conserving atmosphere profile
         solution_atol   = 1e-2          # solver absolute tolerance
         solution_rtol   = 5e-2          # solver relative tolerance
-        overlap_method    = "ee"          # gas overlap method
+        overlap_method  = "ee"          # gas overlap method
         condensation    = true          # volatile condensation
-        real_gas        = true          # use real-gas equations of state
+        real_gas        = false          # use real-gas equations of state
 
 # Volatile escape - physics table
 [escape]
@@ -74,7 +75,7 @@ version = "2.0"
     module = "spider"   # Which interior module to use
 
     [interior.spider]
-        num_levels      = 220       # Number of SPIDER grid levels
+        num_levels      = 200       # Number of SPIDER grid levels
         tolerance       = 1.0e-10   # solver tolerance
         tsurf_atol      = 20.0      # tsurf_poststep_change
         tsurf_rtol      = 0.01      # tsurf_poststep_change_frac
