+++ input/all_options.toml
@@ -215,7 +215,7 @@ version = "2.0"
         dx_max          = 35.0          # maximum temperature step [kelvin] allowed by solver
         max_steps       = 70            # max steps allowed by solver during each iteration
         perturb_all     = true          # updated entire jacobian each step?
-        mlt_criterion   = "l"           # MLT convection stability criterion; (l)edoux or (s)chwarzschild
+        mlt_criterion   = "s"           # MLT convection stability criterion; (l)edoux or (s)chwarzschild
         fastchem_floor        = 150.0   # Minimum temperature allowed to be sent to FC
         fastchem_maxiter_chem = 60000   # Maximum FC iterations (chemistry)
         fastchem_maxiter_solv = 20000   # Maximum FC iterations (internal solver)
