### 2025-11-18 a1a00ef
Update CITATION.cff, and added pre-commit hook
