+++ input/demos/aragog.toml
@@ -1,13 +1,12 @@
 # PROTEUS configuration file (version 2.0)
 version = "2.0"
-author  = "Harrison Nicholls, Tim Lichtenberg"
 
 # ----------------------------------------------------
 # Parameters
 [params]
     # output files
     [params.out]
-        path        = "TOI-1807b"
+        path        = "dummy_aragog"
         logging     = "INFO"
         plot_mod    = 1      # Plotting frequency, 0: wait until completion | n: every n iterations
         plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
@@ -32,19 +31,16 @@ author  = "Harrison Nicholls, Tim Lichtenberg"
 # ----------------------------------------------------
 # Star
 [star]
-    mass    = 0.8       # M_sun
+    mass    = 1.0       # M_sun
     age_ini = 0.100     # Gyr, model initialisation/start age
-    module  = "mors"
-    [star.mors]
-        rot_pcntle = 50.0       # rotation percentile
-        rot_period = "none"     # rotation period [days]
-        tracks  = "spada"       # evolution tracks: spada | baraffe
-        age_now = 4.567         # current age of star [Gyr]
-        spec    = "stellar_spectra/Named/trappist-1.txt" # path to stellar spectrum file
+    module  = "dummy"
+    [star.dummy]
+        radius  = 1.0       # R_sun
+        Teff    = 5772.0    # K
 
 # Orbital system
 [orbit]
-    semimajoraxis   = 0.012       # AU
+    semimajoraxis   = 1.0       # AU
     eccentricity    = 0.0       # dimensionless
     zenith_angle    = 48.19     # degrees
     s0_factor       = 0.375     # dimensionless
@@ -52,12 +48,11 @@ author  = "Harrison Nicholls, Tim Lichtenberg"
 
 # Planetary structure - physics table
 [struct]
-    radius_int  = 1.37       # R_earth
-    corefrac    = 0.55      # non-dim., radius fraction
+    radius_int  = 0.9       # R_earth
+    corefrac    = 0.65      # non-dim., radius fraction
 
 # Atmosphere - physics table
 [atmos_clim]
-    prevent_warming = true      # do not allow the planet to heat up
     surface_d       = 0.01      # m, conductive skin thickness
     surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
     surf_state      = "fixed"   # surface scheme: "mixed_layer" | "fixed" | "skin"
@@ -82,52 +77,45 @@ author  = "Harrison Nicholls, Tim Lichtenberg"
     [atmos_clim.dummy]
         gamma           = 0.01           # atmosphere opacity between 0 and 1
 
-# Volatile escape - physics table
 [escape]
 
-    module = "none"             # Which escape module to use
-    reservoir = "outgas"        # Reservoir that sets escaping gas composition
-
-    [escape.zephyrus]
-        Pxuv        = 5e-5          # Pressure at which XUV radiation become opaque in the planetary atmosphere [bar]
-        efficiency  = 0.1           # Escape efficiency factor
-        tidal       = false         # Tidal contribution enabled
+    reservoir = "bulk"  # Escaping reservoir: "bulk", "outgas", "pxuv".
+    module = "dummy"    # Which escape module to use
 
     [escape.dummy]
-        rate        = 0.0         # Bulk unfractionated escape rate [kg s-1]
+        rate        = 2.0e4         # Bulk unfractionated escape rate [kg s-1]
 
 [interior]
     grain_size      = 0.1               # crystal settling grain size [m]
-    F_initial       = 1e3               # Initial heat flux guess [W m-2]
+    F_initial       = 1e6               # Initial heat flux guess [W m-2]
[…]
