### 2025-11-18 ea7f047
Update trappist and toi6255 configs. Atmos_clim modules return ocean mass for each volatile. Fix deprecated call in logs.py. Store post-rainout Psurf
