### 2025-11-18 0f2392f
Improve model stability for AGNI by starting isothermal. Option for setting agni verbosity
