+++ input/planets/toi6255.toml
@@ -54,6 +54,7 @@ version = "2.0"
     module  = "agni"           # Which atmosphere module to use
 
     [atmos_clim.agni]
+        verbosity       = 2
         p_top           = 1.0e-5        # bar, top of atmosphere grid pressure
         spectral_group  = "Dayspring"   # which gas opacities to include
         spectral_bands  = "48"         # how many spectral bands?
@@ -66,6 +67,7 @@ version = "2.0"
         overlap_method    = "ee"          # gas overlap method
         condensation    = true          # volatile condensation
         real_gas        = true          # use real-gas equations of state
+        ini_profile     = 'isothermal'
 
 # Volatile escape - physics table
 [escape]
