### 2025-11-18 b196023
Update configs to distinguish between rainout and ocean parameters
