+++ input/planets/trappist1c.toml
@@ -60,7 +60,7 @@ version = "2.0"
         solution_atol   = 1e-2          # solver absolute tolerance
         solution_rtol   = 5e-2          # solver relative tolerance
         overlap_method  = "ee"          # gas overlap method
-        condensation    = true          # volatile condensation
+        rainout   = true          # volatile condensation
         real_gas        = false          # use real-gas equations of state
 
 # Volatile escape - physics table
