+++ input/planets/toi6255.toml
@@ -65,7 +65,7 @@ version = "2.0"
         solution_atol   = 1e-2          # solver absolute tolerance
         solution_rtol   = 5e-2          # solver relative tolerance
         overlap_method    = "ee"          # gas overlap method
-        condensation    = true          # volatile condensation
+        rainout   = true          # volatile condensation
         real_gas        = true          # use real-gas equations of state
         ini_profile     = 'isothermal'
 
