### 2025-11-18 0f3efe0
Split the run from the upload
The previous commit yielded "Error: Process completed with exit code 1".
Now we're splitting the two tasks in two blocks: first the run, then the upload.
Hopefully, this works.
