### 2025-11-19 e93af61
Update MUSCLES url with current structure
