### 2025-11-19 7058cf2
Run download_data function during install_all call. Pass surface roughness to AGNI
