### 2025-11-19 d4f6a02
Changed mors spec parameter to mors star_name in config -> will be used for more later. Updated stellar path
