+++ input/all_options.toml
@@ -116,7 +116,7 @@ version = "2.0"
         rot_period  = "none"    # rotation period [days]
         tracks      = "spada"   # evolution tracks: spada | baraffe
         age_now     = 4.567     # current age of star [Gyr]
-        spec        = "stellar_spectra/Named/sun.txt" # path to stellar spectrum file
+        star_name   = "sun"     # star name
 
     [star.dummy]
         radius           = 1.0      # Constant stellar radius [R_sun]
