+++ src/proteus/inference/BO.py
@@ -16,12 +16,16 @@ import time
 
 import matplotlib.pyplot as plt
 import torch
-from botorch.acquisition import LogExpectedImprovement, UpperConfidenceBound, qLogExpectedImprovement
+from botorch.acquisition import (
+    LogExpectedImprovement,
+    UpperConfidenceBound,
+    qLogExpectedImprovement,
+)
 from botorch.fit import fit_gpytorch_mll
-from botorch.optim.fit import fit_gpytorch_mll_torch
 from botorch.models import SingleTaskGP
 from botorch.models.transforms import Normalize, Standardize
 from botorch.optim import optimize_acqf
+from botorch.optim.fit import fit_gpytorch_mll_torch
 from gpytorch.mlls import ExactMarginalLogLikelihood
 from utils import get_kernel_w_prior
 
