+++ tests/inference/dummy.infer.toml
@@ -1,34 +1,34 @@
-# Config file for running a grid of forward models
+# Config file for inference scheme
 
 # Set seed for reproducibility
-seed = 2
+seed = 1
 
 # Path to output folder where inference will be saved (relative to PROTEUS output folder)
-output = "dummy_inference/"
+output = "infer_toy/BO/"
 
 # Path to base (reference) config file relative to PROTEUS root folder
-ref_config = "tests/inference/base.toml"
+ref_config = "input/demos/dummy.toml"
 
 # Method for initialising the inference scheme (one of these must be 'none')
 init_samps = 2        # Number of random samples if starting from scratch.
-init_grid  = 'none'   # Path pre-computed grid (relative to PROTEUS output folder)
+init_grid  = 'none' # grid_demo/'   # Path pre-computed grid (relative to PROTEUS output folder)
 
 # Parameters for Bayesian optimisation
 n_workers  = 2       # Number of parallel workers
-kernel     = "MAT"    # Kernel type for GP, "RBF" | "MAT"
-acqf       = "LogEI"  # Acquisition function, "UCB" | "LogEI"
-n_steps    = 30       # Total number of evaluations (i.e. BO steps)
-n_restarts = 10       # GP optimization restarts
-n_samples  = 1000     # Raw samples for acquisition optimization
+kernel     = "MAT3/2"    # Kernel type for GP, "RBF" | "MAT1/2" | "MAT3/2" | "MAT5/2"
+acqf       = "LogEI"  # Acquisition function, "UCB" | "LogEI" | "E-LogEI"
+n_steps    = 5      # Total number of evaluations (i.e. BO steps)
 
 # Parameters to optimize (with bounds)
 [parameters]
 "struct.mass_tot" = [0.7, 3.0]
 "struct.corefrac" = [0.3, 0.9]
-"delivery.elements.H_ppmw" = [6e3, 2e4]
-"outgas.fO2_shift_IW" = [-3.0, 5.0]
+# "delivery.elements.H_ppmw" = [1e3, 2e4]
+# "outgas.fO2_shift_IW" = [-3.0, 5.0]
+# "escape.dummy.rate" = [1e5, 1e7]
 
 # Target observables to match by optimisation
 [observables]
-"R_obs" = 7.9950245489e6
-"H2O_vmr" = 0.41
+"R_obs" = 7965615.42040000017732382
+# "H2O_vmr" = 0.69457002234999998
+"P_surf" = 16333.19793399999980466
