+++ tools/multiprofile_postprocess.jl
@@ -1,0 +1,272 @@
+#!/usr/bin/env -S julia
+
+PROTEUS_DIR = dirname(dirname(abspath(@__FILE__)))
+println("PROTEUS_DIR = $PROTEUS_DIR")
+
+ROOT_DIR = abspath(PROTEUS_DIR, "AGNI/")
+using Pkg
+Pkg.activate(ROOT_DIR)
+
+using Printf, LoggingExtras, NCDatasets, Glob, Dates, DataStructures
+using AGNI
+import AGNI.atmosphere as atmosphere
+import AGNI.energy as energy
+import AGNI.solver as solver
+
+function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::String, zenith_angle::Float64)
+    @info @sprintf("Setting up atmosphere with custom zenith_angle=%.1f\n", zenith_angle)
+
+    ds = Dataset(ncfile,"r")
+
+    try
+        nlev_c::Int = length(ds["p"][:])
+        input_pl::Array{Float64,1} = ds["pl"][:]
+
+        # Determine min/max pressure and convert to bar
+        p_top_pa = minimum(input_pl)
+        p_boa_pa = maximum(input_pl)
+        p_top_bar = p_top_pa / 1.0e5
+        p_boa_bar = p_boa_pa / 1.0e5
+
+        raw_gases::Array{Char,2} = ds["gases"][:,:]
+        num_gas::Int = size(raw_gases)[2]
+        input_gases::Array{String,1} = []
+        for i in 1:num_gas
+            push!(input_gases, strip(String(raw_gases[:,i])))
+        end
+
+        raw_vmrs::Array{Float64, 2} = ds["x_gas"][:,:]
+        input_vmrs_scalar::Dict{String, Float64} = Dict()
+        for i in 1:num_gas
+            g = input_gases[i]
+            input_vmrs_scalar[g] = raw_vmrs[i, end]
+        end
+
+        input_tsurf::Float64   = ds["tmp_surf"][1]
+        input_radius::Float64  = ds["planet_radius"][1]
+        input_gravity::Float64 = ds["surf_gravity"][1]
+        input_inst::Float64   = ds["instellation"][1]
+        input_s0fact::Float64 = ds["inst_factor"][1]
+        input_albedo::Float64 = ds["bond_albedo"][1]
+
+        input_flag_rayleigh::Bool = true
+        try; input_flag_rayleigh = Bool(ds["flag_rayleigh"][1] == 'y'); catch; end
+        input_flag_thermo::Bool = true
+        try; input_flag_thermo = Bool(ds["thermo_funct"][1] == 'y'); catch; end
+        input_flag_continuum::Bool = true
+        try; input_flag_continuum = Bool(ds["flag_continuum"][1] == 'y'); catch; end
+
+        close(ds)
+
+        if !ispath(spfile)
+            @error("Cannot find spectral file $spfile")
+            exit(1)
+        end
+
+        atmos = atmosphere.Atmos_t()
+        atmosphere.setup!(atmos, ROOT_DIR, output_dir,
+                                spfile,
+                                input_inst, input_s0fact, input_albedo, zenith_angle,
+                                input_tsurf,
+                                input_gravity, input_radius,
+                                nlev_c, p_boa_bar, p_top_bar,
+                                input_vmrs_scalar, "",
+                                flag_gcontinuum=input_flag_continuum,
+                                flag_rayleigh=input_flag_rayleigh,
+                                thermo_functions=input_flag_thermo,
+                                overlap_method="ro"
+                                )
+
+        code = atmosphere.allocate!(atmos, "")
+        if !code
[…]
