### 2025-11-21 295e5a3
Properly commented working script. Ready for first push to github.
