+++ tools/multiprofile_postprocess.jl
@@ -3,15 +3,20 @@
 PROTEUS_DIR = dirname(dirname(abspath(@__FILE__)))
 println("PROTEUS_DIR = $PROTEUS_DIR")
 
+#absolute path to AGNI environment for Julia
 ROOT_DIR = abspath(PROTEUS_DIR, "AGNI/")
 using Pkg
 Pkg.activate(ROOT_DIR)
 
+#import required packages
 using Printf, LoggingExtras, NCDatasets, Glob, Dates, DataStructures
 using AGNI
 import AGNI.atmosphere as atmosphere
 import AGNI.energy as energy
 import AGNI.solver as solver
+
+# function largely based on setup_atmosphere! in atmosphere.jl and setup_atmos_from_nc! from postprocess.jl but
+#for different zenith angles
 
 function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::String, zenith_angle::Float64)
     @info @sprintf("Setting up atmosphere with custom zenith_angle=%.1f\n", zenith_angle)
@@ -22,12 +27,13 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
         nlev_c::Int = length(ds["p"][:])
         input_pl::Array{Float64,1} = ds["pl"][:]
 
-        # Determine min/max pressure and convert to bar
+        # determine min/max pressure and convert to bar
         p_top_pa = minimum(input_pl)
         p_boa_pa = maximum(input_pl)
         p_top_bar = p_top_pa / 1.0e5
         p_boa_bar = p_boa_pa / 1.0e5
 
+        # read in gases and vmrs
         raw_gases::Array{Char,2} = ds["gases"][:,:]
         num_gas::Int = size(raw_gases)[2]
         input_gases::Array{String,1} = []
@@ -42,6 +48,7 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
             input_vmrs_scalar[g] = raw_vmrs[i, end]
         end
 
+        #get other inputs for atmosphere setup from .nc file
         input_tsurf::Float64   = ds["tmp_surf"][1]
         input_radius::Float64  = ds["planet_radius"][1]
         input_gravity::Float64 = ds["surf_gravity"][1]
@@ -49,6 +56,7 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
         input_s0fact::Float64 = ds["inst_factor"][1]
         input_albedo::Float64 = ds["bond_albedo"][1]
 
+        # flags required for atmosphere setup
         input_flag_rayleigh::Bool = true
         try; input_flag_rayleigh = Bool(ds["flag_rayleigh"][1] == 'y'); catch; end
         input_flag_thermo::Bool = true
@@ -58,11 +66,13 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
 
         close(ds)
 
+        #check for user in case the spectral file is missing
         if !ispath(spfile)
             @error("Cannot find spectral file $spfile")
             exit(1)
         end
 
+        # SETUP ATMOSPHERE with function from atmosphere.jl
         atmos = atmosphere.Atmos_t()
         atmosphere.setup!(atmos, ROOT_DIR, output_dir,
                                 spfile,
@@ -83,7 +93,9 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
             exit(1)
         end
 
-        # Load initial state from file
+        # Load initial state (final T-P profile form simulation) from file so we can start RCE from there
+        # to ensure easier convergence than starting from scratch
+
         ds = Dataset(ncfile, "r")
         try
[…]
