### 2025-11-21 d3e76b5
set up download functions for muscles and phoenix spectra
