### 2025-11-22 f2d9d89
Simplified usage and added comments.
