### 2025-11-22 36e8da8
Fix tests for inference. Update gitignore
