### 2025-11-23 0a26ee8
fix missing %s placeholder in warning string by removing % wl
