### 2025-11-24 1a4d5ef
Fix hanging test caused by blocking-process from multiprocessing module
