### 2025-11-24 1fe6390
Restore cache in tests
