### 2025-11-25 a03d77d
added albedo updating agni
