+++ input/albedo_test_cloudfree.toml
@@ -1,0 +1,425 @@
+# PROTEUS configuration file
+
+# This is a comprehensive outline of all configuration options. It includes variables
+# which have default values, in order to showcase the range of potential options available.
+# Variable defaults are defined in `src/proteus/config/*.py`
+
+# Root tables should be physical, with the exception of "params"
+# Software related options should go within the appropriate physical table
+# For configuration see https://fwl-proteus.readthedocs.io/en/latest/config.html
+
+# ----------------------------------------------------
+
+# The general structure is:
+#   [params]        parameters for code execution, output files, time-stepping, convergence
+#   [star]          stellar parameters, model selection
+#   [orbit]         planetary orbital parameters
+#   [struct]        planetary structure (mass, radius)
+#   [atmos_clim]    atmosphere climate parameters, model selection
+#   [atmos_chem]    atmosphere chemistry parameters, model selection
+#   [escape]        escape parameters, model selection
+#   [interior]      magma ocean model selection and parameters
+#   [outgas]        outgassing parameters (fO2) and included volatiles
+#   [delivery]      initial volatile inventory, and delivery model selection
+#   [observe]       synthetic observations
+
+# ----------------------------------------------------
+
+version = "2.0"
+
+# Parameters
+[params]
+    # output files
+    [params.out]
+        path        = "albedo_test_cloudfree"
+        logging     = "INFO"
+        plot_mod    = 3      # Plotting frequency, 0: wait until completion | n: every n iterations | none: do not plot
+        plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
+        write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
+        archive_mod = 5      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf = false    # Remove SOCRATES spectral file when simulation ends.
+
+    # time-stepping
+    [params.dt]
+        minimum      = 1e4    # absolute minimum time-step [years]
+        minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
+        maximum      = 1e7    # maximum time-step [yr]
+        initial      = 3e1    # inital step size [yr]
+        starspec     = 1e9    # interval to re-calculate the stellar spectrum [yr]
+        starinst     = 1e1    # interval to re-calculate the instellation [yr]
+        method       = "adaptive"  # proportional | adaptive | maximum
+
+        [params.dt.proportional]
+            propconst    = 52.0   # Proportionality constant
+
+        [params.dt.adaptive]
+            atol         = 0.02   # Step size atol
+            rtol         = 0.10   # Step size rtol
+
+    # Termination criteria
+    #     Set enabled=true/false in each section to enable/disable that termination criterion
+    [params.stop]
+
+        # Require criteria to be satisfied twice before model will exit?
+        strict = false
+
+        # required number of iterations
+        [params.stop.iters]
+            enabled = true
+            minimum = 5
+            maximum = 9000
+
+        # required time constraints
+        [params.stop.time]
+            enabled = true
+            minimum = 1.0e3      # model will certainly run to t > minimum [yr]
+            maximum = 4.567e+9   # model will terminate when t > maximum [yr]
+
+        # solidification
+        [params.stop.solid]
+            enabled  = true
+            phi_crit = 0.01      # non-dim., model will terminate when global melt fraction < phi_crit
+
+        # radiative equilibrium
+        [params.stop.radeqm]
+            enabled = true
+            atol    = 0.2        # absolute tolerance [W m-2]
[…]
