+++ input/planets/fiducial_sub_Neptune_Tang_comparison.toml
@@ -1,0 +1,281 @@
+# PROTEUS configuration file (version 2.0)
+
+# Root tables should be physical, with the exception of "params"
+# Software related options should go within the appropriate physical table
+
+# The general structure is:
+#   [root]     metadata
+#   [params]   parameters for code execution, output files, time-stepping, convergence
+#   [star]     stellar parameters, model selection
+#   [orbit]    planetary orbital parameters
+#   [struct]   planetary structure (mass, radius)
+#   [atmos]    atmosphere parameters, model selection
+#   [escape]   escape parameters, model selection
+#   [interior] magma ocean model selection and parameters
+#   [outgas]   outgassing parameters (fO2) and included volatiles
+#   [delivery] initial volatile inventory, and delivery model selection
+
+# ----------------------------------------------------
+# Metadata
+version = "2.0"
+
+# ----------------------------------------------------
+# Parameters
+[params]
+    # output files
+    [params.out]
+        path        = "fiducial_sub_Neptune"
+        logging     = "DEBUG"
+        plot_mod    = 1      # Plotting frequency, 0: wait until completion | n: every n iterations
+        plot_fmt    = "png"  # Plotting image file format, "png" or "pdf" recommended
+        write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
+
+    # time-stepping
+    [params.dt]
+        minimum      = 3e2    # yr, minimum time-step
+        maximum      = 1e7    # yr, maximum time-step
+        initial      = 1e3    # yr, inital step size
+        starspec     = 3e8    # yr, interval to re-calculate the stellar spectrum
+        starinst     = 1e2    # yr, interval to re-calculate the instellation
+        method       = "adaptive"  # proportional | adaptive | maximum
+
+        [params.dt.proportional]
+            propconst    = 52.0   # Proportionality constant
+
+        [params.dt.adaptive]
+            atol         = 0.03   # Step size atol
+            rtol         = 0.15   # Step size rtol
+
+    # Termination criteria
+    #     Set enabled=true/false in each section to enable/disable that termination criterion
+    [params.stop]
+
+        # Require criteria to be satisfied twice before model will exit?
+        strict = false
+
+        # required number of iterations
+        [params.stop.iters]
+            enabled = true
+            minimum = 5
+            maximum = 9000
+
+        # required time constraints
+        [params.stop.time]
+            enabled = true
+            minimum = 1.0e3     # yr, model will certainly run to t > minimum
+            maximum = 2.0e+10   # yr, model will terminate when t > maximum
+
+        # solidification
+        [params.stop.solid]
+            enabled  = true
+            phi_crit = 0.01  # non-dim., model will terminate when global melt fraction < phi_crit
+
+        # radiative equilibrium
+        [params.stop.radeqm]
+            enabled = true
+            atol    = 0.2     # absolute tolerance [W m-2]
+            rtol    = 1e-3    # relative tolerance
+
+        [params.stop.escape]
+            enabled   = false
+            mass_frac = 3e-4      # Stop when atm_mass < this frac of initial mass
+
+
+# ----------------------------------------------------
+# Star
+[star]
+
+    # Physical parameters
+    mass    = 0.465      # M_sun
+    age_ini = 0.100     # Gyr, model initialisation/start age
+
+    module  = "dummy"
+    [star.mors]
+        rot_pctle = 50.0    # rotation percentile
+        tracks  = "spada" # evolution tracks: spada | baraffe
[…]
