### 2025-11-26 e684398
Redirect print() calls in BOREAS to PROTEUS log.info stream, and safely catch exceptions
