+++ src/proteus/proteus.py
@@ -424,7 +424,7 @@ class Proteus:
             ############### ESCAPE
             if (self.loops["total"] > self.loops["init_loops"]+2) and (not self.desiccated):
                 PrintHalfSeparator()
-                run_escape(self.config, self.hf_row, self.interior_o.dt)
+                run_escape(self.config, self.hf_row, self.directories, self.interior_o.dt)
 
             ############### / ESCAPE
 
