### 2025-11-26 6bc681c
Fix typo in AGNI wrapper, for getting T_obs
