+++ input/demos/boreas.toml
@@ -79,7 +79,7 @@ version = "2.0"
     surf_greyalbedo  = 0.1      # surface grey albedo
     rayleigh        = false      # enable rayleigh scattering
 
-    module  = "dummy"           # Which atmosphere module to use
+    module  = "agni"           # Which atmosphere module to use
 
     [atmos_clim.dummy]
         gamma = 0.99
