### 2025-11-26 a3447a5
Further changes to script. Still slow.
