+++ tools/multiprofile_debug.jl
@@ -2,31 +2,19 @@
 
 # script to perform radiative-convective equilibrium postprocessing
 # at multiple zenith angles based on a single PROTEUS atmosphere output file
-# similar to postprocess.jl but loops over multiple angles and saves
-# separate .nc files for each angle.
-
-# Usage: julia multiprofile_postprocess.jl <output_dir> <angles>
-# where <output_dir> is the PROTEUS simulation output directory
-# and <angles> is a comma-separated list of zenith angles in degrees
-# Example: julia multiprofile_postprocess.jl ./outputdir 0,30,60,85
 
 PROTEUS_DIR = dirname(dirname(abspath(@__FILE__)))
 println("PROTEUS_DIR = $PROTEUS_DIR")
 
-#absolute path to AGNI environment for Julia
 ROOT_DIR = abspath(PROTEUS_DIR, "AGNI/")
 using Pkg
 Pkg.activate(ROOT_DIR)
 
-#import required packages
 using Printf, LoggingExtras, NCDatasets, Glob, Dates, DataStructures
 using AGNI
 import AGNI.atmosphere as atmosphere
 import AGNI.energy as energy
 import AGNI.solver as solver
-
-# function largely based on setup_atmosphere! in atmosphere.jl and setup_atmos_from_nc! from postprocess.jl but
-# for different zenith angles
 
 function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::String, zenith_angle::Float64)
     @info @sprintf("Setting up atmosphere with custom zenith_angle=%.1f\n", zenith_angle)
@@ -37,13 +25,11 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
         nlev_c::Int = length(ds["p"][:])
         input_pl::Array{Float64,1} = ds["pl"][:]
 
-        # determine min/max pressure and convert to bar
         p_top_pa = minimum(input_pl)
         p_boa_pa = maximum(input_pl)
         p_top_bar = p_top_pa / 1.0e5
         p_boa_bar = p_boa_pa / 1.0e5
 
-        # read in gases and vmrs
         raw_gases::Array{Char,2} = ds["gases"][:,:]
         num_gas::Int = size(raw_gases)[2]
         input_gases::Array{String,1} = []
@@ -58,17 +44,14 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
             input_vmrs_scalar[g] = raw_vmrs[i, end]
         end
 
-        #get other inputs for atmosphere setup from .nc file
         input_tsurf::Float64   = ds["tmp_surf"][1]
         input_radius::Float64  = ds["planet_radius"][1]
         input_gravity::Float64 = ds["surf_gravity"][1]
         input_inst::Float64   = ds["instellation"][1]
         # set instellation factor to 1.0 for RCE postprocessing
-        #input_s0fact::Float64 = 1.0
-        input_s0fact::Float64 = ds["inst_factor"][1]
+        input_s0fact::Float64 = 1.0
         input_albedo::Float64 = ds["bond_albedo"][1]
 
-        # flags required for atmosphere setup
         input_flag_rayleigh::Bool = true
         try; input_flag_rayleigh = Bool(ds["flag_rayleigh"][1] == 'y'); catch; end
         input_flag_thermo::Bool = true
@@ -78,13 +61,11 @@ function setup_atmos_multiangle!(output_dir::String, ncfile::String, spfile::Str
 
         close(ds)
 
-        #check for user in case the spectral file is missing
         if !ispath(spfile)
             @error("Cannot find spectral file $spfile")
             exit(1)
         end
 
[…]
