### 2025-11-28 652c745
add stellar distances dictionary
