### 2025-11-28 b46e2d5
add gj832_synth and fix pc -> cm conversion
