### 2025-11-28 ae2c7f6
make distance parameter default to stars_distance_pc, and use distance_cm after conversion
