### 2025-11-28 70e3d7b
fix NREL url to solar spectrum (still newguey2003.txt)
