### 2025-11-30 03dedd1
add option of downloading all stars in one database ; fix muscles url v23-24; add muscles stars from v23-24
