### 2025-11-30 79c6337
add fits unit (flux, wavelength) test
