### 2025-12-01 25ff51b
Debug FWL data
