### 2025-12-01 855ea47
Debug FWL data
