### 2025-12-01 fb43c0c
Debug FWL data, mkdir output
