+++ .github/workflows/tests.yaml
@@ -137,8 +137,8 @@ jobs:
         run: |
           CUR=$(pwd)
           cd $FWL_DATA
+          tree
           tree > $CUR/output/FWL_DATA.txt
-          echo $CUR/output/FWL_DATA.txt
           cd $CUR
 
       # Run PROTEUS tests
