### 2025-12-01 accaf19
Workflow permissions. And run FWL_DATA tree on failure
