+++ .github/workflows/tests.yaml
@@ -16,7 +16,7 @@ on:
 
 permissions:
   actions: write
-  contents: read
+  contents: write
 
 jobs:
   test:
