### 2025-12-01 7d45e59
command to get interior data
