### 2025-12-01 dcd9f50
Use PROTEUS downloader functions to get aragog data
