### 2025-12-01 a8fb84a
Aragog independence
