### 2025-12-01 55dc018
Pass lookup dir via config. Check files exist.
