+++ src/proteus/interior/aragog.py
@@ -176,13 +176,17 @@ class AragogRunner():
         # Get look up data directory, will be configurable in the future
         LOOK_UP_DIR = (
             FWL_DATA_DIR /
-            "interior_lookup_tables/1TPa-dK09-elec-free/"
-            "MgSiO3_Wolf_Bower_2018_1TPa/"
+            "interior_lookup_tables/" /
+            config.interior.lookup_dir
         )
         MELTING_DIR = (
             FWL_DATA_DIR /
             "interior_lookup_tables/Melting_curves/"
         )
+
+        # check data exist
+        if not (LOOK_UP_DIR / "heat_capacity_melt.dat").is_file():
+            raise FileNotFoundError(f"Aragog lookup data {LOOK_UP_DIR}")
 
         phase_liquid = _PhaseParameters(
             density = LOOK_UP_DIR / "density_melt.dat",
