### 2025-12-01 2f8e9d2
Style
