+++ mkdocs.yml
@@ -62,13 +62,13 @@ plugins:
     handlers:
       python:
         paths: [src]
-        # import:
-        # - https://installer.readthedocs.io/en/stable/objects.inv
-        # - https://docs.python.org/3/objects.inv
-        # - https://numpy.org/doc/stable/objects.inv
-        # - https://docs.scipy.org/doc/scipy/objects.inv
-        # - https://pandas.pydata.org/docs/objects.inv
-        # - https://matplotlib.org/stable/objects.inv
+        inventories:
+        - https://installer.readthedocs.io/en/stable/objects.inv
+        - https://docs.python.org/3/objects.inv
+        - https://numpy.org/doc/stable/objects.inv
+        - https://docs.scipy.org/doc/scipy/objects.inv
+        - https://pandas.pydata.org/docs/objects.inv
+        - https://matplotlib.org/stable/objects.inv
         options:
           docstring_style: numpy
           docstring_options:
