### 2025-12-01 fe99331
Fixed script - normal runtime and correct instellation factor
