### 2025-12-02 13173cd
Corrected convergence for hot lava planets handling
