### 2025-12-02 1e73c4e
handle different url & tar for gj832_synth
