+++ tools/get_stellar_spectrum.py
@@ -8,7 +8,7 @@ import sys
 import numpy as np
 
 stars_online = {
-    "muscles": ["gj1132", "gj1214", "gj15a", "gj163", "gj176", "gj436", "gj551", "gj581", "gj649", "gj667c", "gj674", "gj676a", "gj699", "gj729", "gj832", "gj849", "gj876", "hd40307", "hd85512", "hd97658", "l-980-5", "lhs-2686", "trappist-1", "v-eps-eri", "l-98-59", "hat-p-12", "hat-p-26", "hd-149026", "l-678-39", "lp-791-18", "toi-193", "wasp-127", "wasp-17", "wasp-43", "wasp-77a"], #"gj832_synth" requires different url, is in a zip somewhere else
+    "muscles": ["gj1132", "gj1214", "gj15a", "gj163", "gj176", "gj436", "gj551", "gj581", "gj649", "gj667c", "gj674", "gj676a", "gj699", "gj729", "gj832", "gj832_synth", "gj849", "gj876", "hd40307", "hd85512", "hd97658", "l-980-5", "lhs-2686", "trappist-1", "v-eps-eri", "l-98-59", "hat-p-12", "hat-p-26", "hd-149026", "l-678-39", "lp-791-18", "toi-193", "wasp-127", "wasp-17", "wasp-43", "wasp-77a"],
     "vpl": ["hd128167", "hd114710", "hd206860", "hd22049"],
     "nrel": ["sun"]
 }
@@ -50,12 +50,12 @@ star_distance_pc = {
     "hd-149026":    75.8643,
     "l-678-39":     9.44181,
     "lp-791-18":    26.4927,
-    "toi-173":      80.4373,
+    "toi-193":      80.4373,
     "wasp-127":     159.507,
     "wasp-17":      405.908,
     "wasp-43":      86.7467,
     "wasp-77a":     105.166,
-    "sun": 0.0               # For NREL case, distance is irrelevant
+    "sun":          0.0               # For NREL case, distance is irrelevant
 }
 
 def DownloadModernSpectrum(name, distance=None):
@@ -145,47 +145,88 @@ def DownloadModernSpectrum(name, distance=None):
     match database:
         case 'muscles':
             cert = certifi.where()
+            import tarfile
 
-            if lowres:
-                resstr = "const"
-            else:
-                resstr = "var"
+            if star == "gj832_synth":
+                if lowres:
+                    print("\tNote: lowres option ignored for synthetic GJ 832 spectrum")
 
-            ok = False
-            resp = None
+                tar_url = ("https://archive.stsci.edu/missions/hlsp/muscles/hlsp_muscles_model_multi_gj832_na_v1_synth-spec.tar.gz")
+                print(f"looking for synthetic model: {tar_url}")
+                resp = requests.get(tar_url, verify=cert)
 
-            version_sets = [
-                ("v22",      ["v22"]),
-                ("v23-v24",  ["v23", "v24"]),
-                ("v25",      ["v25"]),
-            ]
-
-            for dir_version, file_versions in version_sets:
-                for file_version in file_versions:
-                    source = (
-                        f"https://archive.stsci.edu/missions/hlsp/muscles/"
-                        f"{dir_version}/{star}/"
-                        f"hlsp_muscles_multi_multi_{star}_broadband_{file_version}_"
-                        f"adapt-{resstr}-res-sed.fits"
+                if resp.status_code == 404:
+                    raise Exception(
+                        "Could not find synthetic spectrum tarball for GJ 832 on MUSCLES website"
                     )
-                    print(f"looking for: {source}")
-                    resp = requests.get(source, verify=cert)
+                if resp.status_code != 200:
+                    print(
+                        "\t WARNING: Request returned with status code '%d' (should be 200/OK)"
[…]
