### 2025-12-02 10281d3
Remove test and make it warning; gj832_synth has no uunits in fits file
