### 2025-12-02 44b0269
change filename of spectra (remove spec_) to be just<star>.txt
