### 2025-12-05 6cd9d85
add download_phoenix which downloads and unzips PHOENIX zip based on alpha and FeH
