### 2025-12-05 f398d7f
add config parameters for PHOENIX (and muscles), will be expanded upon
