+++ src/proteus/config/_star.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from attrs import define, field
-from attrs.validators import ge, gt, in_
+from attrs.validators import ge, gt, in_, optional
 
 from ._converters import none_if_none
 
@@ -46,12 +46,29 @@ class Mors:
         Observed estimated age of the star [Gyr].
     star_name: str
         Name of the star, to find appropriate stellar spectrum. See [documentation](https://fwl-proteus.readthedocs.io/en/latest/data/#stars).
+    spectrum_source: str
+        Source of stellar spectra. Choices: 'named', 'muscles', 'phoenix', 'none'.
+    FeH: float
+        Stellar metallicity [Fe/H] to be used for PHOENIX synthetic spectra,
+        if spectrum_source is 'phoenix'.
+    alpha: float
+        Alpha-element enhancement [alpha/Fe] to be used for PHOENIX synthetic spectra,
+        if spectrum_source is 'phoenix'.
     """
     age_now         = field(default=None)
     star_name       = field(default=None)
     rot_pcntle      = field(default=None, converter=none_if_none)
     rot_period      = field(default=None, converter=none_if_none)
     tracks: str     = field(default='spada', validator=in_(('spada', 'baraffe')))
+
+    spectrum_source: str = field(
+        default="named",
+        validator=in_(("named", "muscles", "phoenix", "none")),
+    )
+
+    # Solar by default
+    FeH: float = field(default=0.0)   # [Fe/H]
+    alpha: float = field(default=0.0) # [alpha/Fe]
 
 def valid_stardummy(instance, attribute, value):
     if instance.module != "dummy":
@@ -99,6 +116,8 @@ class Star:
         Stellar mass [M_sun]. Note that for Mors,
         it should be between 0.1 and 1.25 solar masses.
         Values outside of the valid range will be clipped.
+    radius: float | None
+        Stellar radius [R_sun]. If 'none', radius will be calculated. # add more
     age_ini: float
         Age of system at model initialisation [Gyr].
     module: str | None
@@ -115,6 +134,7 @@ class Star:
     )
 
     mass: float = field(validator=gt(0))
+    radius: float | None = field(default=None, validator=optional(gt(0)))
     age_ini: float = field(validator=gt(0))
 
     mors: Mors       = field(factory=Mors,      validator=valid_mors)
