### 2025-12-05 7caa041
include phoenix/muscles download in get_sufficient if spectrum_source is set
