### 2025-12-05 7c4da84
if spectrum source is phoenix, FeH and alpha should be set (default is solar, only if explicitly set to none this raises an issue)
