### 2025-12-06 dcf957a
make download solar spectrum function and automatically download if spectrum_source is solar, or none
