### 2025-12-06 af29d26
add radius, log_g and Teff overrides to config (otherwise calculated with mors)
