### 2025-12-07 eb0cb12
start phoenix parameter pipeline for download of synthetic spectra
