### 2025-12-07 1f24b52
sun.txt in stellar_spectra/solar
