### 2025-12-07 55c6666
raise error if mass out of range
