+++ src/proteus/star/wrapper.py
@@ -146,6 +146,17 @@ def phoenix_to_grid(*, Teff, logg, FeH, alpha):
 
     return {"Teff": Teff, "logg": logg, "FeH": FeH, "alpha": alpha}
 
+def _phoenix_filename(Teff: float, logg: float, FeH: float, alpha: float) -> str:
+    """
+    Build PHOENIX filename like:
+    LTE_T02300_logg1.00_FeH+0.5_alpha+0.0_phoenixMedRes_R05000.txt
+    """
+    Tstr   = f"{int(round(Teff)):05d}"
+    logg_s = f"{logg:.2f}"
+    feh_s  = f"{FeH:+.1f}"
+    alpha_s = f"{alpha:+.1f}"
+    return f"LTE_T{Tstr}_logg{logg_s}_FeH{feh_s}_alpha{alpha_s}_phoenixMedRes_R05000.txt"
+
 def init_star(handler:Proteus):
     '''
     Star-related things to be done when the simulation begins.
