### 2025-12-08 cb1267c
add star_path for optional custom stellar path
