### 2025-12-08 3cc1a5a
adjust warning mass limits phoenix
