### 2025-12-09 0d22063
add function that finds correct phoenix file, scales and saves it
