### 2025-12-09 27540bc
if no spectrum source mentioned, try muscles first
