### 2025-12-10 bb957c0
mini updates before mergin main
