### 2025-12-10 d3d864b
Create new directory structure for PHOENIX to easily check if data is already present; and fix phoenix_param filename issue
