### 2025-12-10 a132f5b
small log changes
