### 2025-12-10 8371821
new file for phoenix functions (filehandling etc) to clean wrapper.py
