+++ src/proteus/star/phoenix.py
@@ -1,0 +1,262 @@
+# Everything that has to do with PHOENIX stellar spectra
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import numpy as np
+
+from proteus.utils.constants import AU, M_sun, R_sun, const_G
+from proteus.utils.data import GetFWLData, download_phoenix
+from proteus.utils.helper import UpdateStatusfile
+
+log = logging.getLogger("fwl."+__name__)
+
+if TYPE_CHECKING:
+    from proteus import Proteus
+
+# Mass limits on stellar tracks [Msun]
+MASS_LIM = {
+    "spada":   (0.10, 1.25),
+    "baraffe": (0.01, 1.40)
+}
+
+def phoenix_params(handler:Proteus, stellar_track=None, age_yr: float | None = None):
+    """
+    Build PHOENIX parameters.
+
+    Parameters
+    ----------
+    handler : Proteus
+        Proteus object instance
+    stellar_track :
+        MORS stellar track object (mors.Star or BaraffeTrack), if available.
+        If provided, it is used to compute Teff and radius when missing.
+    age_yr : float, optional
+        Stellar age in years at which PHOENIX should represent the star.
+        If None, defaults to config.star.mors.age_now.
+
+    Returns
+    -------
+    dict
+        Dictionary with keys: 'Teff', 'logg', 'radius', 'FeH', 'alpha'.
+
+        - Teff   : effective temperature [K]
+        - logg   : surface gravity log10(g [cgs])
+        - radius : stellar radius [R_sun]
+        - FeH    : [Fe/H]
+        - alpha  : [alpha/M]
+    """
+    star_cfg = handler.config.star
+    mors_cfg = star_cfg.mors
+
+    if age_yr is None:
+        age_yr = mors_cfg.age_now * 1e9  # Gyr -> yr
+
+    # Composition parameters from config
+    FeH   = mors_cfg.FeH
+    alpha = mors_cfg.alpha
+
+    if FeH == 0.0:
+        log.info("PHOENIX: Using solar metallicity [Fe/H]=0.0. "
+                "Set star.mors.FeH to change the composition.")
+    if alpha == 0.0:
+        log.info("PHOENIX: Using solar [alpha/Fe]=0.0. "
+                "Set star.mors.alpha to change the alpha fraction.")
+
+    # Overrides from config
+    Teff   = getattr(mors_cfg, "Teff", None)
+    radius = getattr(mors_cfg, "radius", None)  # [R_sun]
+    logg   = getattr(mors_cfg, "log_g", None)   # log10(g [cgs])
+
+    # If we have a stellar track, use it to fill Teff / radius
+    if stellar_track is not None:
+        age_Myr = age_yr / 1e6
+
+        # Track type: 'spada' vs 'baraffe'
+        track_type = mors_cfg.tracks
+
+        if Teff is None:
+            if track_type == "spada":
+                Teff = float(stellar_track.Value(age_Myr, "Teff"))
+            else:  # baraffe
+                Teff = float(stellar_track.BaraffeStellarTeff(age_yr))
+            log.info(f"Assuming calculated effective temperature {Teff:.0f} K from {track_type} tracks")
+
+        if radius is None:
+            if track_type == "spada":
+                radius = float(stellar_track.Value(age_Myr, "Rstar"))  # [R_sun]
[…]
