### 2025-12-10 e4b8cc0
include PHOENIX: in logs
