### 2025-12-10 90c4578
update spectrum_source to default to None, not solar
