### 2025-12-10 26700d9
adding new way of handling stellar spectra (solar, phoenix, or muscles)
