### 2025-12-10 c1ff2d7
include map to grid for phoenix zip handling
