+++ src/proteus/utils/data.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from time import sleep
 from typing import TYPE_CHECKING
 
+import numpy as np
 import platformdirs
 from osfclient.api import OSF
 
@@ -32,6 +33,7 @@ ARAGOG_BASIC = (
 
 log.debug(f'FWL data location: {FWL_DATA_DIR}')
 
+#### PHOENIX spectra utilities ####
 def _phoenix_param(x: float | int | str, kind: str) -> str:
     """
     Format a PHOENIX parameter.
@@ -52,7 +54,31 @@ def _phoenix_param(x: float | int | str, kind: str) -> str:
     # Normal case, e.g. +0.5, -1.0
     return f"{x:+0.1f}"
 
+def _phoenix_to_grid(FeH: float | int | str, alpha: float | int | str) -> tuple[float, float]:
+    """
+    Map (FeH, alpha) to the nearest PHOENIX composition grid point.
 
+    - FeH grid: [-4.0, -3.0, -2.0, -1.5, ..., +1.0] (step 0.5 from -2 to +1)
+    - alpha grid: -0.2 .. +1.2 in steps of 0.2
+
+    PHOENIX allows non-zero alpha only for FeH <= 0. For FeH > 0, alpha is forced to 0.
+    """
+    FeH = float(FeH)
+    alpha = float(alpha)
+
+    FeH_grid   = np.concatenate([[-4., -3.], np.arange(-2., 1. + 1e-6, 0.5)])
+    alpha_grid = np.arange(-0.2, 1.2 + 1e-6, 0.2)
+
+    # Map FeH to nearest grid value
+    FeH_g = float(FeH_grid[np.abs(FeH_grid - FeH).argmin()])
+
+    # Map alpha only if FeH <= 0; otherwise, force 0.0
+    if FeH_g <= 0.0:
+        alpha_g = float(alpha_grid[np.abs(alpha_grid - alpha).argmin()])
+    else:
+        alpha_g = 0.0
+
+    return FeH_g, alpha_g
 
 def download_zenodo_folder(zenodo_id: str, folder_dir: Path)->bool:
     """
@@ -536,8 +562,11 @@ def download_phoenix(alpha: float | int | str, FeH: float | int | str) -> bool:
     """
     phoenix_zenodo_id = "17674612"
 
-    feh_str   = _phoenix_param(FeH, kind="FeH")
-    alpha_str = _phoenix_param(alpha, kind="alpha")
+    # Map requested composition to nearest grid point
+    FeH_g, alpha_g = _phoenix_to_grid(FeH, alpha)
+
+    feh_str   = _phoenix_param(FeH_g,   kind="FeH")
+    alpha_str = _phoenix_param(alpha_g, kind="alpha")
     zip_name  = f"FeH{feh_str}_alpha{alpha_str}_phoenixMedRes_R05000.zip"
 
     data_dir = GetFWLData() / "stellar_spectra"
