### 2025-12-11 2ecbd7c
test to update the post-processing script to new proteus version. need to fix the phi_crit for solidification function
