+++ src/proteus/grid/post_processing_updated.py
@@ -1,0 +1,301 @@
+from __future__ import annotations
+
+import ast
+import os
+import re
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import toml
+
+
+def load_grid_cases(grid_dir: Path):
+    """
+    Load information for each simulation of a PROTEUS grid.
+    Read 'runtime_helpfile.csv', 'init_coupler.toml' and status
+    files for each simulation of the grid.
+
+    Parameters
+    ----------
+        grid_dir : Path or str
+            Path to the grid directory containing the 'case_*' folders
+
+    Returns
+    ----------
+        combined_data : list
+            List of dictionaries, each containing:
+                - 'init_parameters' (dict): Parameters loaded from `init_coupler.toml`.
+                - 'output_values' (pandas.DataFrame): Data from `runtime_helpfile.csv`.
+                - 'status' (str): Status string from the `status` file, or 'Unknown' if unavailable.
+    """
+
+    combined_data = []
+    grid_dir = Path(grid_dir)
+
+    # Collect and sort the case directories
+    case_dirs = list(grid_dir.glob('case_*'))
+    case_dirs.sort(key=lambda p: int(p.name.split('_')[1]))
+
+    for case in case_dirs:
+        runtime_file = case / 'runtime_helpfile.csv'
+        init_file = case / 'init_coupler.toml'
+        status_file = case / 'status'
+
+        # Load init parameters
+        init_params = {}
+        if init_file.exists():
+            try:
+                init_params = toml.load(open(init_file))
+            except Exception as e:
+                print(f"Error reading init file in {case.name}: {e}")
+
+        # Read runtime_helpfile.csv if available
+        df = None
+        if runtime_file.exists():
+            try:
+                df = pd.read_csv(runtime_file, sep='\t')
+            except Exception as e:
+                print(f"WARNING : Error reading runtime_helpfile.csv for {case.name}: {e}")
+
+        # Read status file if available
+        status = 'Unknown'
+        if status_file.exists():
+            try:
+                raw_lines = [ln.strip() for ln in status_file.read_text(encoding='utf-8').splitlines() if ln.strip()]
+                if len(raw_lines) >= 2:
+                    status = raw_lines[1]
+                elif raw_lines:
+                    status = raw_lines[0]
+                else:
+                    status = 'Empty'
+            except Exception as e:
+                print(f"WARNING : Error reading status file in {case.name}: {e}")
+        else:
+            print(f"WARNING : Missing status file in {case.name}")
+
+        # Combine all info about simulations into a list of dictionaries
+        combined_data.append({
+            'init_parameters': init_params,
+            'output_values'  : df,
+            'status'         : status
+        })
+
+    #
+    statuses = [c['status'] for c in combined_data]
+    status_counts = pd.Series(statuses).value_counts().sort_values(ascending=False)
+
+    print('-----------------------------------------------------------')
+    print(f"Total number of simulations: {len(statuses)}")
+    print('-----------------------------------------------------------')
+    print("Number of simulations per status:")
+    for st, count in status_counts.items():
+        print(f"  - {st:<45} : {count}")
+    print('-----------------------------------------------------------')
+
+    return combined_data
+
+def get_grid_parameters_from_toml(toml_path: str):
+    """
+    Extract grid parameter names and values from a PROTEUS ensemble TOML file.
+
+    Parameters
+    ----------
+    toml_path : str
+        Path to the ensemble TOML file (e.g. input/ensembles/blabla.toml)
+
+    Returns
+    -------
+    param_grid : dict
+        Dictionary where each key is a parameter name and the value is a list
+        of all grid values for that parameter.
+    """
+
[…]
