+++ input/planets/toi561b.toml
@@ -9,18 +9,18 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
     # output files
     [params.out]
         path        = "scratch/toi561b"
-        logging     = "DEBUG"
+        logging     = "INFO"
         plot_mod    = 5         # Plotting frequency, 0: wait until completion | n: every n iterations
         plot_fmt    = "png"     # Plotting image file format, "png" or "pdf" recommended
         write_mod   = 1         # Write CSV frequency, 0: wait until completion | n: every n iterations
-        archive_mod = 0      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
-        remove_sf   = true
+        archive_mod = 0         # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
+        remove_sf   = false      # Remove SOCRATES spectral file when simulation ends.
 
     # time-stepping
     [params.dt]
         minimum      = 3e2    # yr, minimum time-step
         minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
-        maximum      = 1.1e10   # yr, maximum time-step
+        maximum      = 1e7   # yr, maximum time-step
         initial      = 1e3    # yr, inital step size
         starspec     = 1e8    # yr, interval to re-calculate the stellar spectrum
         starinst     = 1e2    # yr, interval to re-calculate the instellation
@@ -64,8 +64,16 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
             rtol    = 1e-3      # relative tolerance
 
         [params.stop.escape]
+            enabled = false
+            p_stop  = 1.0        # Stop surface pressure is less than this value
+
+        # disintegration
+        [params.stop.disint]
             enabled   = false
-            p_stop    = 1.0     # bar, model will terminate with p_surf < p_stop
+            roche_enabled = false
+            offset_roche  = 0 # correction to calculated Roche limit [m]
+            spin_enabled  = false
+            offset_spin   = 0 # correction to calculated Breakup period [s]
 
 
 # ----------------------------------------------------
@@ -100,7 +108,12 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
     s0_factor       = 0.375     # dimensionless
 
     evolve  = false             # whether to evolve the SMaxis and eccentricity
-    module  = "none"
+    module  = "lovepy"          # module used to calculate tidal heating
+
+    axial_period = "none"       # planet's initial day length [hours]; will use orbital period if 'none'
+    satellite = false           # include satellite (moon)?
+    mass_sat = 7.347e+22        # mass of satellite [kg]
+    semimajoraxis_sat = 3e8     # initial SMA of satellite's orbit [m]
 
     [orbit.dummy]
         H_tide  = 1e-11     # Fixed tidal power density [W kg-1]
@@ -121,6 +134,7 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
     module       = "self"       # self | zalmoxis
 
     [struct.zalmoxis]
+        verbose                     = false     # verbose printing?
         coremassfrac                = 0.325     # core mass fraction [non-dim.]
         inner_mantle_mass_fraction  = 0         # inner mantle mass fraction [non-dim.]
         weight_iron_frac            = 0.325     # iron fraction in the planet [non-dim.]
@@ -154,18 +168,41 @@ author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
     module  = "agni"           # Which atmosphere module to use
 
     [atmos_clim.agni]
+        verbosity       = 1             # output verbosity for agni (0:none, 1:info, 2:debug)
[…]

[…]
