+++ src/proteus/utils/data.py
@@ -581,6 +581,13 @@ def download_phoenix(alpha: float | int | str, FeH: float | int | str) -> bool:
     log.info(f"Downloading PHOENIX spectra {zip_name}")
     log.info("This may take a while.")
 
+    # first download readme
+    get_zenodo_file(
+        zenodo_id=phoenix_zenodo_id,
+        folder_dir= data_dir / "PHOENIX",
+        zenodo_path="_readme.md",)
+
+    # then download zip
     if not get_zenodo_file(
         zenodo_id=phoenix_zenodo_id,
         folder_dir=folder_dir,
@@ -611,6 +618,12 @@ def download_muscles(star_name: str) -> bool:
     folder_dir = data_dir / "MUSCLES"
     star_filename = f"{star_name.strip().lower().replace(' ', '-').replace("gj-", 'gj')}.txt" # lowercase, and; "trappist 1" -> "trappist-1", but "gj 876" or "gj-876" -> "gj876"
     log.info(f"Downloading MUSCLES file {star_filename}")
+
+    # first download readme
+    get_zenodo_file(
+        zenodo_id=muscles_zenodo_id,
+        folder_dir= data_dir / "MUSCLES",
+        zenodo_path="_readme.md",)
 
     return get_zenodo_file(
         zenodo_id=muscles_zenodo_id,
@@ -662,6 +675,7 @@ def _get_sufficient(config:Config, clean:bool=False):
         src = config.star.mors.spectrum_source
 
         if src == "solar":
+            log.info("Spectrum source set to 'solar'; downloading solar spectrum.")
             download_solar_spectrum()
 
         elif src is None:
@@ -678,6 +692,7 @@ def _get_sufficient(config:Config, clean:bool=False):
                 download_solar_spectrum()
 
         elif src == "muscles":
+            log.info("Spectrum source set to 'muscles'. Downloading MUSCLES spectrum.")
             muscles_ok = download_muscles(config.star.mors.star_name)
             if not muscles_ok:
                 log.error(f"Could not download MUSCLES spectrum for star {config.star.mors.star_name}.")
@@ -685,6 +700,7 @@ def _get_sufficient(config:Config, clean:bool=False):
                 log.error("If no observed spectrum is available, consider using a PHOENIX synthetic spectrum by setting star.mors.spectrum_source = 'phoenix'.")
 
         elif src == "phoenix":
+            log.info("Spectrum source set to 'phoenix'. Downloading PHOENIX grid.")
             FeH = config.star.mors.FeH
             alpha = config.star.mors.alpha
             log.info(f"Downloading PHOENIX spectra with [Fe/H]={FeH:.2f}, [alpha/Fe]={alpha:.2f}.")
