### 2025-12-11 332a222
introduce phoenix helper file for phoenix file handling / grid mapping
