+++ src/proteus/utils/phoenix_helper.py
@@ -1,0 +1,132 @@
+#PHOENIX spectra utilities
+from __future__ import annotations
+
+import logging
+
+import numpy as np
+
+log = logging.getLogger("fwl."+__name__)
+
+def phoenix_param(x: float | int | str, kind: str) -> str:
+    """
+    Format a PHOENIX parameter.
+
+    kind:
+        "FeH"   -> FeH = 0.0  -> "-0.0" (handles inconsistency in filenames)
+        "alpha" -> alpha = 0.0 -> "+0.0"
+    """
+    x = float(x)
+
+    # zero case
+    if abs(x) < 1e-9:
+        if kind.lower() == "feh":
+            return "-0.0"
+        elif kind.lower() == "alpha":
+            return "+0.0"
+
+    # Normal case, e.g. +0.5, -1.0
+    return f"{x:+0.1f}"
+
+def phoenix_to_grid(*, FeH, alpha, Teff=None, logg=None):
+    """
+    Map (FeH, alpha, Teff, logg) to the nearest PHOENIX grid point.
+
+    Parameters
+    ----------
+    FeH : float
+        Metallicity [Fe/H].
+    alpha : float
+        [alpha/M].
+    Teff : float, optional
+        Effective temperature [K]. If provided, it is snapped to the PHOENIX
+        Teff grid.
+    logg : float, optional
+        Surface gravity log10(g [cgs]). If provided, it is snapped to the
+        PHOENIX log g grid.
+
+    Rules
+    -----
+    - FeH grid: [-4.0, -3.0, -2.0, -1.5, ..., +1.0]
+    - alpha grid: -0.2 .. +1.2 in steps of 0.2
+    - Teff grid: 2300–7000 (100 K), 7200–12000 (200 K)
+    - logg grid: 0.0–6.0 in steps of 0.5
+
+    Alpha is allowed to be nonzero only if:
+      - FeH <= 0, and
+      - (if Teff is given) 3500 K <= Teff <= 8000 K (using the mapped Teff).
+
+    If Teff is not provided, the Teff condition is ignored and only FeH <= 0
+    is enforced.
+
+    Returns
+    -------
+    dict
+        Dictionary with keys: 'Teff', 'logg', 'FeH', 'alpha', where 'Teff' and
+        'logg' may be None if not given.
+    """
+    FeH  = float(FeH)
+    alpha = float(alpha)
+
+    # Grids
+    FeH_grid   = np.concatenate([[-4., -3.], np.arange(-2., 1. + 1e-6, 0.5)])
+    alpha_grid = np.arange(-0.2, 1.2 + 1e-6, 0.2)
+
+    Teff_grid = None
+    logg_grid = None
+
+    # Map FeH
+    FeH_g = float(FeH_grid[np.abs(FeH_grid - FeH).argmin()])
+
+    # Map Teff if provided
+    Teff_g = None
+    if Teff is not None:
+        Teff = float(Teff)
+        Teff_grid = np.concatenate([
+            np.arange(2300., 7000. + 1e-6, 100.),
+            np.arange(7200., 12000. + 1e-6, 200.),
+        ])
+        Teff_g = float(Teff_grid[np.abs(Teff_grid - Teff).argmin()])
+
+    # Map logg if provided
+    logg_g = None
+    if logg is not None:
+        logg = float(logg)
+        logg_grid = np.arange(0., 6. + 1e-6, 0.5)
+        logg_g = float(logg_grid[np.abs(logg_grid - logg).argmin()])
+
+    # Decide whether alpha is allowed to be non-zero
+    if Teff_g is not None:
[…]
