### 2025-12-11 6db8008
make phoenix.py consistent with utils.phoenix_helper and handle offline mode
