### 2025-12-11 7947e04
make phoenix parts consistent with phoenix_helper.py, and fix minor issues
