### 2025-12-11 e84517f
update logging
