### 2025-12-11 9eb39a5
start on updating reference data page for stellar spectra (in progress)
