### 2025-12-12 e57ce5a
more on phoenix, custom and include all muscles star info (stellar spectra)
