+++ docs/data.md
@@ -20,127 +20,533 @@ ls $FWL_DATA/stellar_spectra/solar
 
 For PHOENIX spectra, files are stored under: `$FWL_DATA/stellar_spectra/PHOENIX`, where each subdirectory corresponds to a metallicity / alpha combination, e.g. FeH-0.5_alpha+0.0/.
 
-### Custom stellar spectrum
+### Using solar spectrum
 
-### PHOENIX use
+To use the NREL observed spectrum of the Sun, set `spectrum_source = "solar"` or do not set the parameter at all, and set `star_name = "sun"`.
+
+####  The Sun
+
+* **Star name:**  `sun`
+* **URL:** https://en.wikipedia.org/wiki/Sun
+* **Spectral type:**  G2V
+* **Teff:**  5772 K
+* **Age:**   4.6 Gyr
+* **Luminosity:**   1.0 solar
+* **Mass:**     1.0 solar
+* **Radius:**     1.0 solar
+* **Source:** [Gueymard 2003](https://www.sciencedirect.com/science/article/pii/S0038092X03003967), [NREL](https://www.nrel.gov/grid/solar-resource/spectra.html)
+
+### Using a (Mega-)Muscles observed spectrum
+
+To use an observed stellar spectrum by the Muscles or Mega-Muscles survey, set `spectrum_source = "muscles"` and set `star_name` to the name of the star. Make sure you use star names as listed below, since stars are often known under different names. More information about MUSCLES spectra can be found here: https://zenodo.org/records/17802209 (where the scaled MUSCLES spectra are stored), or here: https://archive.stsci.edu/hlsp/muscles (original MUSCLES spectra).
+
+#### Epsilon Eridani
+
+- **Star name:** `v-eps-eri`
+- **URL:** https://exoplanetarchive.ipac.caltech.edu/overview/eps%20Eri
+- **Spectral type:** K2V
+- **Teff:** 5020 K
+- **Age:** 400–800 Myr
+- **Luminosity:** 0.32 L☉
+- **Mass:** 0.82 M☉
+- **Radius:** 0.759 R☉
+- **Source:** MUSCLES
+
+---
+
+#### GJ 1132
+
+- **Star name:** `gj1132`
+- **URL:** https://exoplanetarchive.ipac.caltech.edu/overview/GJ%201132
+- **Spectral type:** M4.5 V
+- **Teff:** 3229 K
+- **Age:** ?
+- **Luminosity:** 0.005 L☉
+- **Mass:** 0.195 M☉
+- **Radius:** 0.221 R☉
+- **Source:** Mega-MUSCLES
+
+---
+
+#### GJ 1214
+
+- **Star name:** `gj1214`
+- **URL:** https://exoplanetarchive.ipac.caltech.edu/overview/GJ%201214
+- **Spectral type:** M4 V
+- **Teff:** 3100 K
+- **Age:** 5–10 Gyr
+- **Luminosity:** 0.00351 L☉
+- **Mass:** 0.182 M☉
+- **Radius:** 0.216 R☉
+- **Source:** MUSCLES
+
+---
+
+#### HD 85512
+
+- **Star name:** `hd85512`
+- **URL:** https://exoplanetarchive.ipac.caltech.edu/overview/HD%2085512
+- **Spectral type:** M0V
+- **Teff:** ~4400 K
+- **Age:** ~6 Gyr
[…]
