### 2025-12-12 962f4d4
small muscles adjustments (stellar spectra)
