### 2025-12-12 bef6455
small phoenix adjustments (stellar spectra)
