### 2025-12-12 504acba
add non_if_none converter for phoenix params
