### 2025-12-12 5776119
add override star path for custom stellar spectrum
