### 2025-12-12 f130b97
update custom stllar spectrum info
