### 2025-12-12 8842e6d
log when custom spectrum is used
