### 2025-12-12 00b695a
Small logging fix
