### 2025-12-12 aaaa8f3
only download sun.txt if not already present
