### 2025-12-13 c3616e8
update stellar spectra config parameters
