+++ input/all_options.toml
@@ -112,11 +112,22 @@ version = "2.0"
     module  = "mors"
 
     [star.mors]
-        rot_pcntle  = 50.0      # rotation percentile
-        rot_period  = "none"    # rotation period [days]
-        tracks      = "spada"   # evolution tracks: spada | baraffe
-        age_now     = 4.567     # current age of star [Gyr]
-        star_name   = "sun"     # star name
+        rot_pcntle  = 50.0          # rotation percentile
+        rot_period  = "none"        # rotation period [days]
+        tracks      = "spada"       # evolution tracks: spada | baraffe
+        age_now     = 4.567         # current age of star [Gyr]
+        star_name   = "sun"         # star name
+        star_path   = "none"        # optional override star path to custom stellar spectrum
+        spectrum_source = "solar"   # Spectrum source: 'solar' for solar spectra; 'muscles' for MUSCLES spectra; 'phoenix' for synthetic PHOENIX spectrum.
+
+        # PHOENIX parameters, only relevant if spectrum_source = "phoenix". Defaults to solar (0.0).
+        FeH         = 0.0           # metallicity [Fe/H]
+        alpha       = 0.0           # alpha enhancement [α/M]
+
+        # if None, calculated by mors
+        radius      = "none"        # Stellar radius [R_sun] used for PHOENIX spectrum scaling
+        log_g       = "none"        # Stellar surface gravity [dex]
+        Teff        = "none"        # Stellar effective temperature [K]
 
     [star.dummy]
         radius           = 1.0      # Constant stellar radius [R_sun]
