### 2025-12-14 aebdf53
Update Teff, log g and radius from float | None to float | str to avoid test failures due to None
