### 2025-12-14 48d9cec
add simple test script for stellar spectra pipeline
