+++ tests/test_stellar_spectra.py
@@ -1,0 +1,392 @@
+# tests/test_spectrum_pipeline.py
+#
+# Tests for the PROTEUS stellar spectrum pipeline:
+# - phoenix_helper grid/filename/param stuff
+# - PHOENIX raw -> scaled to 1 AU path creation (fake downloads)
+# - init_star() spectrum selection (solar / MUSCLES / star_path / PHOENIX)
+#
+# Can run offline and without the real mors module
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+from types import SimpleNamespace
+
+import numpy as np
+import pytest
+
+# tests/test_stellar_spectra.py
+
+
+
+# Helpers
+
+def _write_spectrum_file(path: Path, wl=(100.0, 200.0), fl=(1.0, 2.0)) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    np.savetxt(path, np.column_stack([np.array(wl), np.array(fl)]))
+
+
+def _make_handler(
+    *,
+    tmp_path: Path,
+    offline: bool,
+    Teff: float = 5800.0,
+    logg: float = 4.5,
+    radius: float = 1.0,
+    FeH: float = 0.0,
+    alpha: float = 0.0,
+    tracks: str = "spada",
+    age_now: float = 4.6,  # Gyr
+    mass: float = 1.0,
+):
+    mors = SimpleNamespace(
+        age_now=age_now,
+        FeH=FeH,
+        alpha=alpha,
+        tracks=tracks,
+        Teff=Teff,
+        radius=radius,
+        log_g=logg,
+    )
+    star = SimpleNamespace(mass=mass, mors=mors)
+    params = SimpleNamespace(offline=offline)
+    config = SimpleNamespace(star=star, params=params)
+
+    outdir = tmp_path / "out"
+    (outdir / "data").mkdir(parents=True, exist_ok=True)
+
+    handler = SimpleNamespace(
+        config=config,
+        directories={
+            "fwl": str(tmp_path),
+            "output": str(outdir),              # <-- add this
+            "output/data": str(outdir / "data") # <-- and this
+        },
+    )
+    return handler
+
+
+def _raw_phoenix_path(tmp_path: Path, raw_name: str, FeH_str: str, alpha_str: str) -> Path:
+    return (
+        tmp_path
+        / "stellar_spectra"
+        / "PHOENIX"
+        / f"FeH{FeH_str}_alpha{alpha_str}"
+        / raw_name
+    )
+
+
+def _install_fake_mors(monkeypatch):
+    """
+    Minimal fake 'mors' module so proteus.star.wrapper.init_star() can run
+    through the Baraffe branch without the real mors install.
+    """
+
+    class FakeBaraffeTrack:
+        def __init__(self, Mstar):
+            self.Mstar = float(Mstar)
+
+        def BaraffeLuminosity(self, age_yr):
+            return 1.0
+
+        # not required by init_star, but harmless if accessed
+        def BaraffeSolarConstant(self, age_yr, sep_au):
+            return 1361.0
+
+        def BaraffeStellarRadius(self, age_yr):
+            return 1.0
+
+        def BaraffeStellarTeff(self, age_yr):
+            return 5800.0
+
+    def ModernSpectrumLoad(star_modern_path, star_backup_path):
+        data = np.loadtxt(star_modern_path)
+        return data[:, 0], data[:, 1]
+
+    fake = SimpleNamespace(
+        BaraffeTrack=FakeBaraffeTrack,
[…]
