+++ src/proteus/grid/post_processing_updated.py
@@ -95,7 +95,7 @@ def load_grid_cases(grid_dir: Path):
 
     return combined_data
 
-def get_grid_parameters_from_toml(toml_path: str):
+def get_grid_parameters_from_toml(grid_dir: str):
     """
     Extract grid parameter names and values from a PROTEUS ensemble TOML file.
 
@@ -110,6 +110,9 @@ def get_grid_parameters_from_toml(toml_path: str):
         Dictionary where each key is a parameter name and the value is a list
         of all grid values for that parameter.
     """
+
+    grid_dir = Path(grid_dir)
+    toml_path = grid_dir / "copy.grid.toml"
 
     if not os.path.exists(toml_path):
         print(f"Error: TOML file not found at {toml_path}")
@@ -229,73 +232,108 @@ def extract_grid_output(cases_data: list, parameter_name: str):
 
     return parameter_values
 
-def extract_solidification_time(cases_data: list):
-    """
-    Extract the solidification time at the time step where the condition
-    'Phi_global' < phi_crit is first satisfied for each planet.
+def load_phi_crit(grid_dir: str | Path):
+    grid_dir = Path(grid_dir)
+    ref_file = grid_dir / "ref_config.toml"
 
-    Parameters
-    ----------
-    cases_data : list
-        List of dictionaries containing simulation data.
+    if not ref_file.exists():
+        raise FileNotFoundError(f"ref_config.toml not found in {grid_dir}")
 
-    phi_crit : float
-        The critical melt fraction value below which a planet is considered solidified.
-        A typical value is 0.005.
+    ref = toml.load(open(ref_file))
 
-    Returns
-    -------
-    solidification_times : list
-        A list containing the solidification times for all solidified planets of the grid.
-        If a planet never solidifies, it will have a NaN in the list.
-    """
+    # Navigate structure safely
+    try:
+        phi_crit = ref["params"]["stop"]["solid"]["phi_crit"]
+    except KeyError:
+        raise KeyError("phi_crit not found in ref_config.toml")
+
+    return phi_crit
+
+def extract_solidification_time(cases_data: list, grid_dir: str | Path):
+    # Load phi_crit once
+    phi_crit = load_phi_crit(grid_dir)
 
     solidification_times = []
     columns_printed = False
 
     for i, case in enumerate(cases_data):
         df = case['output_values']
-        # Check if the required columns exist in the dataframe
+
         if df is None:
-            solidification_times.append(np.nan)  # Append NaN if no output values
+            solidification_times.append(np.nan)
             continue
 
         if 'Phi_global' in df.columns and 'Time' in df.columns:
-            solidification_params = case['init_parameters'].get('params.stop.solid.phi_crit')
-            if solidification_params is None or 'phi_crit' not in solidification_params:
-                raise ValueError(f"Error: 'phi_crit' not found in init_parameters of case {i}. ")
-            phi_crit = solidification_params['phi_crit']
-
             condition = df['Phi_global'] < phi_crit
+
             if condition.any():
-                first_index = condition.idxmax()
-                solid_time = df.loc[first_index, 'Time'] # Get the index of the time at which the condition is first satisfied
-                solidification_times.append(solid_time)
+                idx = condition.idxmax()
+                solidification_times.append(df.loc[idx, 'Time'])
             else:
-                solidification_times.append(np.nan)  # Append NaN if condition is not satisfied
+                solidification_times.append(np.nan)
+
         else:
             if not columns_printed:
-                print("Warning: 'Phi_global' and/or 'Time' columns not found in some cases.")
-                print(f"Available columns: {', '.join(df.columns)}")
+                print("Warning: Missing Phi_global or Time column.")
+                print("Columns available:", df.columns.tolist())
                 columns_printed = True
[…]
