+++ src/proteus/grid/post_processing_updated.py
@@ -1,8 +1,5 @@
 from __future__ import annotations
 
-import ast
-import os
-import re
 from pathlib import Path
 
 import numpy as np
@@ -95,101 +92,66 @@ def load_grid_cases(grid_dir: Path):
 
     return combined_data
 
-def get_grid_parameters_from_toml(grid_dir: str):
+def get_tested_grid_parameters(cases_data: list, grid_dir: str | Path):
     """
-    Extract grid parameter names and values from a PROTEUS ensemble TOML file.
+    Extract tested grid parameters per case using:
+      - copy.grid.toml to determine which parameters were varied
+      - init_parameters already loaded by load_grid_cases
 
     Parameters
     ----------
-    toml_path : str
-        Path to the ensemble TOML file (e.g. input/ensembles/blabla.toml)
+    cases_data : list
+        Output of load_grid_cases.
+    grid_dir : str or Path
+        Path to the grid directory containing copy.grid.toml.
 
     Returns
     -------
-    param_grid : dict
-        Dictionary where each key is a parameter name and the value is a list
-        of all grid values for that parameter.
+    case_params : dict
+        Dictionary mapping case index -> {parameter_name: value}
+    tested_params : dict
+        Dictionary of tested grid parameters and their grid values
+        (directly from copy.grid.toml)
     """
 
     grid_dir = Path(grid_dir)
-    toml_path = grid_dir / "copy.grid.toml"
 
-    if not os.path.exists(toml_path):
-        print(f"Error: TOML file not found at {toml_path}")
-        return {}
+    # ---------------------------------------------------------
+    # 1) Load tested grid parameter definitions
+    # ---------------------------------------------------------
+    tested_params = toml.load(grid_dir / "copy.grid.toml")
 
-    with open(toml_path, "r") as f:
-        lines = f.readlines()
+    # Keep only actual grid parameters (ignore control keys)
+    tested_params = {
+        k: v for k, v in tested_params.items()
+        if "." in k
+    }
 
-    param_grid = {}
+    grid_param_paths = list(tested_params.keys())
 
-    # Regex patterns
-    section_pattern = re.compile(r'^\s*\["(.+?)"\]\s*$')
-    method_pattern = re.compile(r'^\s*method\s*=\s*"(.+?)"\s*$')
-    values_pattern = re.compile(r'^\s*values\s*=\s*(\[.*\])\s*$')
-    start_pattern  = re.compile(r'^\s*start\s*=\s*([0-9.eE+-]+)\s*$')
-    stop_pattern   = re.compile(r'^\s*stop\s*=\s*([0-9.eE+-]+)\s*$')
-    count_pattern  = re.compile(r'^\s*count\s*=\s*(\d+)\s*$')
+    # ---------------------------------------------------------
+    # 2) Extract those parameters from loaded cases
+    # ---------------------------------------------------------
+    case_params = {}
 
-    current_param = None
-    current_method = None
-    start = stop = count = None
+    for idx, case in enumerate(cases_data):
+        params_for_case = {}
+        init_params = case["init_parameters"]
 
-    for line in lines:
-        line = line.strip()
+        for path in grid_param_paths:
+            keys = path.split(".")
+            val = init_params
 
-        # Skip empty lines and comments
-        if not line or line.startswith("#"):
-            continue
-
-        # Detect new parameter block
-        section_match = section_pattern.match(line)
-        if section_match:
-            current_param = section_match.group(1)
-            current_method = None
-            start = stop = count = None
-            continue
-
-        if current_param is None:
-            continue
-
-        # Detect method
-        method_match = method_pattern.match(line)
-        if method_match:
-            current_method = method_match.group(1)
-            continue
-
-        # Direct method values
-        values_match = values_pattern.match(line)
[…]
