+++ src/proteus/grid/test_new_pp_grid.py
@@ -9,7 +9,9 @@ grid_name = "escape_grid_1Msun"
 grid_path = Path(f"/projects/p315557/Paper_1/DATA/Grids/{grid_name}")
 
 data = pp.load_grid_cases(grid_path)
-case_params = pp.get_tested_grid_parameters(data,grid_path)
-pp.generate_summary_csv(data,case_params,grid_path,grid_name)
-pp.generate_completed_summary_csv(data,case_params,grid_path,grid_name)
-pp.generate_running_error_summary_csv(data,case_params,grid_path,grid_name)
+# case_params, tested_params = pp.get_tested_grid_parameters(data,grid_path)
+# pp.generate_summary_csv(data,case_params,grid_path,grid_name)
+# pp.generate_completed_summary_csv(data,case_params,grid_path,grid_name)
+# pp.generate_running_error_summary_csv(data,case_params,grid_path,grid_name)
+
+pp.plot_grid_status(data,grid_path,grid_name)
