+++ src/proteus/grid/post_processing_updated.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
+import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -124,13 +125,14 @@ def get_tested_grid_parameters(cases_data: list, grid_dir: str | Path):
     # ---------------------------------------------------------
     # 1) Load tested grid parameter definitions
     # ---------------------------------------------------------
-    tested_params = toml.load(grid_dir / "copy.grid.toml")
+    raw_params = toml.load(grid_dir / "copy.grid.toml")
 
-    # Keep only actual grid parameters (ignore control keys)
-    tested_params = {
-        k: v for k, v in tested_params.items()
-        if "." in k
-    }
+    # Keep only the parameters and their values (ignore 'method' keys)
+    tested_params = {}
+    for key, value in raw_params.items():
+        if isinstance(value, dict) and "values" in value:
+            # Only store the 'values' list
+            tested_params[key] = value["values"]
 
     grid_param_paths = list(tested_params.keys())
 
@@ -507,8 +509,9 @@ def plot_grid_status(cases_data: list, grid_dir: str | Path, grid_name: str):
     total_simulations = len(cases_data)
     for i, count in enumerate(status_counts.values):
         percentage = (count / total_simulations) * 100
+        offset = 0.005 * status_counts.max()  # 1% of the max count
         ax.text(
-            i, count + 1,
+            i, count + offset,
             f"{count} ({percentage:.1f}%)",
             ha='center', va='bottom', fontsize=14
         )
@@ -519,8 +522,7 @@ def plot_grid_status(cases_data: list, grid_dir: str | Path, grid_name: str):
         f"Total number of simulations : {total_simulations}",
         transform=plt.gca().transAxes,
         ha='right', va='top',
-        fontsize=16,
-        bbox=dict(boxstyle="round,pad=0.3", fc="white", ec="black", lw=1)
+        fontsize=16
     )
 
     plt.grid(alpha=0.2, axis='y')
@@ -533,6 +535,206 @@ def plot_grid_status(cases_data: list, grid_dir: str | Path, grid_name: str):
 
     output_dir = grid_dir / "post_processing" / "grid_plots"
     output_dir.mkdir(parents=True, exist_ok=True)
-    output_file = output_dir / f"{grid_name}_summary_grid_statuses.png"
+    output_file = output_dir / f"summary_grid_statuses_{grid_name}.png"
     plt.savefig(output_file, dpi=300)
     plt.close()
+
+def group_output_by_parameter(df,grid_parameters,outputs):
+    """
+    Groups output values (like solidification times) by a specific grid parameter.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        DataFrame containing simulation results including value of the grid parameter and the corresponding extracted output.
+
+    grid_parameters : str
+        Column name of the grid parameter to group by (like 'escape.zephyrus.Pxuv').
+
+    outputs : str
+        Column name of the output to extract (like 'solidification_time').
+
+    Returns
+    -------
+    dict
+        Dictionary where each key is of the form '[output]_per_[parameter]', and each value is a dict {param_value: [output_values]}.
+    """
+    grouped = {}
+
+    for param in grid_parameters:
+        for output in outputs:
+            key_name = f"{output}_per_{param}"
+            value_dict = {}
+            for param_value in df[param].dropna().unique():
+                subset = df[df[param] == param_value]
+                output_values = subset[output].replace([np.inf, -np.inf], np.nan)
+                output_values = output_values.dropna()
+                output_values = output_values[output_values > 0]  # Remove zeros and negatives
+
[…]
