### 2025-12-16 0ffcfc0
Update input/all_options.toml
Co-authored-by: Tim Lichtenberg <lichtenberg.tim@gmail.com>
