+++ input/all_options.toml
@@ -118,7 +118,7 @@ version = "2.0"
         age_now     = 4.567         # current age of star [Gyr]
         star_name   = "sun"         # star name
         star_path   = "none"        # optional override star path to custom stellar spectrum
-        spectrum_source = "solar"   # Spectrum source: 'solar' for solar spectra; 'muscles' for MUSCLES spectra; 'phoenix' for synthetic PHOENIX spectrum.
+        spectrum_source = "solar"   # Spectrum source: 'solar' for solar spectra; 'muscles' for MUSCLES spectra; 'phoenix' for synthetic PHOENIX spectrum; see docs/data.md
 
         # PHOENIX parameters, only relevant if spectrum_source = "phoenix". Defaults to solar (0.0).
         FeH         = 0.0           # metallicity [Fe/H]
