### 2025-12-16 a277944
Wrong star used (should be trappist-1) -> fix
