### 2025-12-16 a8e3fae
Fix grid error, alpha should be forced to zero when FeH = -4
