### 2025-12-16 d08a9b2
Add get functionality for solar, muscles, phoenix spectra (and log in wrapper.py)
