### 2025-12-17 3f3d641
create toml grid analyze and use it for ecdf plot, clean and update functions comments
