+++ src/proteus/grid/test_new_pp_grid.py
@@ -8,65 +8,51 @@ import post_processing_updated as pp
 
 grid_name = "escape_grid_1Msun"
 grid_path = Path(f"/projects/p315557/Paper_1/DATA/Grids/{grid_name}")
-
 # grid_name= "toi561b_grid"
 # grid_path = Path(f"/projects/p315557/TOI-561b/{grid_name}")
-
 # grid_name= "escape_on_off_janus_agni_1Msun"
 # grid_path = Path(f"/projects/p315557/Paper_1/DATA/Comparison_escape_on_off/{grid_name}")
 
+# Load grid data
 data = pp.load_grid_cases(grid_path)
+# Store the input parameters for each case based on the tested parameters in the grid and store the tested parameters in the grid
 input_param_grid_per_case, tested_params_grid = pp.get_tested_grid_parameters(data,grid_path)
-#pp.generate_summary_csv(data,case_params,grid_path,grid_name)
-#pp.generate_completed_summary_csv(data,case_params,grid_path,grid_name)
-#pp.generate_running_error_summary_csv(data,case_params,grid_path,grid_name)
 
-#pp.plot_grid_status(data,grid_path,grid_name)
-rows = []
-for case_index, case in enumerate(data):
-    row = {}
+# Generate CSV file for all grid, only completed simulations, and running error simulations
+#pp.generate_summary_csv(data,input_param_grid_per_case,grid_path,grid_name)
+#pp.generate_completed_summary_csv(data,input_param_grid_per_case,grid_path,grid_name)
+#pp.generate_running_error_summary_csv(data,input_param_grid_per_case,grid_path,grid_name)
 
-    # Add input parameters
-    init_params = case['init_parameters']
-    for k, v in init_params.items():
-        row[k] = v
+# Load the final extracted data CSV for all simulations
+path_csv_all_simulations = grid_path / "post_processing/extracted_data" / f"{grid_name}_final_extracted_data_all.csv"
+all_simulations_data_csv = pd.read_csv(path_csv_all_simulations, sep="\t")
+# Plot grid status
+pp.plot_grid_status(all_simulations_data_csv,grid_path,grid_name)
 
-    # Add outputs (last timestep)
-    df_out = case['output_values']
-    if df_out is not None:
-        last_row = df_out.iloc[-1].to_dict()
-        row.update(last_row)
+# Load the final extracted data CSV for completed simulations
+path_csv_completed_simulations = grid_path / "post_processing/extracted_data" / f"{grid_name}_final_extracted_data_completed.csv"
+completed_simulations_data_csv = pd.read_csv(path_csv_completed_simulations, sep="\t")
 
-    # Add status
-    row['status'] = case['status']
+# Group output data by tested parameters in the grid
+grouped_data = {}
+columns_output = ['solidification_time', 'Phi_global', 'T_surf', 'P_surf', 'atm_kg_per_mol', 'esc_rate_total']
+for col in columns_output:
+    group = pp.group_output_by_parameter(
+        completed_simulations_data_csv,
+        tested_params_grid,
+        [col]
+    )
+    grouped_data.update(group)
 
-    rows.append(row)
-
-df = pd.DataFrame(rows)
-def flatten_nested_columns(df):
-    flat_data = {}
-    for col in df.columns:
-        if isinstance(df[col].iloc[0], dict):
-            # Expand each dictionary into separate columns
-            nested_df = pd.json_normalize(df[col])
-            nested_df.index = df.index
-            for nested_col in nested_df.columns:
-                flat_data[f"{col}.{nested_col}"] = nested_df[nested_col]
-        else:
-            flat_data[col] = df[col]
-    return pd.DataFrame(flat_data)
-df_flat = flatten_nested_columns(df)
[…]
