+++ src/proteus/grid/post_processing_updated.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 import matplotlib as mpl
+import matplotlib.cm as cm
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -10,8 +11,27 @@ import seaborn as sns
 import toml
 
 # ---------------------------------------------------------
-# Data loading and processing functions
+# Data loading, extraction, and CSV generation functions
 # ---------------------------------------------------------
+
+def get_grid_name(grid_path: str | Path) -> str:
+    """
+    Returns the grid name (last part of the path) from the given grid path.
+
+    Parameters
+    ----------
+    grid_path : str or Path
+        Full path to the grid directory.
+
+    Returns
+    -------
+    grid_name : str
+        Name of the grid directory.
+    """
+    grid_path = Path(grid_path)
+    if not grid_path.is_dir():
+        raise ValueError(f"{grid_path} is not a valid directory")
+    return grid_path.name
 
 def load_grid_cases(grid_dir: Path):
     """
@@ -28,7 +48,7 @@ def load_grid_cases(grid_dir: Path):
     ----------
         combined_data : list
             List of dictionaries, each containing:
-                - 'init_parameters' (dict): Parameters loaded from `init_coupler.toml`.
+                - 'init_parameters' (dict): All input parameters loaded from `init_coupler.toml`.
                 - 'output_values' (pandas.DataFrame): Data from `runtime_helpfile.csv`.
                 - 'status' (str): Status string from the `status` file, or 'Unknown' if unavailable.
     """
@@ -53,7 +73,7 @@ def load_grid_cases(grid_dir: Path):
             except Exception as e:
                 print(f"Error reading init file in {case.name}: {e}")
 
-        # Read runtime_helpfile.csv if available
+        # Read runtime_helpfile.csv
         df = None
         if runtime_file.exists():
             try:
@@ -61,7 +81,7 @@ def load_grid_cases(grid_dir: Path):
             except Exception as e:
                 print(f"WARNING : Error reading runtime_helpfile.csv for {case.name}: {e}")
 
-        # Read status file if available
+        # Read status file
         status = 'Unknown'
         if status_file.exists():
             try:
@@ -84,10 +104,9 @@ def load_grid_cases(grid_dir: Path):
             'status'         : status
         })
 
-    #
+    # Print summary of statuses
     statuses = [c['status'] for c in combined_data]
     status_counts = pd.Series(statuses).value_counts().sort_values(ascending=False)
-
     print('-----------------------------------------------------------')
     print(f"Total number of simulations: {len(statuses)}")
     print('-----------------------------------------------------------')
@@ -101,8 +120,8 @@ def load_grid_cases(grid_dir: Path):
 def get_tested_grid_parameters(cases_data: list, grid_dir: str | Path):
     """
     Extract tested grid parameters per case using:
-      - copy.grid.toml to determine which parameters were varied
-      - init_parameters already loaded by load_grid_cases
+      - copy.grid.toml to determine which parameters were varied in the grid
+      - init_parameters already loaded by load_grid_cases for each simulation of the grid
 
     Parameters
     ----------
@@ -116,15 +135,12 @@ def get_tested_grid_parameters(cases_data: list, grid_dir: str | Path):
     case_params : dict
         Dictionary mapping case index -> {parameter_name: value}
     tested_params : dict
-        Dictionary of tested grid parameters and their grid values
-        (directly from copy.grid.toml)
+        Dictionary of tested grid parameters and their grid values (directly from copy.grid.toml)
     """
 
     grid_dir = Path(grid_dir)
 
-    # ---------------------------------------------------------
[…]

[…]
