### 2025-12-17 4be722a
update the command and doc
