+++ src/proteus/grid/run_grid_analysis.py
@@ -1,213 +1,0 @@
-# This script is used to post-process a grid of PROTEUS simulations.
-# It extracts the output values from the simulation cases, saves them to a CSV file,
-# and generates plots for the grid statuses based on the extracted data. It also generates
-# ECDF single plots and a big grid plot for the input parameters vs extracted outputs.
-
-# The users need to specify the path to the grid directory and the grid name. (see the example below)
-# He also needs to specify the output columns to extract from the 'runtime_helpfile.csv' of each case and
-# update the related plotting variables accordingly. This can be done in the `run_grid_analyze` function (see below).
-from __future__ import annotations
-
-import os
-
-import matplotlib.cm as cm
-
-from proteus.grid.post_processing_grid import (
-    ecdf_grid_plot,
-    ecdf_single_plots,
-    extract_grid_output,
-    extract_solidification_time,
-    get_grid_parameters,
-    group_output_by_parameter,
-    load_extracted_data,
-    load_grid_cases,
-    plot_dir_exists,
-    plot_grid_status,
-    save_completed_cases,
-    save_error_running_cases,
-    save_grid_data_to_csv,
-)
-
-
-def run_grid_analyze(path_to_grid: str, grid_name: str, update_csv: bool = True):
-    """
-    Run the post-processing of a PROTEUS grid, extracting simulation data in a CSV file and generating plots.
-
-    Parameters
-    ----------
-    path_to_grid : str
-        Path to the directory containing the grid folder.
-
-    grid_name : str
-        Name of the grid folder to process.
-
-    update_csv : bool, optional
-        If True, the CSV file will be updated or created. If False, it will skip the CSV extraction step if the file already exists.
-    """
-
-    # ------------------------------------------------------------
-    #  1) Build all the folder/filename strings
-    # ------------------------------------------------------------
-    grid_path = os.path.join(path_to_grid, grid_name) + os.sep
-    postprocess_path = os.path.join(grid_path, "post_processing_grid") + os.sep
-    data_dir = os.path.join(postprocess_path, "extracted_data") + os.sep
-    os.makedirs(data_dir, exist_ok=True)
-
-
-    plots_path = os.path.join(postprocess_path, "plots_grid") + os.sep
-    plot_dir_exists(plots_path)
-
-    csv_file = os.path.join(data_dir, f"{grid_name}_extracted_data.csv")
-
-    # ------------------------------------------------------------
-    #  2) Define which outputs to pull from each case's runtime_helpfile.csv
-    # ------------------------------------------------------------
-    # User choose the output to extract from 'runtime_helpfile.csv' of each case (always the [-1] element of the column).
-    # For the units, check the file src/proteus/utils/coupler.py, lines 348-400 (keys)
-
-    output_to_extract = ['Time','esc_rate_total','Phi_global','P_surf','T_surf','T_eqm','M_planet','R_obs','p_xuv', 'R_xuv', 'F_xuv', 'atm_kg_per_mol',
-                        'H_kg_atm','O_kg_atm','C_kg_atm','N_kg_atm','S_kg_atm', 'Si_kg_atm', 'Mg_kg_atm', 'Fe_kg_atm', 'Na_kg_atm',
-                        'H2O_kg_atm','CO2_kg_atm', 'O2_kg_atm', 'H2_kg_atm', 'CH4_kg_atm', 'CO_kg_atm', 'N2_kg_atm', 'NH3_kg_atm',
-                        'S2_kg_atm', 'SO2_kg_atm', 'H2S_kg_atm', 'SiO_kg_atm','SiO2_kg_atm', 'MgO_kg_atm', 'FeO2_kg_atm', 'runtime']
-
-    # ------------------------------------------------------------
[…]
