### 2025-12-17 64febce
Remove folder_list file and simulations configuration files for my grid
