+++ folder_list.txt
@@ -1,958 +1,0 @@
-case_000000
-case_000001
-case_000002
-case_000003
-case_000004
-case_000005
-case_000006
-case_000007
-case_000008
-case_000009
-case_000010
-case_000011
-case_000012
-case_000013
-case_000014
-case_000015
-case_000016
-case_000017
-case_000018
-case_000019
-case_000020
-case_000021
-case_000022
-case_000023
-case_000024
-case_000025
-case_000026
-case_000027
-case_000028
-case_000029
-case_000030
-case_000031
-case_000032
-case_000033
-case_000034
-case_000035
-case_000036
-case_000037
-case_000038
-case_000039
-case_000040
-case_000041
-case_000042
-case_000043
-case_000044
-case_000045
-case_000046
-case_000047
-case_000048
-case_000049
-case_000050
-case_000051
-case_000052
-case_000053
-case_000054
-case_000055
-case_000056
-case_000057
-case_000058
-case_000059
-case_000060
-case_000061
-case_000062
-case_000063
-case_000064
-case_000065
-case_000066
-case_000067
-case_000068
-case_000069
-case_000070
-case_000071
-case_000072
-case_000073
-case_000074
-case_000075
-case_000076
-case_000077
-case_000078
-case_000079
-case_000080
-case_000081
-case_000082
-case_000083
-case_000084
-case_000085
-case_000086
-case_000087
-case_000088
-case_000089
-case_000090
-case_000091
-case_000092
-case_000093
-case_000094
-case_000095
-case_000096
-case_000097
-case_000098
-case_000099
-case_000100
-case_000101
-case_000102
-case_000103
-case_000104
-case_000105
-case_000106
-case_000107
-case_000108
-case_000109
-case_000110
-case_000111
-case_000112
-case_000113
-case_000114
-case_000115
-case_000116
-case_000117
-case_000118
-case_000119
-case_000120
-case_000121
-case_000122
-case_000123
-case_000124
-case_000125
-case_000126
-case_000127
-case_000128
-case_000129
-case_000130
-case_000131
-case_000132
-case_000133
-case_000134
-case_000135
-case_000136
-case_000137
-case_000138
-case_000139
-case_000140
-case_000141
-case_000142
-case_000143
-case_000144
-case_000145
-case_000146
-case_000147
-case_000148
-case_000149
-case_000150
-case_000151
-case_000152
-case_000153
-case_000154
-case_000155
-case_000156
-case_000157
-case_000158
-case_000159
-case_000160
-case_000161
-case_000162
-case_000163
-case_000164
-case_000165
-case_000166
-case_000167
-case_000168
-case_000169
-case_000170
-case_000171
-case_000172
-case_000173
-case_000174
-case_000175
-case_000176
-case_000177
-case_000178
-case_000179
-case_000180
-case_000181
-case_000182
-case_000183
-case_000184
-case_000185
-case_000186
-case_000187
-case_000188
-case_000189
-case_000190
-case_000191
-case_000192
-case_000193
-case_000194
-case_000195
-case_000196
-case_000197
-case_000198
-case_000199
-case_000200
-case_000201
-case_000202
-case_000203
-case_000204
-case_000205
-case_000206
-case_000207
-case_000208
-case_000209
-case_000210
-case_000211
-case_000212
-case_000213
-case_000214
-case_000215
-case_000216
-case_000217
-case_000218
-case_000219
-case_000220
-case_000221
-case_000222
-case_000223
-case_000224
-case_000225
-case_000226
-case_000227
-case_000228
-case_000229
-case_000230
-case_000231
-case_000232
-case_000233
-case_000234
-case_000235
-case_000236
-case_000237
-case_000238
-case_000239
[…]
