+++ input/demos/escape_grid_1Msun.toml
@@ -1,354 +1,0 @@
-# PROTEUS configuration file (version 2.0)
-
-# Root tables should be physical, with the exception of "params"
-# Software related options should go within the appropriate physical table
-
-# The general structure is:
-#   [root]     metadata
-#   [params]   parameters for code execution, output files, time-stepping, convergence
-#   [star]     stellar parameters, model selection
-#   [orbit]    planetary orbital parameters
-#   [struct]   planetary structure (mass, radius)
-#   [atmos]    atmosphere parameters, model selection
-#   [escape]   escape parameters, model selection
-#   [interior] magma ocean model selection and parameters
-#   [outgas]   outgassing parameters (fO2) and included volatiles
-#   [delivery] initial volatile inventory, and delivery model selection
-#   [observe]  synthetic observations
-
-# ----------------------------------------------------
-# Metadata
-version = "2.0"
-author  = "Harrison Nicholls, Tim Lichtenberg, Emma Postolec"
-
-# ----------------------------------------------------
-# Parameters
-[params]
-    # output files
-    [params.out]
-        path        = "scratch/escape_grid_1Msun"
-        logging     = "INFO"
-        plot_mod    = 5      # Plotting frequency, 0: wait until completion | n: every n iterations
-        plot_fmt    = "pdf"  # Plotting image file format, "png" or "pdf" recommended
-        write_mod   = 1      # Write CSV frequency, 0: wait until completion | n: every n iterations
-        archive_mod = 0      # Archive frequency, 0: wait until completion | n: every n iterations | none: do not archive
-        remove_sf   = true
-
-    # time-stepping
-    [params.dt]
-        minimum      = 3e2    # yr, minimum time-step
-        minimum_rel  = 1e-5   # relative minimum time-step [dimensionless]
-        maximum      = 3e7    # yr, maximum time-step
-        initial      = 1e4    # yr, inital step size
-        starspec     = 100e6    # yr, interval to re-calculate the stellar spectrum
-        starinst     = 100    # yr, interval to re-calculate the instellation
-        method       = "adaptive"  # proportional | adaptive | maximum
-
-        [params.dt.proportional]
-            propconst    = 52.0   # Proportionality constant
-
-        [params.dt.adaptive]
-            atol         = 0.02   # Step size atol
-            rtol         = 0.07   # Step size rtol
-
-    # Termination criteria
-    #     Set enabled=true/false in each section to enable/disable that termination criterion
-    [params.stop]
-
-        # Require criteria to be satisfied twice before model will exit?
-        strict = false
-
-        # required number of iterations
-        [params.stop.iters]
-            enabled = true
-            minimum = 5
-            maximum = 9000
-
-        # required time constraints
-        [params.stop.time]
-            enabled = true
-            minimum = 1.0e3     # yr, model will certainly run to t > minimum
-            maximum = 4.567e+9     # yr, model will terminate when t > maximum
-
-        # solidification
-        [params.stop.solid]
-            enabled  = true
-            phi_crit = 0.005  # non-dim., model will terminate when global melt fraction < phi_crit
-
-        # radiative equilibrium
-        [params.stop.radeqm]
-            enabled = true    # Tim said false in the meeting, true is a try for this grid
-            atol    = 0.2     # absolute tolerance [W m-2]
-            rtol    = 1e-3    # relative tolerance
-
-        # atmospheric escape
-        [params.stop.escape]
-            enabled   = true
-            p_stop    = 1.0  # bar, model will terminate with p_surf < p_stop
-
-
[…]

[…]
