### 2025-12-19 30ba3a5
change phoenix parameter names in config from e.g. alpha to phoenix_alpha, FeH -> phoenix_FeH, etc, and update scripts/tests accordingly
