+++ src/proteus/config/_star.py
@@ -19,8 +19,8 @@ def valid_mors(instance, attribute, value):
     src = instance.mors.spectrum_source
 
     if src == "phoenix":
-        if instance.mors.alpha is None or instance.mors.FeH is None:
-            raise ValueError("mors.alpha and mors.FeH must be set when using PHOENIX spectra")
+        if instance.mors.phoenix_alpha is None or instance.mors.phoenix_FeH is None:
+            raise ValueError("mors.phoenix_alpha and mors.phoenix_FeH must be set when using PHOENIX spectra")
 
     set_pcntle = instance.mors.rot_pcntle is not None
     set_period = instance.mors.rot_period is not None
@@ -51,22 +51,22 @@ class Mors:
     age_now: float
         Observed estimated age of the star [Gyr].
     star_name: str
-        Name of the star, to find appropriate stellar spectrum. See [documentation](https://fwl-proteus.readthedocs.io/en/latest/data/#stars).
+        Name of the star, to find appropriate stellar spectrum. See [documentation](https://fwl-proteus.readthedocs.io/en/latest/data.html).
     star_path: str
         Path to custom stellar spectra. If 'none', star_name will be used to find spectra in default locations.
     spectrum_source: str
         Source of stellar spectra. Choices: 'solar', 'muscles', 'phoenix', 'none'.
-    FeH: float
+    phoenix_FeH: float
         Stellar metallicity [Fe/H] to be used for PHOENIX synthetic spectra,
         if spectrum_source is 'phoenix'.
-    alpha: float
+    phoenix_alpha: float
         Alpha-element enhancement [alpha/Fe] to be used for PHOENIX synthetic spectra,
         if spectrum_source is 'phoenix'.
-    radius: float
+    phoenix_radius: float
         Stellar radius [R_sun]. If 'none', radius will be calculated using mors' stellar tracks, if spectrum_source is 'phoenix'.
-    log_g: float
+    phoenix_log_g: float
         Surface gravity [cgs]. If 'none', log g will be calculated will be calculated using mors' stellar tracks, if spectrum_source is 'phoenix'.
-    Teff: float
+    phoenix_Teff: float
         Effective temperature [K]. If 'none', Teff will be calculated will be calculated using mors' stellar tracks, if spectrum_source is 'phoenix'.
     """
 
@@ -82,13 +82,13 @@ class Mors:
     ### PHOENIX parameters
 
     # Solar by default
-    FeH: float = field(default=0.0)   # [Fe/H]
-    alpha: float = field(default=0.0) # [alpha/Fe]
+    phoenix_FeH: float = field(default=0.0)   # [Fe/H]
+    phoenix_alpha: float = field(default=0.0) # [alpha/Fe]
 
     # calculated if none
-    radius: float | str = field(default=None, validator=optional(gt(0)), converter=none_if_none)
-    log_g: float | str  = field(default=None, validator=optional(gt(0)),  converter=none_if_none)
-    Teff: float | str = field(default=None, validator=optional(gt(0)), converter=none_if_none)
+    phoenix_radius: float | str = field(default=None, validator=optional(gt(0)), converter=none_if_none)
+    phoenix_log_g: float | str  = field(default=None, validator=optional(gt(0)),  converter=none_if_none)
+    phoenix_Teff: float | str = field(default=None, validator=optional(gt(0)), converter=none_if_none)
 
 def valid_stardummy(instance, attribute, value):
     if instance.module != "dummy":
