### 2025-12-21 6f1ebf7
add option for young solar spectra in stellar_spectra/solar, automatically downloaded when spectrum_source = solar. name mapping including to avoid lowercase/uppercase problems. update tests and documentation
