### 2025-12-22 0c61448
Update src/proteus/grid/post_processing.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
