### 2025-12-22 a75fbc5
Update src/proteus/grid/post_processing.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
