### 2025-12-27 f3b6102
Updated configs for chili paper
