### 2025-12-27 7368538
add constraints file to avoid install of too high numpy version, where np.trapz was removed (still used in e.g. mors)
