+++ .github/workflows/contraints.txt
@@ -1,0 +1,1 @@
+numpy<2.4
