### 2025-12-27 d718f32
fix typo in name constrains.txt and add to tests.yaml
