### 2025-12-27 7f13c8f
use other download function that does not look for osf files for solar spectra
