+++ tests/observe/test_observe.py
@@ -1,0 +1,9 @@
+"""
+Tests for proteus.observe module
+"""
+from __future__ import annotations
+
+
+def test_placeholder():
+    """Placeholder test - replace with actual tests"""
+    pass
