+++ tests/atmos_chem/test_atmos_chem.py
@@ -1,0 +1,9 @@
+"""
+Tests for proteus.atmos_chem module
+"""
+from __future__ import annotations
+
+
+def test_placeholder():
+    """Placeholder test - replace with actual tests"""
+    pass
