+++ tests/utils/test_utils.py
@@ -1,0 +1,9 @@
+"""
+Tests for proteus.utils module
+"""
+from __future__ import annotations
+
+
+def test_placeholder():
+    """Placeholder test - replace with actual tests"""
+    pass
