### 2025-12-31 a45c633
Refactor CI workflow for improved testing and compatibility across Python versions and OS platforms
