+++ .github/workflows/ci.yml
@@ -1,123 +1,206 @@
-name: CI # Continuous Integration
+name: Tests for PROTEUS
 
-# Trigger on pushes/PRs to main/develop and manual workflow_dispatch
+# Trigger on pushes/PRs to main and manual workflow_dispatch
+# Aligned with existing tests.yaml structure for consistency
 on:
   push:
-    branches: [ main, develop ]
+    branches:
+      - main
   pull_request:
-    branches: [ main, develop ]
+    branches:
+      - main
+    types:
+      - opened
+      - reopened
+      - synchronize
+      - ready_for_review
   workflow_dispatch:
 
+permissions:
+  actions: write
+  contents: write
+
 jobs:
-  # Test suite with matrix testing across Python versions
-  # Provides fast feedback on code quality and coverage
-  test-matrix:
-    name: Test Suite
+  test:
+    # Test suite with matrix testing across Python versions and OS platforms
+    # Provides comprehensive coverage and ensures cross-platform compatibility
+    name: Run Coverage and Tests
     strategy:
-      fail-fast: false
+      fail-fast: false  # Continue testing all matrix combinations even if one fails
       matrix:
-        # Test against multiple Python versions for compatibility
-        python-version: ['3.11', '3.12', '3.13']
-        os: [ubuntu-latest]
+        os: ['ubuntu-latest', 'macos-14']
+        python-version: ['3.12', '3.13']
+        include:
+          # Ubuntu-specific system dependencies for compiled extensions
+          # netcdf: Required for data I/O operations
+          # libssl-dev: Required for cryptographic operations
+          - os: ubuntu-latest
+            INSTALL_DEPS: sudo apt-get update; sudo apt-get install libnetcdff-dev netcdf-bin libssl-dev tree
+            CC: gcc
+            CXX: g++
+            FC: gfortran
+          # macOS-specific dependencies and compiler setup
+          # Removes pkg-config conflicts and installs gfortran for Fortran compilation
+          - os: macos-14
+            INSTALL_DEPS: brew uninstall --force pkg-config;  rm -f /opt/homebrew/bin/pkg-config; rm -f /opt/homebrew/share/aclocal/pkg.m4; rm -f  /opt/homebrew/share/man/man1/pkg-config.1; brew install gfortran netcdf netcdf-fortran tree
+            CC: gcc
+            CXX: g++
+            FC: gfortran
+
+    env:
+      FWL_DATA: ${{ github.workspace }}/fwl_data
+      PROTEUS_DIR: ${{ github.workspace }}
+      RAD_DIR: ${{ github.workspace }}/socrates
+      AGNI_DIR: ${{ github.workspace }}/AGNI
+      JULIA_NUM_THREADS: 1
 
     runs-on: ${{ matrix.os }}
-
     steps:
-      # Checkout code with full history for accurate coverage analysis
-      # Note: submodules are recursive to include CALLIOPE, JANUS, MORS, VULCAN, ZEPHYRUS, Zalmoxis, aragog, others
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          submodules: recursive
-          fetch-depth: 0
 
-      # Configure Python environment with pip caching for faster builds
+      - name: Free Disk Space (Ubuntu)
+        uses: jlumbroso/free-disk-space@main
+        if: runner.os == 'Linux'
+        with:
+          tool-cache: false
+
+      - name: Free Disk Space (MacOS)
+        if: runner.os == 'macOS'
+        run: |
+          sudo rm -rf /opt/ghc
+          sudo rm -rf "/usr/local/share/boost"
+          sudo rm -rf "$AGENT_TOOLSDIRECTORY"
+
+
[…]
