### 2025-12-31 8bd6722
Update CI workflow to trigger on pushes/PRs to dev branch for improved testing coverage
