### 2025-12-31 804acb7
Add test branch to CI workflow triggers for testing
