### 2025-12-31 86339ef
Install local MORS and aragog in CI to test NumPy 2.0 fixes
