### 2025-12-31 c048960
Clone MORS and aragog repos explicitly in CI
