### 2026-01-02 a1f5441
Fix NumPy 2.0 compatibility in PROTEUS: convert numpy scalar to float in error message
