### 2026-01-02 25f932e
Fix NumPy 2.0 logging conversions in interior wrapper
