### 2026-01-02 52de500
CI: temporarily run only Ubuntu Python 3.13 (commented macOS/3.12)
