### 2026-01-02 41fafbd
Fix NumPy 2.0: convert array outputs to scalars in interior wrapper
