### 2026-01-02 1b0544f
Fix coverage: align pytest --cov path with package name
