+++ pyproject.toml
@@ -117,7 +117,7 @@ omit = [
 [tool.pytest.ini_options]
 minversion = "8.1"
 addopts = [
-    "--cov=src",
+    "--cov=proteus",
     "--cov-report=term-missing",
     "--cov-report=html",
     "--cov-report=xml",
