### 2026-01-02 71ba172
fix: Remove pytest-cov options conflicting with coverage run command
The CI workflow uses 'coverage run -m pytest' to collect coverage data.
Having --cov options in pytest addopts creates a conflict that prevents
coverage measurement. Coverage reporting is still configured in
[tool.coverage.report] section and will now work properly with the
CI command.
