### 2026-01-02 d016472
ci: Update aragog clone to use main branch
Now that NumPy 2.0 fixes are merged to aragog main (PR #5),
remove the temporary test branch reference from CI workflow.

Related: FormingWorlds/aragog#5
