+++ .github/workflows/ci.yml
@@ -109,7 +109,7 @@ jobs:
       - name: Clone MORS and aragog for local testing
         run: |
           git clone https://github.com/FormingWorlds/MORS.git
-          git clone --branch tl/deprecation_fixes_line138 https://github.com/FormingWorlds/aragog.git
+          git clone https://github.com/FormingWorlds/aragog.git
 
       # Get Lovepy
       - name: Get Lovepy
