### 2026-01-02 a89f94b
perf: Add caching for SOCRATES binaries and AGNI Julia depot
Implements performance optimizations to reduce CI runtime:

1. Cache SOCRATES compiled binaries (~7-8 min savings)
   - Caches socrates/ directory with binaries
   - Key based on source file hashes for automatic invalidation
   - Restore-keys for partial cache hits

2. Cache AGNI Julia depot (~3-4 min savings)
   - Caches AGNI/ directory and ~/.julia/ packages
   - Key based on Julia source and manifest files
   - Restore-keys for partial cache hits

Expected improvement: 10-12 minutes saved per CI run (from ~26 to ~14-16 minutes)

T
[…]
