### 2026-01-02 ee81d7c
fix: Correct cache restore/save order for SOCRATES and AGNI
Previous attempt failed because:
- Cache keys used hashFiles() on directories that didn't exist yet
- Result: empty cache keys like 'socrates-bins-Linux-'
- No cache was ever restored or saved properly

New approach:
- Use cache/restore before install-all to load previous build
- Use cache/save after tests to save new build
- Key based on run_id (unique) with restore-keys for prefix matching
- Only save cache if restore missed (avoid duplicate saves)

This allows second and subsequent runs to skip 10-12 minutes of compilation.
