### 2026-01-02 7e3aeb8
fix: Implement hash-based cache invalidation for SOCRATES and AGNI
CRITICAL FIX: Cache keys now depend on source/dependency hashes instead
of run ID. This ensures:

1. Cache is automatically INVALIDATED when source code changes
2. Cache is automatically INVALIDATED when dependencies change
3. Tests always use current SOCRATES and AGNI versions
4. No stale cached code is used if upstream repos change

SOCRATES cache:
- Key: hash of *.f90, *.F90, *.c files, and build_code script
- Invalidates when any Fortran/C source changes

AGNI cache:
- Key: hash of Project.toml and Manifest.toml files
- Inv
[…]
