### 2026-01-02 9597e32
feat: Conditional free disk space step (Option A optimization)
Skip disk cleanup when available space is >80%, saving ~2m30s per run

Changes:
- Add 'Check available disk space' step that calculates usage percentage
- Modify 'Free Disk Space (Ubuntu)' condition to only run if usage >20%
- Threshold can be tuned; 80% is conservative to prevent full disk

Expected savings: 2m 30s per build (disk rarely critical)
Impact: ~15% CI runtime reduction on test branches
Risk: LOW - cleanup still triggers if disk space actually needed
