### 2026-01-02 cfadff8
fix: Clone SOCRATES and AGNI before cache restore for hash-based keys
Critical bug fix: hashFiles() was evaluating on non-existent directories

Root cause:
- Cache restore steps tried to hash 'socrates/**/*.f90' files
- But socrates/ directory didn't exist yet (cloned later in install-all)
- Result: Empty hash → cache key 'socrates-bins-Linux-' (missing hash)
- Cache always missed → SOCRATES recompiled every run (+11-15 min)

Solution:
- Clone SOCRATES and AGNI repos BEFORE cache restore steps
- Now hashFiles() can compute proper hashes
- Cache keys like 'socrates-bins-Linux-abc123def456' work
[…]
