### 2026-01-02 315c6c2
docs: Update testing infrastructure with ecosystem deployment guide and 69% coverage threshold
Priority 1 improvements to testing_infrastructure.md:

Changes:
- Document PROTEUS Phase 1 completion (69.23% coverage achieved)
- Add comprehensive Phase 2 ecosystem integration guide
- Create 4-step quick start deployment for ecosystem modules
- Add advanced hash-based caching strategy documentation
- Update coverage threshold progression from 5% to 69% for PROTEUS
- Change reusable workflow default threshold from 5% to 30% (realistic for new modules)
- Add deployment checklist (~2 hours per module
[…]
