### 2026-01-02 ac0d3ce
docs: Add CALLIOPE Phase 2 improvements to testing infrastructure guide
- Add CALLIOPE as Phase 2 pilot reference implementation
- Document coverage ratcheting mechanism (auto-threshold updates)
- Establish ecosystem integration standards (Codecov, artifacts, test quality)
- Provide 4 direct reference links to CALLIOPE working examples
- Update Phase 2 quick start with CALLIOPE patterns
- Clarify rollout strategy for JANUS/MORS (Phase 2b/2c)
