### 2026-01-02 491a0c3
feat: Add coverage ratcheting mechanism and rename workflow/docs files
- Add tools/update_coverage_threshold.py for automatic threshold updates
- Implement coverage ratcheting step in CI (only increases, never decreases)
- Rename .github/workflows/ci.yml to ci_tests.yml for consistency with CALLIOPE
- Rename docs/testing_infrastructure.md to test_infrastructure.md (shorter, clearer)
- Update all references to renamed files in workflows and documentation
- Update pyproject.toml with ratcheting mechanism comments
- Add test_infrastructure.md to mkdocs.yml navigation

Coverage ratcheting ensures 
[…]
