### 2026-01-02 5d11f05
ci: Run full matrix nightly at 2am
