### 2026-01-02 47edb82
ci: fix workflow run commands
