### 2026-01-02 431e708
ci: gate macOS job to schedule
