### 2026-01-03 8e9b013
refactor: Update coverage requirements and testing documentation for automatic ratcheting
