+++ docs/test_infrastructure.md
@@ -161,7 +161,10 @@ omit = [
 ]
 
 [tool.coverage.report]
-fail_under = 5  # Adjust based on current coverage
+# Coverage threshold - automatically updated by CI when coverage increases (recommended)
+# See: tools/update_coverage_threshold.py and .github/workflows/ci_tests.yml
+# This value can only increase or stay the same (coverage ratcheting mechanism)
+fail_under = 5  # Will auto-ratchet upward as tests are added
 show_missing = true
 precision = 2
 exclude_lines = [
@@ -371,12 +374,12 @@ The testing infrastructure is designed for:
 - **MORS** - Stellar evolution module
 - **VULCAN** - Atmospheric chemistry module
 - **ZEPHYRUS** - Escape module
-- **Zalmoxis** - Interior evolution module
+- **Zalmoxis** - Interior structure module
 - **aragog** - Interior module (alternative)
 
 To be adapted for future modules as needed:
-- **AGNI**
-- **OBLIQUA**
+- **AGNI** (Julia)
+- **OBLIQUA** (Julia)
 - Others
 
 ### Current Status
@@ -486,7 +489,10 @@ omit = [
 ]
 
 [tool.coverage.report]
-fail_under = 30  # Start with realistic threshold; increase by 5-10% quarterly
+# Coverage threshold - automatically updated by CI when coverage increases (recommended)
+# See: tools/update_coverage_threshold.py for ratcheting mechanism
+# Alternative: manually increase by 5-10% quarterly if not using auto-ratcheting
+fail_under = 30  # Start with realistic threshold
 show_missing = true
 precision = 2
 exclude_lines = [
@@ -514,18 +520,30 @@ develop = [
 ```
 
 **Coverage Threshold Guidance:**
+
+**Option 1: Automatic Ratcheting (Recommended - CALLIOPE Pattern)**
+- Set initial baseline (20-30%)
+- Implement `tools/update_coverage_threshold.py`
+- CI automatically increases threshold when coverage improves
+- No manual updates needed
+- See CALLIOPE for reference implementation
+
+**Option 2: Manual Quarterly Updates (Fallback)**
 - **Start:** 20-30% (realistic baseline)
 - **Q2:** Increase to 35-40%
 - **Q4:** Increase to 50-60%
 - **Year 2:** Target 70%+
 
-Example progression for new module:
+Example progression with automatic ratcheting:
 
 ```toml
-fail_under = 30  # January 2026
-fail_under = 40  # April 2026
-fail_under = 50  # July 2026
-fail_under = 60  # October 2026
+# Initial setup
+fail_under = 30  # January 2026 (starting point)
+# After this, CI auto-updates as tests are added:
+fail_under = 34  # Auto-updated by CI
+fail_under = 42  # Auto-updated by CI
+fail_under = 58  # Auto-updated by CI
+# Reaches 70%+ naturally through continuous improvement
 ```
 
 **Advanced: Automatic Coverage Ratcheting (CALLIOPE Innovation)**
@@ -845,13 +863,22 @@ Goal: 80%+ coverage
 
 **Coverage Threshold Growth Plan:**
 
+**Recommended: Automatic Ratcheting (CALLIOPE Pattern)**
+
+```toml
+# PROTEUS Example (auto-ratcheting active)
+fail_under = 69  # Auto-updated by CI as coverage increases
+
+# CALLIOPE Example (auto-ratcheting active)
+fail_under = 18  # Auto-updated by CI as coverage increases
+```
+
+**Alternative: Manual Updates (if not using auto-ratcheting)**
+
 Start with realistic baseline, increase gradually:
 
 ```toml
[…]
