### 2026-01-03 c0b97a2
docs: Update coverage recommendations in test quality gate and infrastructure documentation
