### 2026-01-03 0f571e9
ci: finalize test infra fixes
