+++ tools/validate_test_structure.sh
@@ -39,7 +39,7 @@ for test_dir in tests/*/; do
     module=$(basename "$test_dir")
 
     # Skip special directories
-    if [[ "$module" == "data" || "$module" == "helpers" || "$module" == "__pycache__" ]]; then
+    if [[ "$module" == "data" || "$module" == "helpers" || "$module" == *__pycache__* ]]; then
         continue
     fi
 
@@ -60,7 +60,7 @@ for test_dir in tests/*/; do
     module=$(basename "$test_dir")
 
     # Skip special directories
-    if [[ "$module" == "data" || "$module" == "helpers" || "$module" == "__pycache__" ]]; then
+    if [[ "$module" == "data" || "$module" == "helpers" || "$module" == *__pycache__* ]]; then
         continue
     fi
 
