### 2026-01-03 3a1f2d1
ci: Implement second round of Copilot review fixes
- Remove unused 'import pytest' from placeholder test files (restructure_tests.sh)
- Remove test branches (tl/test_ecosystem_v1/v2) from CI trigger (ci_tests.yml)
- Remove unreachable coverage threshold step from macOS job (ci_tests.yml)
- Add exception logging to NumPy conversion handler (wrapper.py)
- Add explicit error message for missing tomli dependency (update_coverage_threshold.py)
- Update SOCRATES cloning guidance to pin commits for supply-chain security (test_infrastructure.md)
