+++ .github/workflows/ci_tests.yml
@@ -1,16 +1,11 @@
 name: CI Tests for PROTEUS # Continuous Integration Tests for PROTEUS
 
-# Trigger on pushes/PRs to main and dev branches, plus manual workflow_dispatch
-# Aligned with existing tests.yaml structure for consistency
-# Temporarily includes tl/test_ecosystem_v1 and v2 for testing
 # Nightly 02:00 UTC schedule runs the full OS matrix (Ubuntu + macOS)
 on:
   push:
     branches:
       - main
       - dev
-      - tl/test_ecosystem_v1
-      - tl/test_ecosystem_v2
   pull_request:
     branches:
       - main
@@ -499,29 +494,7 @@ jobs:
           echo $'\n```' >> $GITHUB_STEP_SUMMARY
           coverage report
 
-      # Update coverage threshold (automatic ratcheting mechanism)
-      # Only runs on main branch with Python 3.13 to avoid redundant updates
-      # Automatically increases threshold when coverage improves, never decreases
-      - name: Update coverage threshold
-        if: ${{ github.ref == 'refs/heads/main' && env.PYTHON_VERSION == '3.13' && runner.os == 'Linux' && !failure() }}
-        run: |
-          # Automatically ratchet coverage threshold upward when tests pass on main
-          python tools/update_coverage_threshold.py
 
-          # Check if pyproject.toml was modified
-          if git diff --quiet pyproject.toml; then
-            echo "No coverage threshold update needed"
-          else
-            echo "Coverage threshold increased - committing update"
-            git config user.name "github-actions[bot]"
-            git config user.email "github-actions[bot]@users.noreply.github.com"
-            git add pyproject.toml
-            COVERAGE_PCT=$(python -c 'import json; print(json.load(open("coverage.json"))["totals"]["percent_covered"])')
-            git commit -m "ci: Auto-update coverage threshold to ${COVERAGE_PCT}% [skip ci]"
-            git push https://x-access-token:${GITHUB_TOKEN}@github.com/${{ github.repository }}.git HEAD:${{ github.ref_name }}
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          GIST_TOKEN: ${{ secrets.GIST_TOKEN }}
 
       # Create dynamic coverage badge for documentation
       # Only runs on main branch with Python 3.13 to avoid redundant updates
