### 2026-01-03 8155e59
ci: Implement third round of Copilot review fixes (15 issues)
Critical fixes:
- Fix floating-point coverage comparisons using bc instead of bash operators
- Remove unreachable Linux-specific disk space steps from macOS job
- Refactor long INSTALL_DEPS command into multiline format
- Fix incomplete 'Make coverage badge' step

Code quality improvements:
- Break long lines in ci_tests.yml (152, 408) to comply with 96-char limit
- Wrap long cache key example in documentation (651 chars)
- Add shebang to coverage_analysis.sh for consistency
- Add period to test_utils.py docstring for consistency

E
[…]
