### 2026-01-03 0e0fdfa
ci: Fix long line lengths in ci_tests.yml cache keys (lines 152, 397)
- Break SOCRATES cache key onto multiple lines for readability
- All cache keys now under 96-character limit per ruff configuration
- Applies to both Ubuntu and macOS jobs in ci_tests.yml
- Fixes Copilot review comments qwe and qwf
