+++ .github/workflows/ci_tests.yml
@@ -149,7 +149,13 @@ jobs:
         id: cache-socrates-restore
         with:
           path: socrates/
-          key: socrates-bins-${{ runner.os }}-${{ hashFiles('socrates/**/*.f90', 'socrates/**/*.F90', 'socrates/**/*.c', 'socrates/build_code') }}
+          key: |
+            socrates-bins-${{ runner.os }}-${{ hashFiles(
+              'socrates/**/*.f90',
+              'socrates/**/*.F90',
+              'socrates/**/*.c',
+              'socrates/build_code'
+            ) }}
           restore-keys: |
             socrates-bins-${{ runner.os }}-
 
@@ -209,7 +215,13 @@ jobs:
         uses: actions/cache/save@v4
         with:
           path: socrates/
-          key: socrates-bins-${{ runner.os }}-${{ hashFiles('socrates/**/*.f90', 'socrates/**/*.F90', 'socrates/**/*.c', 'socrates/build_code') }}
+          key: |
+            socrates-bins-${{ runner.os }}-${{ hashFiles(
+              'socrates/**/*.f90',
+              'socrates/**/*.F90',
+              'socrates/**/*.c',
+              'socrates/build_code'
+            ) }}
 
       # Save AGNI installation for next run (only if not already cached)
       # Cache key based on AGNI dependency hashes - invalidates when dependencies change
@@ -394,7 +406,13 @@ jobs:
         id: cache-socrates-restore
         with:
           path: socrates/
-          key: socrates-bins-${{ runner.os }}-${{ hashFiles('socrates/**/*.f90', 'socrates/**/*.F90', 'socrates/**/*.c', 'socrates/build_code') }}
+          key: |
+            socrates-bins-${{ runner.os }}-${{ hashFiles(
+              'socrates/**/*.f90',
+              'socrates/**/*.F90',
+              'socrates/**/*.c',
+              'socrates/build_code'
+            ) }}
           restore-keys: |
             socrates-bins-${{ runner.os }}-
 
@@ -454,7 +472,13 @@ jobs:
         uses: actions/cache/save@v4
         with:
           path: socrates/
-          key: socrates-bins-${{ runner.os }}-${{ hashFiles('socrates/**/*.f90', 'socrates/**/*.F90', 'socrates/**/*.c', 'socrates/build_code') }}
+          key: |
+            socrates-bins-${{ runner.os }}-${{ hashFiles(
+              'socrates/**/*.f90',
+              'socrates/**/*.F90',
+              'socrates/**/*.c',
+              'socrates/build_code'
+            ) }}
 
       # Save AGNI installation for next run (only if not already cached)
       # Cache key based on AGNI dependency hashes - invalidates when dependencies change
