### 2026-01-03 dbfb9ae
ci: Fix long AGNI cache key lines (lines 171, 236, 429, 493)
