### 2026-01-03 1dae6ab
fix: Implement all Copilot review feedback (Priority 1, 2, 3)
Priority 1 - Critical Fixes:
- #38: Update coverage threshold documentation for clarity
- #40: Fix chmod +x documentation inconsistency
- #41: Ensure missing_count variable initialization (already present)
- #43: Make wrapper.py exception handling strict (log.error + raise)
- #45: Add tomlkit version constraint (>=0.11.0)
- #49, #51: Fix example thresholds from 5% to 30%

Priority 2 - Risk Mitigation:
- #44: Add git rebase error handling to prevent race conditions
- #46: Add bc availability check before floating-point comparisons
- 
[…]
