+++ src/proteus/interior/wrapper.py
@@ -223,13 +223,14 @@ def run_interior(dirs:dict, config:Config,
                     else:
                         hf_row[k] = val
                 except Exception as exc:
-                    log.warning(
+                    log.error(
                         "Failed to convert output value for key %r (%r) to a NumPy array/scalar (%s: %s)",
                         k,
                         val,
                         type(exc).__name__,
                         exc,
                     )
+                    raise
 
     # Update rheological parameters
     #    Only calculate viscosity here if using dummy module
