### 2026-01-03 7c66722
docs: Update tools/README.md with comprehensive descriptions of all available scripts
Added documentation for:
- Testing & Quality Assurance tools (4 scripts)
- External Repository Management tools (6 scripts)
- Data & Configuration Tools (3 scripts)
- Workflow & Results Management tools (1 script)
- Post-Processing & Analysis tools (4 scripts)

Each tool includes:
- Clear purpose statement
- What it does (bullet points)
- Usage examples with code blocks
- Requirements where applicable
- Exit codes or expected outputs
