+++ tools/README.md
@@ -1,170 +1,339 @@
 # PROTEUS Tools
 
-This directory contains utility scripts and tools for PROTEUS development.
+This directory contains utility scripts and tools for PROTEUS development, configuration management, data retrieval, and testing.
 
-## Available Tools
+## Testing & Quality Assurance
 
-### `validate_test_structure.sh`
+### validate_test_structure.sh
 
 **Purpose:** Validate that the `tests/` directory properly mirrors the `src/proteus/` structure.
 
 **What it does:**
-1. Checks for missing test directories
-2. Verifies test files exist in each directory
-3. Ensures `__init__.py` files are present
-4. Provides a summary report with colored output
+- Checks for missing test directories
+- Verifies test files exist in each directory
+- Ensures `__init__.py` files are present
+- Provides a summary report with colored output
 
 **Usage:**
+
 ```bash
 # From repository root
 bash tools/validate_test_structure.sh
-```
-
-**Example output:**
-```
-🔍 Validating test structure...
-
-Checking for missing test directories...
-✓ Found: tests/config
-✗ Missing: tests/escape (for src/proteus/escape)
-✓ Found: tests/grid
-
-Summary:
-  Test directories found: 10
-  Test directories missing: 3
-  __init__.py files missing: 2
-
-⚠ Run 'bash tools/restructure_tests.sh' to fix issues
 ```
 
 **Exit codes:**
 - `0`: All checks passed
 - `1`: Issues found (missing directories or __init__.py files)
 
-### `restructure_tests.sh`
+### restructure_tests.sh
 
 **Purpose:** Restructure the `tests/` directory to mirror the `src/proteus/` structure.
 
 **What it does:**
-1. Creates missing test directories for all source modules
-2. Moves misplaced test files to appropriate subdirectories
-3. Creates placeholder test files for untested modules
-4. Adds `__init__.py` files for proper Python package structure
+- Creates missing test directories for all source modules
+- Moves misplaced test files to appropriate subdirectories
+- Creates placeholder test files for untested modules
+- Adds `__init__.py` files for proper Python package structure
 
 **Usage:**
+
 ```bash
 # From repository root
 bash tools/restructure_tests.sh
 ```
 
-**Before:**
-```
-tests/
-├── conftest.py
-├── grid/
-├── inference/
-├── integration/
-├── test_cli.py
-├── test_config.py
-├── test_cpl_colours.py
-└── test_cpl_helpers.py
-```
-
-**After:**
-```
-tests/
-├── conftest.py
-├── atmos_chem/
-│   └── test_atmos_chem.py
-├── atmos_clim/
-│   └── test_atmos_clim.py
-├── config/
-│   └── test_config.py
-├── escape/
-│   └── test_escape.py
-├── grid/
-│   └── test_grid.py
-├── inference/
-│   └── test_inference.py
-├── interior/
-│   └── test_interior.py
-├── observe/
-│   └── test_observe.py
-├── orbit/
-│   └── test_orbit.py
-├── outgas/
-│   └── test_outgas.py
-├── plot/
-│   ├── test_cpl_colours.py
-│   └── test_cpl_helpers.py
-├── star/
-│   └── test_star.py
-├── utils/
-│   └── test_utils.py
-├── integration/
-│   └── ... (unchanged)
-├── test_cli.py (stays at root)
-└── test_init.py (stays at root)
-```
-
 **Safe to run multiple times:** The script checks for existing files before moving them.
 
-### `coverage_analysis.sh`
+### coverage_analysis.sh
 
 **Purpose:** Analyze test coverage by module and identify testing priorities.
 
 **What it does:**
-1. Runs pytest with coverage
-2. Shows coverage percentage for each module
[…]
