### 2026-01-03 6ff875e
docs: Update testing documentation for accurate pytest-cov vs coverage tool usage
- Update docs/test_infrastructure.md pytest configuration example to match actual pyproject.toml
- Clarify that global coverage flags were removed from pytest addopts
- Update coverage threshold example from 30% to current 69% (auto-ratcheted)
- Improve pyproject.toml comments on coverage[toml] vs pytest-cov distinction
- Add coverage tool options to copilot-instructions.md
- Document that both 'pytest --cov' and 'coverage run -m pytest' are supported
