### 2026-01-03 922eb67
ci: build socrates and init agni
