### 2026-01-03 c7d728a
ci: generate socrates julia wrappers
