### 2026-01-03 ac029f3
Switch CI to Python 3.12 and developer install pathway
- Change Python version from 3.13 to 3.12
- Replace pypi-based install-all with developer install for all submodules
- Install JANUS, CALLIOPE, ZEPHYRUS as editable packages
- Use tools/get_socrates.sh for SOCRATES build
- Use AGNI's get_agni.sh for AGNI build (includes wrapper gen and lib build)
- Clone SPIDER repo without building (skipping PETSc requirement)
- Follow installation.md developer pathway consistently across all modules
