### 2026-01-03 b459136
ci: Add test_infrastructure.md compliance enhancements
- Add test structure validation (bash tools/validate_test_structure.sh)
- Add pytest test discovery validation (pytest --collect-only)
- Add Codecov integration for coverage reporting
- Add HTML coverage artifact uploads (30-day retention)
- Fix Python version in auto-ratcheting conditions (3.13 -> 3.12)
- Apply validation steps to both test-linux and test-macos jobs

These changes ensure workflow compliance with test_infrastructure.md guidelines
for test organization, coverage tracking, and ecosystem integration standards.
