### 2026-01-03 1c6ad5b
fix: Resolve critical workflow issues for CI test execution
**Critical Fixes:**
1. Move pytest --collect-only validation AFTER Python and packages installed
   - Was: Running before Python setup (would fail with 'command not found')
   - Now: Runs after all dependencies installed for accurate test discovery

2. Add coverage.xml and coverage.html generation
   - Was: Only generating coverage.json
   - Now: Generates xml for Codecov, html for artifacts
   - Codecov upload now has files to upload
   - HTML artifact upload now has directory to upload

**Execution Flow (Both Jobs):**
✅ 1. Test with
[…]
