### 2026-01-03 8f797d8
fix: Remove Unicode emoji from validation script for CI compatibility
