### 2026-01-03 ae85486
CI: fast-fail test structure
