+++ src/proteus/cli.py
@@ -414,13 +414,11 @@ def _update_input_data(config_path: Path):
         # Only try data download if a config file is present.
         configuration = read_config_object(config_path)
         download_sufficient_data(configuration, clean=True)
-        click.secho("✅ Additional data has been downloaded.", fg="green")
+        click.secho("[+] Additional data has been downloaded.", fg="green")
         return True
 
     else:
-        click.echo(
-            f"⚠️ No config file found at {config_path}, skipping data download."
-        )
+        click.echo(f"[!] No config file found at {config_path}, skipping data download.")
         return False
 
 @cli.command()
@@ -442,7 +440,7 @@ def install_all(export_env: bool, config_path: Path):
     required_disk_space_in_GB = 5
     if not available_disk_space_in_GB > required_disk_space_in_GB:
         click.secho(
-            f"⚠️ You have {available_disk_space_in_GB:.3f} GB of disk space at your disposal.",
+            f"[!] You have {available_disk_space_in_GB:.3f} GB of disk space at your disposal.",
             fg="yellow",
         )
         click.secho(
@@ -450,7 +448,7 @@ def install_all(export_env: bool, config_path: Path):
             fg="yellow",
         )
         click.secho(
-            "❌ Aborting installation — 'proteus install-all'.",
+            "[x] Aborting installation -- 'proteus install-all'.",
             fg="red",
         )
         raise SystemExit(1)
@@ -460,21 +458,21 @@ def install_all(export_env: bool, config_path: Path):
     # --- Step 1: FWL_DATA directory ---
     fwl_data = resolve_fwl_data_dir()
     fwl_data.mkdir(parents=True, exist_ok=True)
-    click.secho(f"✅ FWL_DATA directory: {fwl_data}", fg="green")
+    click.secho(f"[+] FWL_DATA directory: {fwl_data}", fg="green")
 
     # --- Step 2: Install SOCRATES ---
     root = Path.cwd()
     socrates_dir = root / "socrates"
     if not socrates_dir.exists():
-        click.secho("🌤️ Installing SOCRATES...", fg="blue")
+        click.secho("[+] Installing SOCRATES...", fg="blue")
         try:
             subprocess.run(["bash", "tools/get_socrates.sh"], check=True)
         except subprocess.CalledProcessError as e:
-            click.secho("❌ Failed to install SOCRATES", fg="red")
+            click.secho("[x] Failed to install SOCRATES", fg="red")
             click.echo(e)
             raise SystemExit(1)
     else:
-        click.secho("✅ SOCRATES already present", fg="green")
+        click.secho("[+] SOCRATES already present", fg="green")
 
     rad_dir = socrates_dir.resolve()
     os.environ.setdefault("RAD_DIR", str(rad_dir))
@@ -483,7 +481,7 @@ def install_all(export_env: bool, config_path: Path):
 
     # --- Step 3: Julia check ---
     if not is_julia_installed():
-        click.secho("⚠️ Julia not found in PATH.", fg="yellow")
+        click.secho("[!] Julia not found in PATH.", fg="yellow")
         click.secho(
             "   Proteus requires Julia for AGNI.",
             fg="yellow",
@@ -494,7 +492,7 @@ def install_all(export_env: bool, config_path: Path):
         )
         click.secho(f'   Current PATH: {os.environ["PATH"]}', fg="white")
         click.secho(
-            "❌ Aborting installation — 'proteus install-all' cannot proceed without Julia.",
+            "[x] Aborting installation -- 'proteus install-all' cannot proceed without Julia.",
             fg="red",
         )
         raise SystemExit(1)
@@ -502,7 +500,7 @@ def install_all(export_env: bool, config_path: Path):
[…]
