### 2026-01-18 #8758 @timlichtenberg
Thanks @warrickball and @TobiBu for moving the submission further. Indeed, it has been quite some wait, but we understand the issues you have been facing. The new sections and weighting make sense to us. Will we receive reviews now first on the previous submission, or do you want us to update the submission before reviewers will have a look at it? We would prefer the former, as we are currently undergo a substantial developer team reorganisation and have been working on major updates to our documentation and automated testing frameworks in the last few weeks. Given these significant updates, rewriting and agreeing on a new version right now will take us a significant amount of time; and I cannot see this happening before ~March to April to be fully transparent. With that in mind, it would be beneficial if we could get the views of the reviewers now first, and then together with their comments rework the submission to adhere to the new editorial requirements, including all the new updates to documentation and testing standards. Would that be an option? Thanks a lot in advance!
