### 2026-01-19 #8758 @TobiBu
> > 👋 [@lorenaacuna](https://github.com/lorenaacuna) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Hi [@warrickball](https://github.com/warrickball), [@timlichtenberg](https://github.com/timlichtenberg). Thank you so much for suggesting me as a reviewer for this software, it look very exciting. Unfortunately, I am already committed to other service responsibilities this year and will not be available. I would, however, be very pleased to review future software submissions in 2027+.

Dear @lorenaacuna,

thanks a lot for your quick response. That is much appreciated!
