### 2026-01-20 #8758 @TobiBu
👋 @MartianColonist @kathlandgren @zkbt - Would one of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
