### 2026-01-20 #8758 @TobiBu
> Thanks [@warrickball](https://github.com/warrickball) and [@TobiBu](https://github.com/TobiBu) for moving the submission further. Indeed, it has been quite some wait, but we understand the issues you have been facing. The new sections and weighting make sense to us. Will we receive reviews now first on the previous submission, or do you want us to update the submission before reviewers will have a look at it? We would prefer the former, as we are currently undergo a substantial developer team reorganisation and have been working on major updates to our documentation and automated testing frameworks in the last few weeks. Given these significant updates, rewriting and agreeing on a new version right now will take us a significant amount of time; and I cannot see this happening before ~March to April to be fully transparent. With that in mind, it would be beneficial if we could get the views of the reviewers now first, and then together with their comments rework the submission to adhere to the new editorial requirements, including all the new updates to documentation and testing standards. Would that be an option? Thanks a lot in advance!

Dear @timlichtenberg,

I understand the difficulty of reorganising the developer team and rewriting a paper version.
However, in the end the reviewers will need to look at the paper version that includes all new paragraphs.
From reading your paper, I think much of what is needed under the new paragraphs is already there.
In the end we all want to minimise the mutual work needed to finish the review process. From this perspective it would be mindful to let the reviewer look only at one paper version instead of at the current and then a revised version including new paragraphs.

If it's completely impossible for you to provide a new paper version, we might start the review anyways communicating to the reviewers that they will need to look at a revised paper version that includes new paragraphs in depth in the end.
