### 2026-01-21 #8758 @nenasedk
Hi Tobias,

Sorry for taking so long to get back to you, I got the emails from a
different JOSS review mixed up with this one. Unfortunately, I'm already
committed to other reviews at the moment, and am unable to help with this
one.

Best,
Evert

On Fri, Jan 16, 2026 at 7:54 PM Tobias Buck ***@***.***>
wrote:

> *TobiBu* left a comment (openjournals/joss-reviews#8758)
> <https://github.com/openjournals/joss-reviews/issues/8758#issuecomment-3761568626>
>
> 👋 @nenasedk <https://github.com/nenasedk> - Would you be willing to
> review this submission to JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
