### 2026-01-21 #8758 @timlichtenberg
> Hi [@TobiBu](https://github.com/TobiBu) and [@timlichtenberg](https://github.com/timlichtenberg) -- I'm happy to do this review, but I'm totally swamped through about mid February. If a review sometime late February or early March is OK, please sign me up!
>
> I'm fine knowing the specific paragraphs will be revised for the new JOSS standards, and I'd start with trying to give helpful comments on the code and documentation first.

Awesome, thanks @zkbt! From my side that would be completely fine.
