### 2026-01-22 #8758 @TobiBu
> > Hi [@TobiBu](https://github.com/TobiBu) and [@timlichtenberg](https://github.com/timlichtenberg) -- I'm happy to do this review, but I'm totally swamped through about mid February. If a review sometime late February or early March is OK, please sign me up!
> > I'm fine knowing the specific paragraphs will be revised for the new JOSS standards, and I'd start with trying to give helpful comments on the code and documentation first.
>
> Awesome, thanks [@zkbt](https://github.com/zkbt)! From my side that would be completely fine.

@zkbt also thanks a lot from my side. This should be fine with me as well.
It might anyways take few more days until we can start the review thread. I am still looking via GitHub and mail for potential reviewers and only keep getting declines. Let's hope for the best soon.
