### 2026-01-28 #8758 @TobiBu
👋 @cjcollin37 @chandan-kumarsahu - Would one of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
