### 2026-01-29 #9917 @editorialbot
[REVIEW]: PROTEUS: A modular framework for simulating planetary evolution
**Submitting author:** @timlichtenberg (<a href="http://orcid.org/0000-0002-3286-7683">Tim Lichtenberg</a>)
**Repository:** https://github.com/FormingWorlds/PROTEUS
**Branch with paper.md** (empty if default branch):
**Version:** 25.07.31
