### 2026-01-29 #9917 @TobiBu
👋 @timlichtenberg @rieder @zkbt - This is the review thread for the paper. All of our communications will happen here from now on.

Both reviewers should create their own checklists as described at the top of this thread (in that first comment) with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention <LINK> so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

Note, the paper PDF currently does comply with the recently changed JOSS scope but we agreed with the authors to do a first round of reviewer comments and @timlichtenberg will then make sure to update the paper PDF to the new scope requirements (including shortening to the recommended word count).
