### 2026-02-27 #9917 @rieder
@timlichtenberg I see that some required sections are missing from the paper (see https://github.com/openjournals/joss-reviews/issues/9917#issuecomment-3820406418), are you working on adding these?
