### 2026-03-05 #9917 @rieder
Some confusing bits in the installation guide that I would like to see improved:
- `gfortran` and `nf-config` should be listed as a dependency (probably best installed from conda-forge?), without it I get stuck on a clean install at the part where I need to install Socrates. ~I'm not sure how to install `nf-config` yet~ `nf-config` requires the conda-forge packages `netcdf-fortran` and `libnetcdf`.
- "Julia 1.12+ is not yet supported due to OpenSSL library incompatibilities with Python" -> Please monitor if this remains the case, I would expect it not to be globally true. It seems that there are other issues, maybe link to the following tracking issue: https://github.com/nichollsh/AGNI/issues/145
- "You should only obtain Julia using the official installer, not via your computer's package manager." -> Why is that? Is this really necessary, and how do I prevent conflicts with a Julia that was already installed that way?
- `wget` is also a dependency (for installing AGNI)
- `zalmoxis` requires the python package `osfclient`
