### 2026-03-10 #9917 @rieder
@timlichtenberg Could you check https://github.com/openjournals/joss-reviews/issues/9917#issuecomment-4005305326? If this is addressed I could probably tick off "Installation: Does installation proceed as outlined in the documentation?".
