### 2026-03-10 #9917 @timlichtenberg
@rieder: I will, saw the comments, very useful! I think it would be ok if we address this together with all the other changes likely necessary to the paper and code. We're just doing anyway a lot of documentation updates throughout, and can incorporate this into the next one; probably sometime next week.
