### 2026-03-15 #9917 @timlichtenberg
@rieder: we now implemented already some of your comments in the installation guidelines: https://proteus-framework.org/PROTEUS/How-to/installation.html (thanks @stuitje!)
