### 2026-03-17 #9917 @rieder
> [@rieder](https://github.com/rieder): we now implemented already some of your comments in the installation guidelines: https://proteus-framework.org/PROTEUS/How-to/installation.html (thanks [@stuitje](https://github.com/stuitje)!)

thanks, I'll check again.
