### 2026-08-29 #9917 @TobiBu
@timlichtenberg I hope the summer went well so far.
I am just checking in on the status of the paper updates? Any progress on your side? I think the paper updates are the only thing missing before we can move forward with the submission.
